### 2025-07-10 #8571 @editorialbot
[PRE REVIEW]: gkwreg: An R Package for Generalized Kumaraswamy Regression Models for Bounded Data
**Submitting author:** @evandeilton (<a href="http://orcid.org/0009-0007-5887-4084">José Evandeilton Lopes</a>)
**Repository:** https://github.com/evandeilton/gkwreg
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.10
