### 2025-07-10 #8571 @evandeilton
Hi @crvernon .
My mistake.

The repo link for the article files is wrong.
Correct one is this <https://github.com/evandeilton/gkwreg-joss>

Is it possible to fix this in the submission page from JOSS?
