### 2025-07-12 2e65900
Add files for compiling paper
