+++ paper.md
@@ -1,0 +1,856 @@
+---
+# Example from https://joss.readthedocs.io/en/latest/submitting.html
+title: 'gkwreg: An R Package for Generalized Kumaraswamy Regression Models for Bounded Data'
+tags:
+  - R
+  - gkwreg
+  - generalized kumaraswamy
+  - regression
+  - maximum likelihood
+  - bounded data
+  - TMB
+authors:
+  - name: José Evandeilton Lopes
+    orcid: 0009-0007-5887-4084
+    affiliation: "1"
+  - name: Wagner Hugo Bonat
+    orcid: 0000-0002-0349-7054
+    affiliation: "1"
+affiliations:
+ - name: Paraná Federal University, Brazil
+   index: 1
+ # - name: Universidade Federal do Paraná (UFPR)
+ #   index: 2
+citation_author: Lopes, J.E et. al.
+date: '10 julho 2025'
+year: '2025'
+bibliography: paper.bib
+output: rticles::joss_article
+csl: apa.csl
+journal: JOSS
+header-includes:
+  - \usepackage{booktabs}
+  - \usepackage{array}
+  # - \usepackage{amsmath}
+  - \usepackage{amssymb}
+  - \usepackage{amsfonts}
+  - \usepackage{multirow}
+  - \usepackage{longtable}
+  - \usepackage{caption}
+  - \usepackage{float}
+  # - \usepackage{colortbl} # Uncomment if necessary
+  # - \usepackage{threeparttable} # Uncomment if necessary
+---
+
+
+
+
+
+
+
+
+# Summary
+
+Modeling data restricted to the unit interval $(0,1)$, such as proportions and rates, is a common challenge. The beta distribution, although predominant, has limitations regarding flexibility and analytical tractability (CDF), motivating alternatives, especially in regression contexts. The Kumaraswamy (Kw) distribution offers analytical simplicity (closed-form CDF/PDF), but it may have insufficient flexibility. In this context, the Generalized Kumaraswamy (GKw) distribution, a five-parameter flexible model nesting Beta, Kw, and other relevant distributions, emerges as a promising alternative. However, the complexity of its likelihood function makes maximum likelihood estimation (MLE) difficult, particularly in multiparametric regression. This article introduces the $\textbf{R}$ package `gkwreg`, developed for efficient (MLE) estimation and inference in GKw regression models and its seven main subfamilies. `gkwreg` uses the robust Template Model Builder (`TMB`) framework, employing automatic differentiation for fast and accurate gradient/Hessian computation and effective optimization algorithms. This integrated approach via `TMB` enables reliable, robust, and efficient estimation of complex models, even when multiple parameters are linked to covariates by various link functions. The package also includes routines in `RcppArmadillo` and `C++` that implement distribution functions `p`, `d`, `r`, and `q`, as well as their respective log-likelihood functions, gradients, and analytic Hessians for the seven GKw subfamilies and a function for fitting distributions to data. We detail the statistical foundations of the GKw family, its hierarchy, the regression framework, inference, diagnostics, and the package’s computational implementation. Real data examples demonstrate `gkwreg`’s ability to capture complex distributional patterns that simpler models may overlook. The package thus provides a unified, flexible, and computationally efficient tool for analyzing responses restricted to the unit interval.
+
+
+# Introduction
+
[…]
