### 2025-07-12 cfe2d10
Clear some not used files
