+++ paper.Rmd
@@ -1,838 +1,0 @@
----
-# Example from https://joss.readthedocs.io/en/latest/submitting.html
-title: 'gkwreg: An R Package for Generalized Kumaraswamy Regression Models for Bounded Data'
-tags:
-  - R
-  - gkwreg
-  - generalized kumaraswamy
-  - regression
-  - maximum likelihood
-  - bounded data
-  - TMB
-authors:
-  - name: José Evandeilton Lopes
-    orcid: 0009-0007-5887-4084
-    affiliation: "1"
-  - name: Wagner Hugo Bonat
-    orcid: 0000-0002-0349-7054
-    affiliation: "1"
-affiliations:
- - name: Paraná Federal University, Brazil
-   index: 1
- # - name: Universidade Federal do Paraná (UFPR)
- #   index: 2
-citation_author: Lopes, J.E et. al.
-date: '`r format(Sys.time(), "%d %B %Y")`'
-year: '`r format(Sys.time(), "%Y")`'
-bibliography: paper.bib
-output: rticles::joss_article
-csl: apa.csl
-journal: JOSS
-header-includes:
-  - \usepackage{booktabs}
-  - \usepackage{array}
-  # - \usepackage{amsmath}
-  - \usepackage{amssymb}
-  - \usepackage{amsfonts}
-  - \usepackage{multirow}
-  - \usepackage{longtable}
-  - \usepackage{caption}
-  - \usepackage{float}
-  # - \usepackage{colortbl} # Uncomment if necessary
-  # - \usepackage{threeparttable} # Uncomment if necessary
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE,
-                      prompt = 'R> ',
-                      continue = '+ ',
-                      fig.align = 'center', # Center figures by default
-                      message = FALSE,      # Hide messages
-                      warning = FALSE,       # Hide warnings
-                      knitr.table.format = "latex"
-                      )
-```
-
-
-```{r load-packages, echo=FALSE, message=FALSE, warning=FALSE, include=FALSE}
-# Silently load required packages
-library(gkwreg)
-library(simplexreg)
-library(betareg)
-library(knitr)
-library(dplyr)
-source("gkwreg-examples.R")
-```
-
-
-
-# Summary
-
[…]
