+++ gkwreg-examples.R
@@ -1,285 +1,0 @@
-## -------------------------------------------------------------------------- ##
-## gkwreg: Generalized Kumaraswamy Regression Models for Bounded Data
-## Author: Lopes, J. E. // 2025
-## -------------------------------------------------------------------------- ##
-
-if(!require(gkwreg)) install.packages("gkwreg")
-if(!require(skimr)) install.packages("skimr")
-
-## -------------------------------------------------------------------------- ##
-## Extra functions for analysis
-## -------------------------------------------------------------------------- ##
-
-# Wrapper function to compare different gkw models for bounded data
-fit_and_compare_models_2p <- function(
-    formula, data, method = "BFGS",
-    families = c("gkw", "bkw", "kkw", "ekw", "mc", "kw", "beta"),
-    use_betareg = TRUE) {
-  
-  # Map family codes to descriptive names
-  family_names <- c(
-    gkw = "Generalized Kumaraswamy", bkw = "Beta-Kumaraswamy",
-    kkw = "Kumaraswamy-Kumaraswamy", ekw = "Exponential Kumaraswamy",
-    mc = "McDonald/Beta Power", kw = "Kumaraswamy", beta = "Beta (gkwreg)"
-  )
-  
-  # Number of parameters for each family
-  family_params <- list(
-    gkw = 5, bkw = 4, kkw = 4, ekw = 3, mc = 3, kw = 2, beta = 2
-  )
-  
-  results <- list()
-  
-  # Fit gkwreg models
-  for (fam in families) {
-    num_links <- family_params[[fam]]
-    
-    # Try to fit model with basic error handling
-    fit <- try(gkwreg(formula, data = data, family = fam,
-                      link = rep("log", num_links),
-                      method = method, silent = TRUE), silent = TRUE)
-    
-    # Extract model statistics if fit successful
-    if (!inherits(fit, "try-error")) {
-      results[[fam]] <- data.frame(
-        Model = family_names[[fam]],
-        Code = fam,
-        Parameters = length(unlist(coef(fit))),
-        logLik = as.numeric(logLik(fit)),
-        AIC = AIC(fit),
-        BIC = BIC(fit),
-        Conv = paste0(as.numeric(fit$convergence), collapse = ";"),
-        row.names = fam
-      )
-    }
-  }
-  
-  # Fit betareg model if requested
-  if (use_betareg) {
-    
-    # Fit standard beta regression
-    br_fit <- try(
-      betareg::betareg(formula, data = data,
-                       link = "log", link.phi = "log"),
-      silent = TRUE)
-    
-    # Add betareg results if successful
-    if (!inherits(br_fit, "try-error")) {
-      results[["betareg"]] <- data.frame(
-        Model = "Beta (betareg)",
-        Code = "betareg",
-        Parameters = length(coef(br_fit)),
-        logLik = as.numeric(logLik(br_fit)),
-        AIC = AIC(br_fit),
-        BIC = BIC(br_fit),
-        # Conv = max(as.numeric(br_fit$converged)),
-        Conv = paste0(as.numeric(br_fit$converged), collapse = ";"),
-        row.names = NULL
-      )
-    }
-  }
-  
-  # Combine all results and sort by AIC
-  results_df <- do.call(rbind, results)
-  results_df <- results_df[order(results_df$AIC), ]
-  rownames(results_df) <- NULL
-  
-  return(results_df)
-}
-
-# Function to fit and compare models with 3+ parameters
-fit_and_compare_models_3p <- function(formula, data, method = "nlminb",
-                               families = c("gkw", "bkw", "kkw", "ekw", "mc")) {
-  # Map family codes to descriptive names  
-  family_names <- c(
[…]
