+++ simulacao/draft/gkw_bias_simulation.R
@@ -1,409 +1,0 @@
-# ---------------------------------------------------------------------------- #
-# Title: Simulation Script for GKw MLE Bias Study
-# Description: This script runs the simulation study to evaluate bias in MLEs
-#              for the Generalized Kumaraswamy distribution family using four
-#              estimation methods: TMB, Newton-Raphson, nlminb, and optim.
-# ---------------------------------------------------------------------------- #
-
-# Load required packages
-library(tidyverse)
-library(parallel)
-library(foreach)
-library(doParallel)
-library(ggh4x)    # For facet_nested
-library(latex2exp) # For rendering LaTeX in plots
-library(gkwreg)   # Assumed to contain all GKw distribution functions and fit methods
-
-# Load helper functions
-source("gkw_bias_extra_functions.R")
-
-# ---------------------------------------------------------------------------- #
-# Setup Configuration
-# ---------------------------------------------------------------------------- #
-
-# Define the configurations to test
-configurations <- list(
-  # GKw (5 parameters)
-  gkw_bimodal = list(family = "gkw", params = list(alpha = 0.3, beta = 0.5, gamma = 1, delta = 1, lambda = 3)),
-  gkw_u_shape = list(family = "gkw", params = list(alpha = 0.5, beta = 0.5, gamma = 0.5, delta = 0.5, lambda = 1)),
-  gkw_j_shape = list(family = "gkw", params = list(alpha = 2, beta = 0.5, gamma = 1, delta = 1, lambda = 1)),
-
-  # BKw (4 parameters)
-  bkw_l_shape = list(family = "bkw", params = list(alpha = 1, beta = 3, gamma = 0.3, delta = 3)),
-  bkw_unimodal = list(family = "bkw", params = list(alpha = 2, beta = 2, gamma = 2, delta = 2)),
-
-  # KKw (4 parameters)
-  kkw_j_shape = list(family = "kkw", params = list(alpha = 2, beta = 0.5, delta = 1, lambda = 1)),
-  kkw_bimodal = list(family = "kkw", params = list(alpha = 0.5, beta = 5, delta = 30, lambda = 1)),
-
-  # EKw (3 parameters)
-  ekw_asym = list(family = "ekw", params = list(alpha = 2, beta = 3, lambda = 1.5)),
-  ekw_extreme = list(family = "ekw", params = list(alpha = 0.5, beta = 0.5, lambda = 10)),
-
-  # Mc (3 parameters)
-  mc_u_shape = list(family = "mc", params = list(gamma = 0.5, delta = 0.5, lambda = 1)),
-  mc_unimodal = list(family = "mc", params = list(gamma = 2, delta = 2, lambda = 1)),
-
-  # Kw (2 parameters)
-  kw_j_shape = list(family = "kw", params = list(alpha = 0.5, beta = 3)),
-  kw_unimodal = list(family = "kw", params = list(alpha = 2, beta = 2)),
-
-  # Beta (2 parameters)
-  beta_u_shape = list(family = "beta", params = list(gamma = 0.5, delta = 0.5)),
-  beta_unimodal = list(family = "beta", params = list(gamma = 2, delta = 2))
-)
-
-# Estimation methods to compare
-methods <- c("tmb", "nr", "nlminb", "optim")
-
-# Sample sizes to test
-# sample_sizes <- c(50, 100, 250, 500, 1000)
-sample_sizes <- c(250, 500)
-
-# Number of simulation replications
-# n_sim <- 1000
-n_sim <- 200
-
-# Create timestamp for this run
-run_timestamp <- format(Sys.time(), "%Y%m%d_%H%M%S")
-
[…]

[…]
