+++ inst/paper/simulacao/draft/gkw_bias_extra_functions.R
@@ -1,2715 +1,0 @@
-# ---------------------------------------------------------------------------- #
-# Title: Utility Functions for GKw MLE Bias Study
-# Description: Contains all functions needed for analysis, visualization, and
-#              reporting in the GKw MLE bias simulation study
-# ---------------------------------------------------------------------------- #
-
-# ---------------------------------------------------------------------------- #
-# Utility Functions
-# ---------------------------------------------------------------------------- #
-
-#' Generate a deterministic seed based on configuration
-#'
-#' @param i Iteration number
-#' @param config_name Configuration name
-#' @param method Estimation method
-#' @param sample_size Sample size
-#' @param base_seed Base seed to start from
-#' @return Integer seed value
-generate_seed <- function(i, config_name, method, sample_size, base_seed = 12345) {
-  # Encode method as a number (1-4)
-  method_num <- match(method, c("tmb", "nr", "nlminb", "optim"))
-
-  # Encode config_name as a number based on its position
-  config_num <- which(names(configurations) == config_name)
-
-  # Create a deterministic but unique seed
-  seed <- base_seed + i * 10000 + config_num * 100 + method_num * 10 +
-    as.numeric(as.factor(sample_size))
-
-  # Ensure it's within integer limits
-  seed <- seed %% 2147483647  # Max integer in R
-
-  return(seed)
-}
-
-#' Create a data frame from the parameter results with relative bias
-#'
-#' @param param_name Parameter name
-#' @param param_value True parameter value
-#' @param method Estimation method
-#' @param family Distribution family
-#' @param config Configuration name
-#' @param sample_size Sample size
-#' @param bias Absolute bias
-#' @param se Standard error
-#' @param emp_se Empirical standard error
-#' @param coverage Coverage probability
-#' @return Data frame with all bias metrics
-create_param_df <- function(param_name, param_value, method, family, config,
-                            sample_size, bias, se, emp_se, coverage) {
-  # Calculate bias_z outside the data.frame call
-  bias_z_val <- bias / se
-
-  # Calculate relative bias (as percentage of true value)
-  rel_bias <- ifelse(param_value != 0,
-                     (bias / param_value) * 100,
-                     NA)  # Avoid division by zero
-
-  # Calculate relative standard error as percentage of true value
-  rel_se <- ifelse(param_value != 0,
-                   (se / param_value) * 100,
-                   NA)
-
-  # Calculate relative empirical standard error
-  rel_emp_se <- ifelse(param_value != 0,
-                       (emp_se / param_value) * 100,
-                       NA)
-
-  # Create data frame with all metrics
-  data.frame(
-    parameter = param_name,
-    true_value = param_value,
-    method = method,
-    family = family,
-    config = config,
-    sample_size = sample_size,
-    bias = bias,                    # Absolute bias
-    rel_bias = rel_bias,            # Relative bias (%)
-    se = se,                        # Standard error
-    emp_se = emp_se,                # Empirical standard error
-    rel_se = rel_se,                # Relative standard error (%)
-    rel_emp_se = rel_emp_se,        # Relative empirical standard error (%)
-    coverage = coverage,            # Coverage probability
-    bias_z = bias_z_val,            # Standardized bias
-    bias_lower_z = bias_z_val - 1,  # Lower bound for standardized bias
-    bias_upper_z = bias_z_val + 1,  # Upper bound for standardized bias
-    stringsAsFactors = FALSE
-  )
-}
-
-#' Compute coverage rate
-#'
[…]
