+++ inst/paper/simulacao/gkw_bias_analysis.R
@@ -1,1201 +1,0 @@
-# ---------------------------------------------------------------------------- #
-# Title: Analysis Script for GKw MLE Bias Study
-# Description: This script analyzes the results of the GKw MLE bias simulation
-#              study, generates plots and tables for all parameters and methods.
-# ---------------------------------------------------------------------------- #
-
-# Load required packages
-library(tidyverse)
-library(ggplot2)
-library(patchwork)
-library(ggh4x)    # For facet_nested
-library(latex2exp) # For rendering LaTeX in plots
-library(knitr)
-library(kableExtra)
-
-# Load helper functions
-source("gkw_bias_extra_functions.R")
-
-# ---------------------------------------------------------------------------- #
-# Load Results
-# ---------------------------------------------------------------------------- #
-
-# Create output directories if they don't exist
-plots_dir <- "plots"
-param_plots_dir <- file.path(plots_dir, "parameter_focus")
-csv_dir <- "csv"
-
-dir_list <- c(plots_dir, param_plots_dir, csv_dir)
-for (dir_path in dir_list) {
-  if (!dir.exists(dir_path)) {
-    dir.create(dir_path, recursive = TRUE)
-  }
-}
-
-# Load the aggregated results
-param_results <- readRDS("rds/gkw_bias_param_results.rds")
-convergence_df <- readRDS("rds/gkw_bias_convergence_results.rds")
-run_config <- readRDS("rds/run_config.rds")
-
-# Extract simulation parameters from config
-methods <- run_config$methods
-sample_sizes <- run_config$sample_sizes
-configurations <- run_config$configurations
-
-# Verify data loaded correctly
-cat("Loaded parameter results:", nrow(param_results), "rows\n")
-cat("Loaded convergence results:", nrow(convergence_df), "rows\n")
-
-# Extract unique parameters, families, and configurations
-unique_params <- unique(param_results$parameter)
-unique_families <- unique(param_results$family)
-unique_configs <- unique(param_results$config)
-
-cat("Parameters in results:", paste(unique_params, collapse=", "), "\n")
-cat("Families in results:", paste(unique_families, collapse=", "), "\n")
-
-# ---------------------------------------------------------------------------- #
-# Generate Summary Tables
-# ---------------------------------------------------------------------------- #
-
-cat("Generating summary tables...\n")
-
-# 1. Bias summary table
-bias_summary <- create_bias_summary(param_results)
-write_csv(bias_summary, file.path(csv_dir, "gkw_bias_summary.csv"))
-
-# 2. Coverage summary table
-coverage_summary <- create_coverage_summary(param_results)
-write_csv(coverage_summary, file.path(csv_dir, "gkw_coverage_summary.csv"))
-
-# 3. Convergence and timing summary
-convergence_summary <- create_convergence_summary(convergence_df)
-write_csv(convergence_summary, file.path(csv_dir, "gkw_convergence_summary.csv"))
-
-# 4. Method comparison summary
-method_comparison <- create_method_comparison(param_results, convergence_df)
-write_csv(method_comparison, file.path(csv_dir, "gkw_method_comparison.csv"))
-
-# 5. Method rankings
-method_rankings <- rank_methods(param_results, convergence_df,
-                                rel_bias_weight = 0.4,
-                                coverage_weight = 0.3,
[…]

[…]
