### 2025-07-13 ffcd213
Revert "Fix minors in fitted funtion and update versions"
This reverts commit ba028ad462224514753eedb5da9432324ab52cbc.
