+++ NAMESPACE
@@ -8,7 +8,6 @@ S3method(anova,gkwfit)
 S3method(coef,gkwfit)
 S3method(coef,gkwreg)
 S3method(confint,gkwfit)
-S3method(fitted,gkwreg)
 S3method(logLik,gkwfit)
 S3method(logLik,gkwreg)
 S3method(plot,gkwfit)
@@ -40,6 +39,7 @@ export(dkkw)
 export(dkw)
 export(dmc)
 export(extract_gof_stats)
+export(fitted.gkwreg)
 export(get_bounded_datasets)
 export(gkwfit)
 export(gkwfitall)
