### 2025-07-13 53a5c70
Revert "Increment version number to 1.0.10"
This reverts commit 26152a0c950025399571c9eb4a2540ba8d10ed2d.
