+++ README.Rmd
@@ -1,251 +1,0 @@
----
-output: github_document
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.path = "man/figures/README-",
-  out.width = "100%"
-)
-```
-
-# gkwreg: Generalized Kumaraswamy Regression Models for Bounded Data <img src="man/figures/gkwreg.png" align="right" height="138" />
-
-[![CRAN status](https://www.r-pkg.org/badges/version/gkwreg)](https://CRAN.R-project.org/package=gkwreg)
-[![R-CMD-check](https://github.com/evandeilton/gkwreg/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/evandeilton/gkwreg/actions/workflows/R-CMD-check.yaml)
-[![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/gkwreg)](https://cran.r-project.org/package=gkwreg)
-[![License:MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
-
-## Overview
-
-The `gkwreg` package provides a robust and efficient framework for modeling data restricted to the standard unit interval $(0, 1)$, such as proportions, rates, fractions, or indices. While the Beta distribution is commonly used for such data, `gkwreg` focuses on the **Generalized Kumaraswamy (GKw)** distribution family, offering enhanced flexibility by encompassing several important bounded distributions (including Beta and Kumaraswamy) as special cases.
-
-The package facilitates both distribution fitting and regression modeling with potentially all distribution parameters modeled as functions of covariates using various link functions. Estimation is performed efficiently via **Maximum Likelihood** leveraging the **Template Model Builder (TMB)** framework, which utilizes automatic differentiation for superior speed, accuracy, and stability.
-
-## Key Features
-
-* **Flexible Distribution Family:** Model data using the 5-parameter Generalized Kumaraswamy (GKw) distribution and its seven key nested sub-families:
-  
-  | Distribution | Code | Parameters Modeled | Fixed Parameters | # Par. |
-  |:-------------|:-----|:-------------------|:-----------------|:-------|
-  | Generalized Kumaraswamy | `gkw` | `alpha`, `beta`, `gamma`, `delta`, `lambda` | `None` | 5 |
-  | Beta-Kumaraswamy | `bkw` | `alpha`, `beta`, `gamma`, `delta` | `lambda` = 1 | 4 |
-  | Kumaraswamy-Kumaraswamy | `kkw` | `alpha`, `beta`, `delta`, `lambda` | `gamma` = 1 | 4 |
-  | Exponentiated Kumaraswamy | `ekw` | `alpha`, `beta`, `lambda` | `gamma` = 1, `delta` = 0 | 3 |
-  | McDonald / Beta Power | `mc` | `gamma`, `delta`, `lambda` | `alpha` = 1, `beta` = 1 | 3 |
-  | Kumaraswamy | `kw` | `alpha`, `beta` | `gamma` = 1, `delta` = 0, `lambda` = 1 | 2 |
-  | Beta | `beta` | `gamma`, `delta` | `alpha` = 1, `beta` = 1, `lambda` = 1 | 2 |
-
-* **Advanced Regression Modeling (`gkwreg`):** Independently model each relevant distribution parameter as a function of covariates using a flexible formula interface:
-  ```r
-  y ~ alpha_terms | beta_terms | gamma_terms | delta_terms | lambda_terms
-  ```
-
-* **Multiple Link Functions:** Choose appropriate link functions for each parameter, including:
[…]
