### 2025-07-13 6b02b23
Remove simplexreg package
