+++ man/gkwreg.Rd
@@ -295,34 +295,7 @@ kw_model_gas <- gkwreg(formu_gy, gasoline_data,
 summary(kw_model_gas)
 plot(kw_model_gas, use_ggplot = TRUE, arrange_plots = TRUE, sub.caption = "")
 
-## Example 3: SDAC Cancer Data
-# Load cancer survival dataset
-sdac_data <- get_bounded_datasets("sdac")
-
-# Formula: relative cumulative density ~ age adjustment + chemotherapy
-formu_sd <- rcd ~ ageadj + chemo
-
-# Fit Extended Kumaraswamy model
-ekw_model_gas <- gkwreg(formu_sd, sdac_data, family = "ekw", method = "BFGS")
-summary(ekw_model_gas)
-plot(ekw_model_gas, use_ggplot = TRUE, arrange_plots = TRUE, sub.caption = "")
-
-## Example 4: Retinal Data
-# Load retinal dataset
-retinal_data <- get_bounded_datasets("retinal")
-
-# Formula for three parameters with different predictors
-# alpha ~ LogT + LogT2 + Level
-# beta  ~ LogT + Level
-# gamma ~ Time
-formu_rt <- Gas ~ LogT + LogT2 + Level | LogT + Level | Time
-
-# Fit Extended Kumaraswamy model
-ekw_model_ret <- gkwreg(formu_rt, retinal_data, family = "ekw", method = "nlminb")
-summary(ekw_model_ret)
-plot(ekw_model_ret, use_ggplot = TRUE, arrange_plots = TRUE, sub.caption = "")
-
-## Example 5: Weather Task Agreement Data
+## Example 3: Weather Task Agreement Data
 # Load the WeatherTask dataset
 df_weather <- get_bounded_datasets("WeatherTask")
 
