+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: gkwreg
 Title: Generalized Kumaraswamy Regression Models for Bounded Data
-Version: 1.0.10
+Version: 1.0.11
 Authors@R: 
     person("Lopes", "J. E.", , "evandeilton@gmail.com", role = c("aut", "cre"))
 Description: Implements regression models for bounded continuous data in the open interval (0,1) using the five-parameter Generalized Kumaraswamy distribution. Supports modeling all distribution parameters (alpha, beta, gamma, delta, lambda) as functions of predictors through various link functions. Provides efficient maximum likelihood estimation via Template Model Builder ('TMB'), offering comprehensive diagnostics, model comparison tools, and simulation methods. Particularly useful for analyzing proportions, rates, indices, and other bounded response data with complex distributional features not adequately captured by simpler models.
@@ -12,7 +12,7 @@ Imports: Formula, stats, graphics, TMB, Rcpp, RcppArmadillo, RcppEigen,
         magrittr, ggplot2, ggpubr, gridExtra, rappdirs, patchwork,
         fmsb, scales, tidyr, reshape2, numDeriv
 LinkingTo: Rcpp, RcppArmadillo, RcppEigen
-Suggests: betareg, knitr, rmarkdown, simplexreg, testthat (>= 3.0.0)
+Suggests: betareg, knitr, rmarkdown, testthat (>= 3.0.0)
 Config/testthat/edition: 3
 NeedsCompilation: yes
 Packaged: 2025-07-08 23:44:58 UTC; jlopes
