### 2025-07-13 0e2382e
Update paper text
