### 2025-07-18 #8571 @crvernon
@evandeilton - we need the paper word count down to <= 1000 words and it must include a "Statement of Need" section.  Please revisit the submission requirements if needed.
