### 2025-07-19 15c8113
Add paper files for JOSS.
