+++ inst/paper/paper.bib
@@ -484,3 +484,11 @@
   year      = {1985},
   address   = {Oxford}
 }
+
+
+@Manual{gkwreg,
+  title = {gkwreg: An R Package for Generalized Kumaraswamy Regression Models for Bounded Data},
+  author = {Lopes, JE},
+  year = {2025},
+  note = {R package version 1.0.11},
+}
