+++ inst/paper/paper.md
@@ -22,7 +22,7 @@ affiliations:
  # - name: Universidade Federal do Paraná (UFPR)
  #   index: 2
 citation_author: Lopes, JE & Bonat, WH
-date: '10 julho 2025'
+date: '19 julho 2025'
 year: '2025'
 bibliography: paper.bib
 output: rticles::joss_article
@@ -42,815 +42,27 @@ header-includes:
   # - \usepackage{threeparttable} # Uncomment if necessary
 ---
 
-
-
-
-
-
-
-
 # Summary
 
-Modeling data restricted to the unit interval $(0,1)$, such as proportions and rates, is a common challenge. The beta distribution, although predominant, has limitations regarding flexibility and analytical tractability (CDF), motivating alternatives, especially in regression contexts. The Kumaraswamy (Kw) distribution offers analytical simplicity (closed-form CDF/PDF), but it may have insufficient flexibility. In this context, the Generalized Kumaraswamy (GKw) distribution, a five-parameter flexible model nesting Beta, Kw, and other relevant distributions, emerges as a promising alternative. However, the complexity of its likelihood function makes maximum likelihood estimation (MLE) difficult, particularly in multiparametric regression. This article introduces the $\textbf{R}$ package `gkwreg`, developed for efficient (MLE) estimation and inference in GKw regression models and its seven main subfamilies. `gkwreg` uses the robust Template Model Builder (`TMB`) framework, employing automatic differentiation for fast and accurate gradient/Hessian computation and effective optimization algorithms. This integrated approach via `TMB` enables reliable, robust, and efficient estimation of complex models, even when multiple parameters are linked to covariates by various link functions. The package also includes routines in `RcppArmadillo` and `C++` that implement distribution functions `p`, `d`, `r`, and `q`, as well as their respective log-likelihood functions, gradients, and analytic Hessians for the seven GKw subfamilies and a function for fitting distributions to data. We detail the statistical foundations of the GKw family, its hierarchy, the regression framework, inference, diagnostics, and the package’s computational implementation. Real data examples demonstrate `gkwreg`’s ability to capture complex distributional patterns that simpler models may overlook. The package thus provides a unified, flexible, and computationally efficient tool for analyzing responses restricted to the unit interval.
+`gkwreg` is an R package for fitting regression models to data restricted to the unit interval $(0,1)$, such as proportions, rates, and indices. The package implements the flexible five-parameter Generalized Kumaraswamy (GKw) distribution and its seven main subfamilies, including the widely used Beta and Kumaraswamy distributions. A key feature of `gkwreg` is its use of the Template Model Builder (`TMB`) framework, which leverages automatic differentiation and C++ templates **to provide fast, stable, and accurate maximum likelihood estimation**. This overcomes the significant computational challenges typically associated with such complex multiparametric models, making them accessible for practical application. The package provides a user-friendly interface with standard R methods for model specification, inference, and diagnostics.
 
+# Statement of need
 
-# Introduction
+Statistical modeling of data bounded in the interval $(0,1)$ is frequent across fields such as economics, epidemiology, and social sciences. Traditional methods like variable transformations followed by linear regression often present interpretability issues and fail near boundary points.
 
[…]
