+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: gkwreg
 Title: Generalized Kumaraswamy Regression Models for Bounded Data
 Version: 1.0.11
 Authors@R: 
-    person("Lopes", "J. E.", , "evandeilton@gmail.com", role = c("aut", "cre"))
+    person("Lopes, J.E.", "evandeilton@gmail.com", role = c("aut", "cre"))
 Description: Implements regression models for bounded continuous data in the open interval (0,1) using the five-parameter Generalized Kumaraswamy distribution. Supports modeling all distribution parameters (alpha, beta, gamma, delta, lambda) as functions of predictors through various link functions. Provides efficient maximum likelihood estimation via Template Model Builder ('TMB'), offering comprehensive diagnostics, model comparison tools, and simulation methods. Particularly useful for analyzing proportions, rates, indices, and other bounded response data with complex distributional features not adequately captured by simpler models.
 License: MIT + file LICENSE
 Encoding: UTF-8
