### 2025-08-21 #8571 @jbytecode
@evandeilton - Thank you for submitting your work to JOSS, I'm the handling editor of your submission. First, I'll try to invite the suggested reviewers by our editorialbot. Meanwhile, you can perform these operations:

- The Mathematics section in the manuscript looks strange. It could be more convenient, e.g., Mathematical Background?
- In the manuscript, methods like p*, q*, r*, d*, are short for, for example pnorm, qnorm, rnorm, dnorm, right? This would be more expressive, for example, the methods for probability (pgkw), quantile (qgkw), random numbers (rgkw), and distribution (dgkw) .... are implemented... is a more expressive one
- After inviting suitable reviewers, we will assign at least two reviewers from those who responded and accepted. You can also use the list of people link provided at the top of this page to search and filter suitable reviewers. If I fail to assign reviewers in a reasonable time, I'll ask you if you have any suggestions.

Thank you in advance!
