### 2025-08-29 #8571 @jbytecode
👋👋👋 Dear @strengejacke 👋👋👋

**[[PRE REVIEW]: gkwreg: An R Package for Generalized Kumaraswamy Regression Models for Bounded Data](https://github.com/openjournals/joss-reviews/issues/8571#top)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/8571)).

Thank you in advance!
