### 2025-08-29 #8571 @strengejacke
I guess the paper is a bit beyond my expertise, as I'm struggling to fully understand the formulas in the _Mathematics_ section. Since this is the main and only section, I cannot judge the paper's content very well. I could look at the other technical aspects of the repository, if that's enough, else I'd suggest you may try to find another reviewer.
