### 2025-10-02 #8991 @evandeilton
Hi @jbytecode, please use gkwreg, Version 1.0.11 from GitHub. The DOI issues have been fixed in this version.
Thank you in advance.
