### 2025-10-03 #8991 @jbytecode
@evandeilton - I'm sorry if I'm missing something. What does, 'Please use gkwreg, Version 1.0.11 from GitHub' mean? I need to clarify that we don't need to use that, or any other software, to check the references. The task is automated: the editorial bot reads the .bib file in the defined branch and checks whether the DOIs (Digital Object Identifiers) are correct, incorrect, or missing.

Please, to go further, correct the DOI's as the editorialbot suggests, and let them pass the usual checks.

Thank you.
