+++ inst/paper/paper.bib
@@ -1,39 +1,12 @@
-@article{Akaike1974,
-  author = {Akaike, H.},
-  title = {A new look at the statistical model identification},
-  journal = {IEEE Transactions on Automatic Control},
-  volume = {19},
-  number = {6},
-  pages = {716--723},
-  year = {1974},
-  doi = {10.1109/TAC.1974.1100705}
-}
-
-@article{Bell2007,
-  author = {Bell, B. M.},
-  title = {CppAD: A Package for C++ Algorithmic Differentiation},
-  journal = {Computational Infrastructure for Operations Research},
-  year = {2007},
-  url = {http://www.coin-or.org/CppAD}
-}
-
-@article{branscum2007,
-  author = {Branscum, A. J. and Johnson, W. O. and Thurmond, M. C.},
-  title = {Bayesian beta regression: Applications to household expenditure data and genetic distance between foot-and-mouth disease viruses},
-  journal = {Australian \& New Zealand Journal of Statistics},
-  volume = {49},
-  number = {3},
-  pages = {287--301},
-  year = {2007},
-  doi = {10.1111/j.1467-842X.2007.00481.x}
-}
-
-@book{Burnham2002,
-  author = {Burnham, K. P. and Anderson, D. R.},
-  title = {Model Selection and Multimodel Inference: A Practical Information-Theoretic Approach},
-  publisher = {Springer},
-  year = {2002},
-  edition = {2}
+@article{kumaraswamy1980,
+  author = {Kumaraswamy, P.},
+  title = {A generalized probability density function for double-bounded random processes},
+  journal = {Journal of Hydrology},
+  volume = {46},
+  number = {1-2},
+  pages = {79--88},
+  year = {1980},
+  doi = {10.1016/0022-1694(80)90036-0}
 }
 
 @article{carrasco2010,
@@ -43,156 +16,11 @@
   year = {2010}
 }
 
-@article{Cordeiro2011,
-  author = {Cordeiro, G. M. and de Castro, M.},
-  title = {A new family of generalized distributions},
-  journal = {Journal of Statistical Computation and Simulation},
-  volume = {81},
-  number = {7},
-  pages = {883--898},
-  year = {2011},
-  doi = {10.1080/00949650903530745}
-}
-
-@article{cribari2010,
-  author = {Cribari-Neto, F. and Zeileis, A.},
-  title = {Beta Regression in {R}},
-  journal = {Journal of Statistical Software},
-  volume = {34},
-  number = {2},
-  pages = {1--24},
-  year = {2010},
-  doi = {10.18637/jss.v034.i02}
-}
-
-@article{Dunn1996,
-  author = {Dunn, P. K. and Smyth, G. K.},
-  title = {Randomized Quantile Residuals},
-  journal = {Journal of Computational and Graphical Statistics},
-  volume = {5},
-  number = {3},
-  pages = {236--244},
-  year = {1996},
-  doi = {10.1080/10618600.1996.10474708}
-}
-
-@article{dutka1981,
-  author = {Dutka, J.},
-  title = {The incomplete beta function—a historical profile},
-  journal = {Archive for History of Exact Sciences},
-  volume = {24},
-  number = {1},
-  pages = {11--29},
-  year = {1981},
-  doi = {10.1007/BF00327713}
-}
-
-@article{Eddelbuettel2014,
-  author = {Eddelbuettel, D. and Sanderson, C.},
-  title = {RcppArmadillo: Accelerating {R} with high-performance {C++} linear algebra},
-  journal = {Computational Statistics and Data Analysis},
-  volume = {71},
-  pages = {1054--1063},
-  year = {2014},
[…]
