### 2025-10-06 #8991 @jbytecode
@JonasMoss, @faosorios - You can now start your review by typing

`@editorialbot generate my checklist`

This command will create your own task list. Thank you in advance.
