### 2025-10-06 #8991 @JonasMoss
This looks very impressive, @evandeilton! I think this package has the potential to be very useful and be a nice companion to betareg, and I can definitely see myself using it. I am currently reading through the code, and there's *a lot* of code here! This of course makes the reviewing task take longer time and be more difficult.

In any case, the major problem with your package, per the checklist above, is right now _the lack of tests_.

The second problem is the size of the package. It is very, _very_ big right now, actually overwhelming. While I definitely **won't try to deny a publication** here due to these size problems, I think you might consider trimming the package down a great deal along with adding tests, essentially making it more user-friendly. You have implemented a lot of functionality, and there's a risk of feature creep. My thoughts can be split into:

1. **Shrink the public surface:** Keep the generics (predict/summary/coef/vcov/logLik/AIC/BIC/residuals/plot/confint), probably  without documentation (they should be self-documenting), move the rest internal. Push knobs into a single control list to make the interface for the public-facing functions tidier. Consider removing redundant features such as plot and confint = TRUE in the main functions. Possibly remove handy but not really major utility functions such as `gkwfitall`.
2. **Remove internal functions that are not used in external functions.** At the very least move them to their own files, but removal is probably the best option.
3. **Split density functions into a different package?** These clutter your namespace and are not the main contribution of your package. You can probably make a PR to extraDistr or make another package. I realize this is asking a lot, but I think it's a good idea. These functions are potentially of interested even if you don't need the whole regression setup, and should ideally be available without that burden.
4. **Split up the files.** The R files are many thousands of lines long right now. They could be split up into say, generics, internals, and public-facing API.
5. **Remove clutter.** Remove commented-out code (in several files, both R and c++). There are too many `rd` files, remove them by putting `@noRD` https://style.tidyverse.org/documentation.html Moreover, many of the examples are *too large*. In general there's quite a bit of clutter in the source files.

I think the [principle of least surprise](https://en.wikipedia.org/wiki/Principle_of_least_astonishment) (which you follow well, but not 100%) would be a good guideline here. Say, make the functions mimic the GLM functions, with full generic support such as confint taking care of confidence intervals etc. But deally with even less complexity. Remove the huge number of knobs, and make sensible choices for the user, e.g., nlminb could be hard-coded. KISS is another possible guideline.

To be honest, if you clean up the exported functions and the docs to make it more focused and less overwhelming, I think that Journal of Statistical Software could be a good venue for a more fleshed-out software paper after JOSS, especially if you can make the case this is sometimes a substantial improvement on existing regression methods.

## Some comments
**Main takeaway.** There is too much stuff that takes focus away from your main contributions `gkreg` (and `gkfit`  I take it? though this is of much smaller importance). Many functions have arguably too many arguments (it woud be a good idea to add a `control` argument as in e.g. `?betareg::betareg` or `optim`, even multiple controls). It's not immediately clear wether to use `nrgkw` or the fit function. The documentation is full of irrelevant functions, and some functions (`gkwfitall`) have, perhaps, suboptimal design choices.

* The documentation looks very good from a cursory glance, both in terms of details and examples. I will read through it later on.
* There are a lot of internal functions, which is fine, but some of them are plotting and diagnostics functions. These look like either  (a) functions that should be removed. (b) maybe functions that should be exported (Example: ".plot_gkwreg_ggplot".)
* The size of the code files are absolutely enormous, and they mix internals and exported functions haphazardly. It would be good to split them up more I'd say.
* In gkwreg you mix up parameter and data at least once? ("Parameter in (0,1)" should probably be data in (0,1) or maybe [0,1]?).
* "Conf.level" is an argument to gkwreg; I'd suggest you remove this, it can be handled with `confint` anyway? `plot` is also superfluous. All in all, `gkwreg` is too complex, and would be better with way fewer arguments I'd say.
* Ah, `conf.int` isn't there for `gkwreg`; you might mimic confint.lm here.
* `print.gkwreg` isn't implemented.
* You might consider having fewer exports. This is especially important for R packages because all exported functions are collected into one huge documentation file, making it very hard for users to find the correct functions to use when its overcrowded with noise. This is a problem with how R works (it doesn't allow subpackages which would allow us to organize things better), but is still *strongly* worth keeping in mind.
* For instance, AIC, BIC, loglik, coef, summary, vcov, anova, fitted, predict, print, plot, residuals are generic functions that do not *always* need to have documentation. (Maybe plot and summary?)
* The hessian matrices, gradients, if considered important by themselves (outside being returned by a fitting procedure) can plausibly be merged into one function. I would suggest dropping them from the exports.
* `get_bounded_datasets` is most likely not important enough to be exported.
* The `d`, `p`, `q`, `r` functions are of course standard in R, and can't be avoided even though they make the documentation messy. I would suggest importing the kumaraswamy density from `extraDistr` if possible though, as this package is not mainly about providing such functions. It could also be possible to divide the package in two, one for densities, one for regression and fitting. Or make a PR in extraDistr (which I think is not actively maintained but the author might accept PRs).
* What is the point of `nrgkw` and what's it's relationship to `gkfit`? I think you should tone down "industrial strength" here. You’ve clearly done a huge amount of work here; perhaps you could emphasize what the functions do rather than how they do it.
* `utils_tmb.R` contains only commented code. The function appears to have been placed in zzz.R.
* `gkwmle.R` also contains loads of commented out code. I haven't thoroughly checked every file for this.
* There are too many `rd` files, remove them by putting `@noRD` https://style.tidyverse.org/documentation.html
* It would be nice with some "sanity checks" say vs. the betareg package.
* Do you want to do this? "## usethis namespace: start
#' @importFrom numDeriv grad hessian
## usethis namespace: end
NULL" I suppose it's ... OK, but best practice is surely to do numDeriv::grad etc? I am not even sure you need to use this.

## The main functions
I interpret `gkwreg`, `gkwfit`, `gkwgof`, and `gkfitall` as the main user-facing functions. I understand the need for `gkwfit` and `gkwreg`, but I am not sure `gkwfitall` is really needed, and we do not see that sort of function in other packages. It's also constructed to leave a report? I am not sure about this one -- it's ok, but think about the focus of the package and whether it might be better to have it in a vignette.

Ok, about my feeling of getting overwhelmed: Let's have a look at one of your functions:
```gkwfit(
  data,
  family = "gkw",
  start = NULL,
  fixed = NULL,
  method = "nlminb",
  use_moments = FALSE,
  hessian = TRUE,
  profile = FALSE,
  npoints = 20,
  plot = TRUE,
  conf.level = 0.95,
  optimizer.control = list(),
  submodels = FALSE,
  silent = TRUE,
  ...
)
```
I'd say that plot and conf.level shoud be removed since they violate separation of concerns. Fit the model and get confint with summary / confint and plots with plot. `hessian` should probably be fixed to true; this is more like an internal knob. Profie and submodels can probably also be removed, start and use_moments can be merged, `family` should probably be a vector of strings using `match.arg`, n_points can be removed along with profile (or actually merged with profile if you want to keep it), `fixed` also feels like something internal; your whole deal is fixing different parameters with the generalized Kumaraswamy anyway. `start` can be put into the control list, and we get something like:

```gkw_density_ml(
  x,
  family = c("gkw", "bkw", "kkw", "ekw", "mc", "kw", "beta"),
  method = "nlminb",
  control = list(),
  silent = TRUE
)
```

This function is less flexible but also much less scary and nicer for your inexperienced users (like me!) to use. Now I can just see "ah, generalized Kumaraswamy density maximum likelihood, let's put in some data". I see immediately that only *family* likely matters, and play.
