### 2025-10-07 #8991 @evandeilton
Hi @JonasMoss , thank you for the careful read and the thorough feedback. It clearly highlights how to make the package more useful and predictable.

I admit that after your review, my TODO list is about the size of a CRAN package. Great roadmap! I’ll address everything thoughtfully. I’ll go through each point carefully, especially the lack of tests and the wide public surface to ensure the final solution is lean, GLM-consistent, and robust.

Next, I’ll prioritize building a comprehensive test suite, streamlining/organizing the public API (moving knobs into a `control` list), auditing exports and docs, removing commented code and splitting files, evaluating a separate package/PR for densities, and adding sanity checks against `betareg`.

Thanks again for the time and generosity; your suggestions will be incorporated with care, and I’m happy to discuss design choices as I move forward.
