### 2025-10-15 a40f5b8
PACKAGE FULL REFACTOR AFTER JOSS REVIEW
