+++ R/gkwreg-other-methods.R
@@ -1,0 +1,667 @@
+#' @title Update and Re-fit a GKw Regression Model
+#'
+#' @description
+#' Updates and (by default) re-fits a Generalized Kumaraswamy regression model.
+#' This method allows modification of the model formula, data, or other arguments
+#' without having to completely re-specify the model call. Supports formulas with
+#' up to 5 parts (alpha, beta, gamma, delta, lambda) using the \pkg{Formula} package.
+#'
+#' @param object An object of class \code{"gkwreg"}, typically obtained from
+#'   \code{\link{gkwreg}}.
+#' @param formula. Changes to the formula. This is a formula where \code{.} refers
+#'   to the corresponding part of the old formula. For multi-part formulas
+#'   (e.g., \code{y ~ x1 | x2 | x3}), you can update each part separately using
+#'   the \code{|} separator.
+#' @param data. Optional. A new data frame in which to evaluate the updated model.
+#'   If omitted, the original data is used.
+#' @param ... Additional arguments to the call, or arguments with changed values.
+#'   Use \code{name = NULL} to remove an argument.
+#' @param evaluate Logical. If \code{TRUE} (default), the updated model is fitted.
+#'   If \code{FALSE}, the updated call is returned without fitting.
+#'
+#' @details
+#' The \code{update} method allows you to modify a fitted model and re-fit it
+#' with the changes. The GKw regression model supports formulas with up to 5 parts:
+#' \code{y ~ model_alpha | model_beta | model_gamma | model_delta | model_lambda}
+#'
+#' Each part can be updated independently using \code{.} to refer to the current
+#' specification:
+#' \itemize{
+#'   \item \code{. ~ . + x | . | . | . | .} - Add \code{x} to alpha only
+#'   \item \code{. ~ . | . + x | . | . | .} - Add \code{x} to beta only
+#'   \item \code{. ~ . | . | . + x | . | .} - Add \code{x} to gamma only
+#'   \item \code{. ~ . + x | . + x | . | . | .} - Add \code{x} to alpha and beta
+#'   \item \code{. ~ . - x | . | . | . | .} - Remove \code{x} from alpha
+#' }
+#'
+#' Omitting parts at the end is allowed (they default to \code{.}):
+#' \itemize{
+#'   \item \code{. ~ . + x | .} is equivalent to \code{. ~ . + x | . | . | . | .}
+#'   \item \code{. ~ . | . + x} is equivalent to \code{. ~ . | . + x | . | . | .}
+#' }
+#'
+#' @return If \code{evaluate = TRUE}, a new fitted model object of class
+#'   \code{"gkwreg"}. If \code{evaluate = FALSE}, an updated call.
+#'
+#' @author Lopes, J. E.
+#'
+#' @seealso \code{\link{gkwreg}}, \code{\link[stats]{update}},
+#'   \code{\link[Formula]{Formula}}
+#'
+#' @examples
+#' \donttest{
+#' # Load example data
+#' data(GasolineYield)
+#'
+#' # ===========================================================================
+#' # EXAMPLE 1: Simple formulas (1 part - alpha only)
+#' # ===========================================================================
+#'
+#' m1_0 <- gkwreg(yield ~ 1, data = GasolineYield, family = "kw")
+#' m1_1 <- update(m1_0, . ~ . + temp)
+#' m1_2 <- update(m1_1, . ~ . + batch)
[…]
