+++ R/utils_tmb.R
@@ -1,288 +1,0 @@
-#' #' Check and Compile TMB Model Code with Persistent Cache
-#' #'
-#' #' @description
-#' #' This utility function verifies whether the TMB model shared object (\code{.so/.dll}) file
-#' #' has already been compiled for a specified DLL name. If not, it compiles the
-#' #' corresponding C++ file and caches it in a persistent directory across R sessions.
-#' #'
-#' #' @details
-#' #' The function works through the following steps:
-#' #' \enumerate{
-#' #'   \item Creates a persistent cache directory for storing compiled TMB models.
-#' #'   \item Checks if a compiled file for the specified DLL already exists in the
-#' #'         cache directory and whether it's up-to-date compared to the source code.
-#' #'   \item If a valid compiled file exists, it loads it directly.
-#' #'   \item If not, the function locates the corresponding C++ file inside the
-#' #'         package, compiles it, and stores the result in the cache directory.
-#' #'   \item Provides diagnostic messages regarding compilation status and exported symbols.
-#' #' }
-#' #'
-#' #' @param dll_name A character string specifying the base name of the C++ file
-#' #'   and the resulting DLL. The function assumes the code file is \code{dll_name.cpp}
-#' #'   located in the \code{inst/tmb/} directory of the package.
-#' #' @param pkg_name A character string specifying the package name. Defaults to "gkwreg".
-#' #' @param force_recompile Logical; if \code{TRUE}, forces recompilation even if
-#' #'   a valid compiled file exists (default is \code{FALSE}).
-#' #' @param verbose Logical; if \code{TRUE}, prints detailed status messages
-#' #'   (default is \code{TRUE}).
-#' #'
-#' #' @return
-#' #' Returns (invisibly) a list with information about the compiled model, including path,
-#' #' normalized path, name, and compilation status.
-#' #' If any step fails, an error is thrown.
-#' #'
-#' #' @importFrom TMB compile dynlib
-#' #' @importFrom tools file_path_sans_ext
-#' #' @keywords internal
-#' .check_and_compile_TMB_code <- function(dll_name,
-#'                                         pkg_name = "gkwreg",
-#'                                         force_recompile = FALSE,
-#'                                         verbose = FALSE) {
-#'   # Helper function for logging
-#'   log_msg <- function(...) {
-#'     if (verbose) cat(...)
-#'   }
-#'
-#'   if (verbose) log_msg("Checking TMB model status for ", dll_name, "...\n")
-#'
-#'   # Ensure .gkwreg_env exists
-#'   if (!exists(".gkwreg_env")) {
-#'     .gkwreg_env <- new.env(parent = emptyenv())
-#'   }
-#'
-#'   # 1. Create a persistent cache directory for TMB compilation
-#'   # Use rappdirs if available, otherwise create directory in user's home folder
-#'   if (requireNamespace("rappdirs", quietly = TRUE)) {
-#'     cache_dir <- file.path(rappdirs::user_cache_dir(pkg_name), "tmb_cache")
-#'   } else {
-#'     # Fallback to a directory in the user's home folder
-#'     cache_dir <- file.path(path.expand("~"), ".R", paste0(pkg_name, "_tmb_cache"))
-#'   }
-#'
-#'   if (!dir.exists(cache_dir)) {
-#'     dir.create(cache_dir, recursive = TRUE)
-#'     if (verbose) log_msg("Created persistent cache directory: ", cache_dir, "\n")
[…]
