+++ R/gkwreg-plot.R
@@ -1,0 +1,3567 @@
+#' @title Diagnostic Plots for Generalized Kumaraswamy Regression Models
+#'
+#' @description
+#' Produces a comprehensive set of diagnostic plots for assessing the adequacy
+#' of a fitted Generalized Kumaraswamy (GKw) regression model (objects of class
+#' \code{"gkwreg"}). The function offers flexible plot selection, multiple
+#' residual types, and support for both base R graphics and \code{ggplot2} with
+#' extensive customization options. Designed for thorough model evaluation
+#' including residual analysis, influence diagnostics, and goodness-of-fit
+#' assessment.
+#'
+#' @param x An object of class \code{"gkwreg"}, typically the result of a call
+#'   to \code{\link{gkwreg}}.
+#'
+#' @param which Integer vector specifying which diagnostic plots to produce.
+#'   If a subset of the plots is required, specify a subset of the numbers 1:6.
+#'   Defaults to \code{1:6} (all plots). The plots correspond to:
+#'   \enumerate{
+#'     \item \strong{Residuals vs. Observation Indices}: Checks for temporal
+#'       patterns, trends, or autocorrelation in residuals across observation order.
+#'     \item \strong{Cook's Distance Plot}: Identifies influential observations
+#'       that have disproportionate impact on model estimates. Points exceeding
+#'       the 4/n threshold warrant investigation.
+#'     \item \strong{Generalized Leverage vs. Fitted Values}: Identifies high
+#'       leverage points with unusual predictor combinations. Points exceeding
+#'       2p/n threshold may be influential.
+#'     \item \strong{Residuals vs. Linear Predictor}: Checks for non-linearity
+#'       in the predictor-response relationship and heteroscedasticity
+#'       (non-constant variance).
+#'     \item \strong{Half-Normal Plot with Simulated Envelope}: Assesses
+#'       normality of residuals (particularly useful for quantile residuals)
+#'       by comparing observed residuals against simulated quantiles. Points
+#'       outside the envelope indicate potential model misspecification.
+#'     \item \strong{Predicted vs. Observed Values}: Overall goodness-of-fit
+#'       check showing model prediction accuracy and systematic bias.
+#'   }
+#'
+#' @param type Character string indicating the type of residuals to be used for
+#'   plotting. Defaults to \code{"quantile"}. Valid options are:
+#'   \itemize{
+#'     \item \code{"quantile"}: Randomized quantile residuals (Dunn & Smyth, 1996).
+#'       \strong{Recommended for bounded responses} as they should be approximately
+#'       N(0,1) if the model is correctly specified. Most interpretable with
+#'       standard diagnostic tools.
+#'     \item \code{"pearson"}: Pearson residuals (response residual standardized
+#'       by estimated standard deviation). Useful for checking the variance function
+#'       and identifying heteroscedasticity patterns.
+#'     \item \code{"deviance"}: Deviance residuals. Related to the log-likelihood
+#'       contribution of each observation. Sum of squared deviance residuals equals
+#'       the model deviance.
+#'   }
+#'
+#' @param family Character string specifying the distribution family assumptions
+#'   to use when calculating residuals and other diagnostics. If \code{NULL}
+#'   (default), the family stored within the fitted \code{object} is used.
+#'   Specifying a different family can be useful for diagnostic comparisons
+#'   across competing model specifications. Available options match those in
[…]

[…]
