+++ R/zzz.R
@@ -458,6 +458,7 @@ NULL
 #' @importFrom Rcpp sourceCpp evalCpp
 #' @import RcppArmadillo
 #' @import graphics
+#' @import gkwdist
 ## usethis namespace: end
 NULL
 
