+++ pkgdown.yml
@@ -1,0 +1,8 @@
+pandoc: 3.1.11
+pkgdown: 2.1.3
+pkgdown_sha: ~
+articles: {}
+last_built: 2025-10-15T12:48Z
+urls:
+  reference: https://evandeilton.github.io/gkwreg/reference
+  article: https://evandeilton.github.io/gkwreg/articles
