+++ R/gkwreg-inference.R
@@ -380,503 +380,3 @@ BIC.gkwreg <- function(object, ...) {
 
   return(-2 * as.numeric(ll) + df * log(nobs))
 }
-
-
-
-
-
-
-
-
-
-
-
-
-#'
-#' #' @title Extract Log-Likelihood from a Generalized Kumaraswamy Regression Model
-#' #'
-#' #' @description
-#' #' This function extracts the maximized log-likelihood value from a fitted Generalized
-#' #' Kumaraswamy (GKw) regression model object (class \code{"gkwreg"}). The result is
-#' #' returned as an object of class \code{"logLik"}, which includes attributes for
-#' #' degrees of freedom and number of observations, suitable for use with model
-#' #' selection criteria like AIC and BIC.
-#' #'
-#' #' @param object An object of class \code{"gkwreg"}, typically the result of a call
-#' #'   to \code{\link{gkwreg}}.
-#' #' @param ... Additional arguments, currently ignored by this method.
-#' #'
-#' #' @details
-#' #' The log-likelihood value is typically computed during the model fitting process
-#' #' (e.g., by \code{\link{gkwreg}}) and stored within the resulting object. This
-#' #' method retrieves this stored value. If the value is not directly available, it
-#' #' attempts to calculate it from the stored deviance (\eqn{logLik = -deviance / 2}).
-#' #'
-#' #' The log-likelihood for a GKw family model with parameters \eqn{\theta} is
-#' #' generally defined as the sum of the log-density contributions for each observation:
-#' #' \deqn{l(\theta | y) = \sum_{i=1}^n \log f(y_i; \alpha_i, \beta_i, \gamma_i, \delta_i, \lambda_i)}
-#' #' where \eqn{f(y; \dots)} is the probability density function (PDF) of the specific
-#' #' distribution from the GKw family used in the model (determined by the \code{family}
-#' #' argument in \code{gkwreg}), and parameters (\eqn{\alpha_i, \dots, \lambda_i}) may
-#' #' depend on covariates.
-#' #'
-#' #' The function also extracts the number of estimated parameters (\code{df}) and the
-#' #' number of observations (\code{nobs}) used in the fit, storing them as attributes
-#' #' of the returned \code{"logLik"} object, which is essential for functions like
-#' #' \code{\link[stats]{AIC}} and \code{\link[stats]{BIC}}. It attempts to find \code{df}
-#' #' and \code{nobs} from various components within the \code{object} if they are not
-#' #' directly stored as \code{npar} and \code{nobs}.
-#' #'
-#' #' @return An object of class \code{"logLik"} representing the maximized
-#' #'   log-likelihood value. It has the following attributes:
-#' #'   \itemize{
-#' #'     \item \code{df}: (numeric) The number of estimated parameters in the model
-#' #'       (coefficients).
-#' #'     \item \code{nobs}: (numeric) The number of observations used for fitting the model.
-#' #'   }
-#' #'
-#' #' @author Lopes, J. E.
-#' #'
-#' #' @seealso \code{\link{gkwreg}}, \code{\link{AIC.gkwreg}}, \code{\link{BIC.gkwreg}},
-#' #'   \code{\link[stats]{logLik}}, \code{\link[stats]{AIC}}, \code{\link[stats]{BIC}}
-#' #'
-#' #' @keywords log-likelihood models likelihood
-#' #'
-#' #' @examples
-#' #' \donttest{
[…]
