+++ R/gkwreg-methods.R
@@ -731,254 +731,3 @@ confint.gkwreg <- function(object, parm, level = 0.95, ...) {
 
   return(ci)
 }
-
-
-
-#'
-#' #' Confidence Intervals for Generalized Kumaraswamy Regression Parameters
-#' #'
-#' #' Computes confidence intervals for model parameters in fitted \code{gkwreg}
-#' #' objects. Supports Wald (normal approximation) intervals based on asymptotic
-#' #' theory. Follows the same interface as \code{\link[stats]{confint.lm}} and
-#' #' \code{\link[stats]{confint.glm}}.
-#' #'
-#' #' @param object An object of class \code{"gkwreg"} from \code{\link{gkwreg}}.
-#' #' @param parm Optional specification of which parameters to compute intervals for.
-#' #'   Can be \code{NULL} (default, intervals for all parameters), a character vector
-#' #'   of parameter names or patterns, or a numeric vector of parameter indices.
-#' #' @param level Numeric. The confidence level required. Must be between 0 and 1.
-#' #'   Default is 0.95 (95\% confidence intervals).
-#' #' @param ... Additional arguments (currently unused).
-#' #'
-#' #' @details
-#' #' \strong{Wald Intervals:}
-#' #'
-#' #' The confidence intervals are computed using the Wald method based on
-#' #' asymptotic normality of maximum likelihood estimators:
-#' #' \deqn{CI = \hat{\theta} \pm z_{\alpha/2} \times SE(\hat{\theta})}
-#' #' where \eqn{z_{\alpha/2}} is the appropriate normal quantile and
-#' #' \eqn{SE(\hat{\theta})} is the standard error from the Hessian matrix.
-#' #'
-#' #' These intervals are:
-#' #' \itemize{
-#' #'   \item Fast to compute (no refitting required)
-#' #'   \item Appropriate for large samples
-#' #'   \item Valid when likelihood is approximately normal
-#' #' }
-#' #'
-#' #' \strong{Requirements:}
-#' #'
-#' #' The model must have been fitted with \code{hessian = TRUE} (the default)
-#' #' in \code{\link{gkw_control}}. If standard errors are not available, an
-#' #' error is raised with instructions to refit the model.
-#' #'
-#' #' \strong{Parameter Selection:}
-#' #'
-#' #' The \code{parm} argument allows flexible parameter selection:
-#' #' \itemize{
-#' #'   \item \code{NULL}: All parameters (default)
-#' #'   \item Character vector: Exact parameter names (e.g., \code{"alpha:(Intercept)"})
-#' #'   \item Numeric vector: Parameter indices (e.g., \code{1:3})
-#' #' }
-#' #'
-#' #' \strong{Flexibility:}
-#' #'
-#' #' Unlike the deprecated \code{conf.level} argument in older versions of
-#' #' \code{gkwreg()}, this function can compute intervals at any confidence
-#' #' level without refitting the model. Multiple confidence levels can be
-#' #' obtained efficiently.
-#' #'
-#' #' @return A matrix with columns giving lower and upper confidence limits
-#' #'   for each parameter. Row names correspond to parameter names. Column names
-#' #'   indicate the confidence level (e.g., "2.5 \%", "97.5 \%").
-#' #'
-#' #' @examples
-#' #' \donttest{
-#' #' # Fit model
-#' #' data(GasolineYield)
-#' #' fit <- gkwreg(yield ~ batch + temp, data = GasolineYield, family = "kw")
-#' #'
-#' #' # Default: 95\% intervals for all parameters
[…]
