+++ README.Rmd
@@ -1,0 +1,570 @@
+---
+output: github_document
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  echo = TRUE,
+  eval = FALSE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
+)
+```
+
+
+# gkwreg: Generalized Kumaraswamy Regression Models for Bounded Data <img src="man/figures/gkwreg.png" align="right" height="138" />
+
+[![CRAN status](https://www.r-pkg.org/badges/version/gkwreg)](https://CRAN.R-project.org/package=gkwreg)
+[![R-CMD-check](https://github.com/evandeilton/gkwreg/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/evandeilton/gkwreg/actions/workflows/R-CMD-check.yaml)
+[![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/gkwreg)](https://cran.r-project.org/package=gkwreg)
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+
+## Overview
+
+The **gkwreg** package provides a comprehensive and computationally efficient framework for regression modeling of data restricted to the standard unit interval (0, 1), including proportions, rates, fractions, percentages, and bounded indices. While Beta regression is the traditional approach for such data, **gkwreg** focuses on the **Generalized Kumaraswamy (GKw)** distribution family, offering exceptional flexibility by encompassing seven important bounded distributions—including Beta and Kumaraswamy—as special or limiting cases.
+
+The package enables full **distributional regression**, where *all* relevant parameters can be modeled as functions of covariates through flexible link functions. Maximum Likelihood estimation is performed efficiently via the **Template Model Builder (TMB)** framework, leveraging **Automatic Differentiation (AD)** for superior computational speed, numerical accuracy, and optimization stability.
+
+---
+
+## Key Features
+
+### 📊 Flexible Distribution Hierarchy
+
+Model bounded data using the **5-parameter Generalized Kumaraswamy (GKw)** distribution and its **seven nested subfamilies**:
+
+| Distribution                | Code   | Parameters Modeled                        | Fixed Parameters                         | # Params |
+|:----------------------------|:-------|:------------------------------------------|:-----------------------------------------|:---------|
+| Generalized Kumaraswamy     | `gkw`  | α, β, γ, δ, λ                            | None                                     | 5        |
+| Beta-Kumaraswamy            | `bkw`  | α, β, γ, δ                               | λ = 1                                    | 4        |
+| Kumaraswamy-Kumaraswamy     | `kkw`  | α, β, δ, λ                               | γ = 1                                    | 4        |
+| Exponentiated Kumaraswamy   | `ekw`  | α, β, λ                                  | γ = 1, δ = 0                             | 3        |
+| McDonald (Beta Power)       | `mc`   | γ, δ, λ                                  | α = 1, β = 1                             | 3        |
[…]

[…]
