+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-10-15T12:57Z
+last_built: 2025-10-15T17:23Z
 urls:
   reference: https://evandeilton.github.io/gkwreg/reference
   article: https://evandeilton.github.io/gkwreg/articles
