+++ README.Rmd
@@ -168,12 +168,12 @@ x1 <- runif(n, -2, 2)
 x2 <- rnorm(n)
 
 # True parameters (log link)
-alpha_true <- exp(0.8 + 0.3*x1)
-beta_true  <- exp(1.2 - 0.2*x2)
+alpha_true <- exp(0.8 + 0.3 * x1)
+beta_true <- exp(1.2 - 0.2 * x2)
 
 # Generate response from Kumaraswamy distribution
 y <- rkw(n, alpha = alpha_true, beta = beta_true)
-y <- pmax(pmin(y, 1 - 1e-7), 1e-7)  # Ensure strict bounds
+y <- pmax(pmin(y, 1 - 1e-7), 1e-7) # Ensure strict bounds
 df <- data.frame(y = y, x1 = x1, x2 = x2)
 
 # Fit Kumaraswamy regression
@@ -194,25 +194,27 @@ newdata <- data.frame(
 )
 
 # Predict different quantities
-pred_mean  <- predict(fit_kw, newdata, type = "response")    # E(Y|X)
-pred_var   <- predict(fit_kw, newdata, type = "variance")    # Var(Y|X)
-pred_alpha <- predict(fit_kw, newdata, type = "alpha")       # α parameter
-pred_params <- predict(fit_kw, newdata, type = "parameter")  # All parameters
+pred_mean <- predict(fit_kw, newdata, type = "response") # E(Y|X)
+pred_var <- predict(fit_kw, newdata, type = "variance") # Var(Y|X)
+pred_alpha <- predict(fit_kw, newdata, type = "alpha") # α parameter
+pred_params <- predict(fit_kw, newdata, type = "parameter") # All parameters
 
 # Evaluate density at y = 0.5 for each observation
 dens_values <- predict(fit_kw, newdata, type = "density", at = 0.5)
 
 # Compute quantiles (10th, 50th, 90th percentiles)
-quantiles <- predict(fit_kw, newdata, type = "quantile", 
-                     at = c(0.1, 0.5, 0.9), elementwise = FALSE)
+quantiles <- predict(fit_kw, newdata,
+  type = "quantile",
+  at = c(0.1, 0.5, 0.9), elementwise = FALSE
+)
 ```
 
 ### Model Comparison
 
 ```{r}
 # Fit nested models
-fit0 <- gkwreg(y ~ 1,      data = df, family = "kw")  # Null model
-fit1 <- gkwreg(y ~ x1,     data = df, family = "kw")  # + x1
+fit0 <- gkwreg(y ~ 1, data = df, family = "kw") # Null model
+fit1 <- gkwreg(y ~ x1, data = df, family = "kw") # + x1
 fit2 <- gkwreg(y ~ x1 | x2, data = df, family = "kw") # + x2 on beta
 
 # Information criteria comparison
@@ -230,25 +232,27 @@ par(mfrow = c(3, 2))
 plot(fit_kw, ask = FALSE)
 
 # Select specific plots with customization
-plot(fit_kw, 
-     which = c(2, 5, 6),           # Cook's distance, Half-normal, Pred vs Obs
-     type = "quantile",             # Quantile residuals (recommended)
-     caption = list(
-       "2" = "Influential Points",
-       "5" = "Distributional Check"
-     ),
-     nsim = 200,                    # More accurate envelope
-     level = 0.95)                  # 95% confidence
+plot(fit_kw,
+  which = c(2, 5, 6), # Cook's distance, Half-normal, Pred vs Obs
+  type = "quantile", # Quantile residuals (recommended)
+  caption = list(
+    "2" = "Influential Points",
+    "5" = "Distributional Check"
+  ),
+  nsim = 200, # More accurate envelope
[…]
