+++ NEWS.md
@@ -1,0 +1,79 @@
+# gkwreg 2.0.0
+
+## Major Changes
+
+### Package Restructuring
+
+* **Complete package reformulation** following JOSS reviewer feedback to reduce complexity and improve maintainability.
+
+* **Distribution functions moved to separate package** `gkwdist`: All `d*`, `p*`, `q*`, `r*` density/CDF/quantile/random generation functions have been extracted to the companion package `gkwdist` for cleaner namespace and reduced dependencies. The `gkwreg` package now focuses exclusively on regression modeling.
+
+* **Univariate fitting functions removed**: `gkwfit()`, `gkwgof()`, and `gkwfitall()` have been removed to maintain package focus on regression. Users needing univariate distribution fitting should use the `gkwdist` package directly or standard MLE approaches.
+
+### Simplified Interface
+
+* **Introduced `gkw_control()`**: All technical/optimization parameters (method, start, fixed, hessian, maxit, tolerances, etc.) are now consolidated in a dedicated control function following the `glm.control()` design pattern. This dramatically simplifies the main `gkwreg()` interface.
+
+* **Removed arguments violating separation of concerns** from `gkwreg()`: 
+  - `plot` argument removed (use `plot()` method instead)
+  - `conf.level` argument removed (use `confint()` method instead)
+  - `profile`, `submodels`, `npoints` arguments removed (focused functionality)
+
+* **Streamlined `gkwreg()` signature**: Reduced from 15+ arguments to ~12 core arguments, with technical options delegated to `control`.
+
+### Complete S3 Method Implementation
+
+* **Standard methods suite**: Implemented complete S3 methods following R conventions:
+  - `print.gkwreg()`, `summary.gkwreg()`, `print.summary.gkwreg()`
+  - `coef.gkwreg()`, `vcov.gkwreg()`, `fitted.gkwreg()`
+  - `logLik.gkwreg()`, `AIC.gkwreg()`, `BIC.gkwreg()`, `nobs.gkwreg()`
+  - `confint.gkwreg()`, `residuals.gkwreg()`, `predict.gkwreg()`
+  - `anova.gkwreg()`, `print.anova.gkwreg()`, `lrtest()`
+
+### Enhanced Diagnostics
+
+* **Comprehensive `plot.gkwreg()` method** with 6 diagnostic plot types:
+  1. Residuals vs Observation Indices
+  2. Cook's Distance
+  3. Generalized Leverage vs Fitted Values
+  4. Residuals vs Linear Predictor
+  5. Half-Normal Plot with Simulated Envelope
+  6. Predicted vs Observed Values
+
+* **Dual graphics system support**: Base R graphics (default) or ggplot2 with automatic grid arrangement via `gridExtra`/`ggpubr`.
+
+* **Advanced customization**: Named-list interface for plot captions (partial customization without repeating all titles), theme control, sampling for large datasets.
+
+### Powerful Prediction
+
+* **`predict.gkwreg()` with 9 prediction types**:
+  - `"response"`, `"variance"`, `"link"`, `"parameter"`
+  - Individual parameters: `"alpha"`, `"beta"`, `"gamma"`, `"delta"`, `"lambda"`
+  - Distribution functions: `"density"`, `"probability"`, `"quantile"`
+
+* **Element-wise and vectorized modes**: Flexible evaluation via `elementwise` argument for CDF/PDF/quantile calculations.
+
+### Model Comparison Tools
+
+* **Likelihood ratio tests**: `anova.gkwreg()` for comparing nested models with automatic ordering and chi-squared tests; dedicated `lrtest()` function for pairwise comparisons.
+
+* **Information criteria**: `AIC.gkwreg()` and `BIC.gkwreg()` with multi-model comparison support returning data frames.
+
[…]

[…]
