+++ R/gkwreg-control.R
@@ -25,11 +25,13 @@
 #'   matrix. Set to \code{FALSE} for faster fitting when standard errors are not needed.
 #'
 #' @param maxit Integer. Maximum number of iterations for the optimizer.
-#'   Default 500. Increase for difficult optimization problems.
+#'   Default 500 for derivative-based methods, 10000 for SANN.
+#'   Increase for difficult optimization problems.
 #'
 #' @param reltol Numeric. Relative convergence tolerance for the optimizer.
 #'   Default \code{sqrt(.Machine$double.eps)} approx. 1.5e-8. Smaller values require
-#'   tighter convergence but may increase computation time.
+#'   tighter convergence but may increase computation time. Used by Nelder-Mead,
+#'   BFGS, and CG methods.
 #'
 #' @param abstol Numeric. Absolute convergence tolerance. Default 0.
 #'   Used by some optimization methods as an additional stopping criterion.
@@ -38,7 +40,7 @@
 #'   \itemize{
 #'     \item 0: Silent (default)
 #'     \item 1: Print iteration progress
-#'     \item 2+: Print detailed diagnostic information
+#'     \item 2+: Print detailed diagnostic information (up to 6 for L-BFGS-B)
 #'   }
 #'   Ignored if \code{silent = TRUE}.
 #'
@@ -65,6 +67,56 @@
 #'   Default \code{sqrt(.Machine$double.eps)}. Alternative specification of
 #'   relative tolerance.
 #'
+#' @param alpha Numeric. Reflection factor for Nelder-Mead method. Default 1.0.
+#'   Only used when \code{method = "Nelder-Mead"}.
+#'
+#' @param beta Numeric. Contraction factor for Nelder-Mead method. Default 0.5.
+#'   Only used when \code{method = "Nelder-Mead"}.
+#'
+#' @param gamma Numeric. Expansion factor for Nelder-Mead method. Default 2.0.
+#'   Only used when \code{method = "Nelder-Mead"}.
+#'
+#' @param type Integer. Update formula for CG method. Options:
+#'   \itemize{
+#'     \item 1: Fletcher-Reeves update
+#'     \item 2: Polak-Ribiere update
+#'     \item 3: Beale-Sorenson update
+#'   }
+#'   Default 1. Only used when \code{method = "CG"}.
+#'
+#' @param temp Numeric. Starting temperature for SANN method. Default 10.
+#'   Only used when \code{method = "SANN"}.
+#'
+#' @param tmax Integer. Number of function evaluations at each temperature
+#'   for SANN method. Default 10. Only used when \code{method = "SANN"}.
+#'
+#' @param lmm Integer. Number of BFGS updates retained in L-BFGS-B method.
+#'   Default 5. Only used when \code{method = "L-BFGS-B"}.
+#'
+#' @param factr Numeric. Convergence tolerance factor for L-BFGS-B method.
+#'   Convergence occurs when the reduction in the objective is within this
+#'   factor of the machine tolerance. Default 1e7 (tolerance ~1e-8).
+#'   Only used when \code{method = "L-BFGS-B"}.
+#'
+#' @param pgtol Numeric. Tolerance on the projected gradient for L-BFGS-B method.
+#'   Default 0 (check suppressed). Only used when \code{method = "L-BFGS-B"}.
+#'
+#' @param REPORT Integer. Frequency of progress reports for BFGS, L-BFGS-B and SANN
+#'   methods when \code{trace > 0}. Default 10 for BFGS/L-BFGS-B, 100 for SANN.
+#'
[…]
