+++ R/zzz.R
@@ -457,6 +457,7 @@ NULL
 ## usethis namespace: start
 #' @importFrom Rcpp sourceCpp evalCpp
 #' @import RcppArmadillo
+#' @import RcppEigen
 #' @import graphics
 #' @import gkwdist
 ## usethis namespace: end
