+++ man/gkw_control.Rd
@@ -21,6 +21,20 @@ gkw_control(
   step.max = 1,
   x.tol = 1.5e-08,
   rel.tol = sqrt(.Machine$double.eps),
+  alpha = 1,
+  beta = 0.5,
+  gamma = 2,
+  warn.1d.NelderMead = TRUE,
+  type = 1,
+  temp = 10,
+  tmax = 10,
+  lmm = 5,
+  factr = 1e+07,
+  pgtol = 0,
+  REPORT = NULL,
+  fnscale = 1,
+  parscale = NULL,
+  ndeps = NULL,
   ...
 )
 
@@ -45,11 +59,13 @@ via \code{\link[TMB]{sdreport}} to obtain standard errors and variance-covarianc
 matrix. Set to \code{FALSE} for faster fitting when standard errors are not needed.}
 
 \item{maxit}{Integer. Maximum number of iterations for the optimizer.
-Default 500. Increase for difficult optimization problems.}
+Default 500 for derivative-based methods, 10000 for SANN.
+Increase for difficult optimization problems.}
 
 \item{reltol}{Numeric. Relative convergence tolerance for the optimizer.
 Default \code{sqrt(.Machine$double.eps)} approx. 1.5e-8. Smaller values require
-tighter convergence but may increase computation time.}
+tighter convergence but may increase computation time. Used by Nelder-Mead,
+BFGS, and CG methods.}
 
 \item{abstol}{Numeric. Absolute convergence tolerance. Default 0.
 Used by some optimization methods as an additional stopping criterion.}
@@ -58,7 +74,7 @@ Used by some optimization methods as an additional stopping criterion.}
 \itemize{
 \item 0: Silent (default)
 \item 1: Print iteration progress
-\item 2+: Print detailed diagnostic information
+\item 2+: Print detailed diagnostic information (up to 6 for L-BFGS-B)
 }
 Ignored if \code{silent = TRUE}.}
 
@@ -84,6 +100,56 @@ Default 1.5e-8. Optimizer stops if parameter changes are smaller than this.}
 \item{rel.tol}{Numeric. Relative tolerance for function value (nlminb only).
 Default \code{sqrt(.Machine$double.eps)}. Alternative specification of
 relative tolerance.}
+
+\item{alpha}{Numeric. Reflection factor for Nelder-Mead method. Default 1.0.
+Only used when \code{method = "Nelder-Mead"}.}
+
+\item{beta}{Numeric. Contraction factor for Nelder-Mead method. Default 0.5.
+Only used when \code{method = "Nelder-Mead"}.}
+
+\item{gamma}{Numeric. Expansion factor for Nelder-Mead method. Default 2.0.
+Only used when \code{method = "Nelder-Mead"}.}
+
+\item{warn.1d.NelderMead}{Logical. Whether to warn when Nelder-Mead is used
+for one-dimensional optimization. Default \code{TRUE}.}
+
+\item{type}{Integer. Update formula for CG method. Options:
+\itemize{
+\item 1: Fletcher-Reeves update
+\item 2: Polak-Ribiere update
+\item 3: Beale-Sorenson update
+}
+Default 1. Only used when \code{method = "CG"}.}
+
+\item{temp}{Numeric. Starting temperature for SANN method. Default 10.
+Only used when \code{method = "SANN"}.}
+
+\item{tmax}{Integer. Number of function evaluations at each temperature
+for SANN method. Default 10. Only used when \code{method = "SANN"}.}
+
+\item{lmm}{Integer. Number of BFGS updates retained in L-BFGS-B method.
+Default 5. Only used when \code{method = "L-BFGS-B"}.}
+
[…]
