+++ NAMESPACE
@@ -31,6 +31,7 @@ export(gkwreg)
 export(lrtest)
 export(response)
 import(RcppArmadillo)
+import(RcppEigen)
 import(gkwdist)
 import(graphics)
 importFrom(Formula,as.Formula)

[…]
