### 2025-10-17 6d21aba
Update gkwreg() and plot() documentation.
