+++ R/gkwreg-core.R
@@ -349,6 +349,8 @@
 #'     Four-parameter alternatives when three parameters insufficient
 #'   \item \strong{McDonald (mc)}: Beta generalization via power parameter, useful
 #'     for J-shaped distributions
+#'   \item \strong{Kumaraswamy-Kumaraswamy (kkw)}: Most flexible, use only when
+#'     simpler families inadequate. It extends kw
 #'   \item \strong{Generalized Kumaraswamy (gkw)}: Most flexible, use only when
 #'     simpler families inadequate
 #' }
