+++ man/plot.gkwreg.Rd
@@ -404,114 +404,125 @@ df <- data.frame(y = y, x1 = x1, x2 = x2)
 model <- gkwreg(y ~ x1 | x2, data = df, family = "kw")
 
 # Generate all diagnostic plots with defaults
-plot(model)
+par(mfrow = c(3,2))
+plot(model, ask = FALSE)
 
 # EXAMPLE 2: Selective Plots with Custom Residual Type
 
 # Focus on key diagnostic plots only
+par(mfrow = c(3,1))
 plot(model,
-  which = c(2, 4, 5), # Cook's distance, Resid vs LinPred, Half-normal
-  type = "pearson"
+     which = c(2, 4, 5), # Cook's distance, Resid vs LinPred, Half-normal
+     type = "pearson"
 ) # Use Pearson residuals
 
 # Check for influential points (plot 2) and non-linearity (plot 4)
+par(mfrow = c(2,1))
 plot(model,
-  which = c(2, 4),
-  type = "deviance"
+     which = c(2, 4),
+     type = "deviance"
 )
 
 # EXAMPLE 3: Caption Customization - New Named List Interface
 
 # Customize only specific plot titles (RECOMMENDED NEW WAY)
+par(mfrow = c(3,1))
 plot(model,
-  which = c(1, 4, 6),
-  caption = list(
-    "1" = "Time Pattern Check",
-    "4" = "Linearity Assessment",
-    "6" = "Predictive Accuracy"
-  )
+     which = c(1, 4, 6),
+     caption = list(
+       "1" = "Time Pattern Check",
+       "4" = "Linearity Assessment",
+       "6" = "Predictive Accuracy"
+     )
 )
 
 # Customize subtitle and main title
+par(mfrow = c(2,1))
 plot(model,
-  which = c(1, 5),
-  main = "Model Diagnostics",
-  sub.caption = "Kumaraswamy Regression - Training Data",
-  caption = list("5" = "Normality Check with 95\% Envelope")
+     which = c(1, 5),
+     main = "Model Diagnostics",
+     sub.caption = "Kumaraswamy Regression - Training Data",
+     caption = list("5" = "Normality Check with 95\% Envelope")
 )
 
 # Suppress subtitle entirely
+par(mfrow = c(3,2))
 plot(model, sub.caption = "")
 
 # EXAMPLE 4: Backward Compatible Caption (Vector Interface)
 
 # OLD WAY - still fully supported
+par(mfrow = c(3,2))
 plot(model,
-  which = 1:6,
-  caption = c(
-    "Residual Pattern Analysis",
-    "Influence Diagnostics",
-    "Leverage Assessment",
-    "Linearity Check",
-    "Distributional Fit",
-    "Prediction Quality"
-  )
+     which = 1:6,
+     caption = c(
+       "Residual Pattern Analysis",
+       "Influence Diagnostics",
+       "Leverage Assessment",
+       "Linearity Check",
+       "Distributional Fit",
+       "Prediction Quality"
+     )
 )
 
 # EXAMPLE 5: ggplot2 Graphics with Theming
 
 # Modern publication-quality plots
 plot(model,
-  use_ggplot = TRUE
+     use_ggplot = TRUE,
+     arrange_plots = TRUE
 )
 
 # With custom theme
 plot(model,
-  use_ggplot = TRUE,
-  theme_fn = ggplot2::theme_bw
+     use_ggplot = TRUE,
+     theme_fn = ggplot2::theme_bw,
+     arrange_plots = TRUE
 )
 
 # With classic theme and custom colors (via ...)
 plot(model,
-  use_ggplot = TRUE,
-  theme_fn = ggplot2::theme_classic
+     use_ggplot = TRUE,
+     theme_fn = ggplot2::theme_classic,
+     arrange_plots = TRUE
 )
 
[…]
