+++ NAMESPACE
@@ -31,7 +31,6 @@ export(gkwreg)
 export(lrtest)
 export(response)
 import(RcppArmadillo)
-import(RcppEigen)
 import(gkwdist)
 import(graphics)
 importFrom(Formula,as.Formula)
