### 2025-10-18 0416de7
Add a tests for several methods.
