+++ tests/testthat/test-gkwreg-control-tests.R
@@ -1,0 +1,951 @@
+library(testthat)
+library(gkwreg)
+
+# =============================================================================
+# Basic Object Creation and Structure
+# =============================================================================
+
+test_that("Default control object is created successfully", {
+  # Create default control
+  ctrl <- gkw_control()
+
+  expect_s3_class(ctrl, "gkw_control")
+  expect_type(ctrl, "list")
+  expect_true(all(c("method", "start", "fixed", "hessian", "silent") %in% names(ctrl)))
+})
+
+test_that("Control object has correct default values", {
+  ctrl <- gkw_control()
+
+  expect_equal(ctrl$method, "nlminb")
+  expect_null(ctrl$start)
+  expect_null(ctrl$fixed)
+  expect_true(ctrl$hessian)
+  expect_true(ctrl$silent)
+  expect_equal(ctrl$maxit, 500L)
+})
+
+test_that("Print method works for control objects", {
+  ctrl <- gkw_control()
+
+  # Should not error
+  expect_output(print(ctrl), "Generalized Kumaraswamy Control Parameters")
+  expect_output(print(ctrl), "Method:")
+  expect_output(print(ctrl), "nlminb")
+})
+
+# =============================================================================
+# Method Selection and Validation
+# =============================================================================
+
+test_that("All optimization methods are accepted", {
+  methods <- c("nlminb", "BFGS", "Nelder-Mead", "CG", "SANN", "L-BFGS-B")
+
+  for (m in methods) {
+    ctrl <- gkw_control(method = m)
+    expect_equal(ctrl$method, m)
+    expect_s3_class(ctrl, "gkw_control")
+  }
+})
+
+test_that("Invalid method throws error", {
+  expect_error(
+    gkw_control(method = "invalid_method"),
+    "'arg' should be one of"
+  )
+})
+
+test_that("SANN method gets higher default maxit", {
+  ctrl_sann <- gkw_control(method = "SANN")
+  ctrl_default <- gkw_control()
+
+  expect_equal(ctrl_sann$maxit, 10000L)
+  expect_equal(ctrl_default$maxit, 500L)
+})
+
+# =============================================================================
+# Numeric Parameter Validation
+# =============================================================================
+
+test_that("maxit parameter is validated correctly", {
+  # Valid maxit
+  expect_silent(gkw_control(maxit = 100))
+  expect_silent(gkw_control(maxit = 1000))
+
+  # Invalid maxit
+  expect_error(gkw_control(maxit = -10), "must be a single positive integer")
+  expect_error(gkw_control(maxit = 0), "must be a single positive integer")
+  expect_error(gkw_control(maxit = 10.5), "must be a single positive integer")
+  expect_error(gkw_control(maxit = c(10, 20)), "must be a single positive integer")
+  expect_error(gkw_control(maxit = "100"), "must be a single positive integer")
+})
+
+test_that("Tolerance parameters are validated correctly", {
+  # Valid tolerances
+  expect_silent(gkw_control(reltol = 1e-8))
+  expect_silent(gkw_control(abstol = 1e-10))
+  expect_silent(gkw_control(x.tol = 1e-6))
+
+  # Invalid tolerances (negative)
+  expect_error(gkw_control(reltol = -1e-8), "must be a single non-negative")
+  expect_error(gkw_control(abstol = -0.01), "must be a single non-negative")
+  expect_error(gkw_control(x.tol = -1), "must be a single non-negative")
+
+  # Invalid tolerances (wrong type)
+  expect_error(gkw_control(reltol = "1e-8"), "must be a single non-negative")
+  expect_error(gkw_control(abstol = c(0, 1)), "must be a single non-negative")
+})
+
[…]
