### 2025-10-18 80ff0fc
Update DESCRIPTION and NEWS about tests.
