+++ inst/tmb/kkwreg.cpp
@@ -13,14 +13,12 @@
 //  F(x; α, β, δ, λ) = 1 - {1 - [1 - (1 - x^α)^β]^λ}^(δ+1)
 // ---------------------------------------------------------------------
 
+
 #include <TMB.hpp>
-#include <vector>
 #include <unordered_map>
-#include <cmath>
-#include <cassert>
 
 // ========================
-// Numeric constants
+// Numeric constants - MINIMAL SET
 // ========================
 template <class Type>
 struct Constants {
@@ -43,187 +41,166 @@ template <class Type>
 const Type Constants<Type>::max_exp  = Type(30);
 
 // ========================
-// Safe numerical operations
+// Safe numerical operations - INLINE, NO CondExp
 // ========================
 template <class Type>
-Type safeLog(const Type &x) {
-  // Avoid log(<=0)
-  if (x <= Type(0.0)) return -Constants<Type>::inf_repl;
-  return log(x + Constants<Type>::eps_log);
+inline Type safeLog(const Type &x) {
+  return (x <= Constants<Type>::eps_pos) ? 
+  -Constants<Type>::inf_repl : log(x);
 }
 
 template <class Type>
-Type safeExp(const Type &x) {
-  // Clamp exponent to avoid overflow
-  if (x >  Constants<Type>::max_exp)  return exp(Constants<Type>::max_exp);
-  if (x < -Constants<Type>::max_exp)  return exp(-Constants<Type>::max_exp);
+inline Type safeExp(const Type &x) {
+  double x_d = asDouble(x);
+  if (x_d >  30.0) return exp(Type(30.0));
+  if (x_d < -30.0) return exp(Type(-30.0));
   return exp(x);
 }
 
-template <class Type>
-Type safePow(const Type &base, const Type &exponent) {
-  // Protect base^exponent in extreme cases
-  if (base <= Constants<Type>::eps_pos) {
-    return (exponent > Type(0.0)) ? Type(0.0) : Constants<Type>::inf_repl;
-  }
-  if (fabs(exponent) > Type(1.0)) {
-    return safeExp(exponent * safeLog(base));
-  }
-  return pow(base, exponent);
-}
-
 // ========================
-// Enforce domain
+// Link functions - SIMPLIFIED, INLINE, NO CondExp
 // ========================
 template <class Type>
-Type enforceMin(const Type &x, const Type &min_val) {
-  // max(x, min_val)
-  return CppAD::CondExpLt(x, min_val, min_val, x);
-}
-
-template <class Type>
-Type enforceProbability(const Type &x) {
-  // Clamps x into (eps_prob, 1 - eps_prob)
-  Type eps   = Constants<Type>::eps_prob;
-  Type upper = Type(1.0) - eps;
-  Type clippedUp = CppAD::CondExpGt(x, upper, upper, x);
-  return CppAD::CondExpLt(clippedUp, eps, eps, clippedUp);
-}
-
-// ========================
-// Link functions
-// (1=log, 2=logit, 3=probit, 4=cauchy, 5=cloglog,
-//  6=identity, 7=sqrt, 8=inverse, 9=inverse-square)
-// ========================
-template <class Type>
-Type fast_log_link(const Type &eta) {
-  return safeExp(eta);
-}
-
-template <class Type>
-Type fast_logit_link(const Type &eta) {
-  if (eta >  Constants<Type>::max_exp) return Type(1.0) - Constants<Type>::eps_prob;
-  if (eta < -Constants<Type>::max_exp) return Constants<Type>::eps_prob;
[…]
