+++ inst/tmb/gkwreg.cpp
@@ -1,26 +1,36 @@
-// File: gkwreg_v6.cpp
-// ---------------------------------------------------------------
-//  Regression Model for the Generalized Kumaraswamy (GKw) Distribution
-//  using TMB, focusing on scalability with big data and many covariates.
+// File: gkwreg_optimized.cpp
+// ---------------------------------------------------------------------
+//  OPTIMIZED Regression Model for the Generalized Kumaraswamy (GKw) Distribution
 //
-//  Changes from v5:
-//   - Removed all OpenMP dependencies for better portability
-//   - Maintains all optimizations including caching with unordered_map
-//   - Preserves computation scheduling flexibility via user chunk size
-// ---------------------------------------------------------------
+//  Mathematical model:
+//  f(x; α, β, γ, δ, λ) = 
+//    (λαβ x^(α-1) (1-x^α)^(β-1)) / B(γ, δ+1) *
+//    [1-(1-x^α)^β]^(γλ-1) *
+//    [1-[1-(1-x^α)^β]^λ]^δ
+//
+//  where 0 < x < 1, α, β, γ, δ, λ > 0
+//
+//  Optimization level: Elite (targeting 25-35% speedup)
+//  Based on proven techniques from kwreg_optimized.cpp
+//  
+//  Key optimizations:
+//  - Removed ALL CppAD::CondExp overhead (major performance killer)
+//  - Using native pow() instead of exp(log()) wherever possible
+//  - Maintained if(isDouble<Type>::value) optimization (compile-time branch)
+//  - Simplified validation using asDouble() checks (zero-cost in double mode)
+//  - All functions marked inline
+//  - Minimal overhead in hot paths
+//  - 100% compatible with original interface
+//  - Exact analytical mean if available (none known for GKw)
+//  - Precomputed link functions
+//  - Single unified loop structure
+// ---------------------------------------------------------------------
 
 #include <TMB.hpp>
-#include <vector>
 #include <unordered_map>
-#include <cmath>
-#include <cassert>
-
-// Provide stub functions to replace OpenMP functions
-// inline int omp_get_max_threads() { return 1; }
-// inline int omp_get_thread_num() { return 0; }
 
 // ========================
-// Numeric constants
+// Numeric constants - MINIMAL SET
 // ========================
 template <class Type>
 struct Constants {
@@ -43,192 +53,174 @@ template <class Type>
 const Type Constants<Type>::max_exp  = Type(30);
 
 // ========================
-// Safe numerical operations
+// Safe numerical operations - INLINE, NO CondExp
 // ========================
 template <class Type>
-Type safeLog(const Type &x) {
-  // Avoid log(<=0)
-  if (x <= Type(0.0)) return -Constants<Type>::inf_repl;
-  return log(x + Constants<Type>::eps_log);
+inline Type safeLog(const Type &x) {
+  return (x <= Constants<Type>::eps_pos) ? 
+  -Constants<Type>::inf_repl : log(x);
 }
 
 template <class Type>
-Type safeExp(const Type &x) {
-  // Clamp exponent to avoid overflow
-  if (x >  Constants<Type>::max_exp)  return exp(Constants<Type>::max_exp);
-  if (x < -Constants<Type>::max_exp)  return exp(-Constants<Type>::max_exp);
+inline Type safeExp(const Type &x) {
+  double x_d = asDouble(x);
+  if (x_d >  30.0) return exp(Type(30.0));
[…]

[…]
