+++ pkgdown.yml
@@ -1,8 +1,9 @@
 pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
-articles: {}
-last_built: 2025-10-18T03:05Z
+articles:
+  gkwreg-vs-betareg: gkwreg-vs-betareg.html
+last_built: 2025-10-20T21:08Z
 urls:
   reference: https://evandeilton.github.io/gkwreg/reference
   article: https://evandeilton.github.io/gkwreg/articles
