### 2025-10-21 e583375
IMPROVE NUMERIC STABILTY TO gkw FAMILY. FIX DOCUMENTATION.
