+++ R/zzz.R
@@ -288,8 +288,8 @@ utils::globalVariables(c(
           TMB::compile(cpp_basename, # Use only the filename
             safebounds = FALSE,
             safeunload = FALSE,
-            verbose = verbose,
-            flags = "-O3 -march=native -ffast-math -funroll-loops"
+            verbose = verbose
+            # flags = "-O3 -march=native -ffast-math -funroll-loops"
           )
         },
         silent = !verbose
@@ -301,8 +301,8 @@ utils::globalVariables(c(
           TMB::compile(temp_cpp_file,
             safebounds = FALSE,
             safeunload = FALSE,
-            verbose = verbose,
-            flags = "-O3 -march=native -ffast-math -funroll-loops"
+            verbose = verbose
+            # flags = "-O3 -march=native -ffast-math -funroll-loops"
           )
         },
         silent = !verbose
@@ -321,8 +321,8 @@ utils::globalVariables(c(
             TMB::compile("tmb_file.cpp",
               safebounds = FALSE,
               safeunload = FALSE,
-              verbose = verbose,
-              flags = "-O3 -march=native -ffast-math -funroll-loops"
+              verbose = verbose
+              # flags = "-O3 -march=native -ffast-math -funroll-loops"
             )
           },
           silent = !verbose
