+++ vignettes/gkwreg-vs-betareg.Rmd
@@ -85,10 +85,10 @@ library(ggplot2)
 
 # Standardized color palette for all comparisons
 MODEL_COLORS <- c(
-  "Beta (betareg)" = "#D32F2F",      # Red
-  "Beta (gkwreg)" = "#1976D2",       # Blue  
-  "Kumaraswamy" = "#388E3C",         # Green
-  "Exp. Kumaraswamy" = "#7B1FA2"     # Purple
+  "Beta (betareg)" = "#D32F2F", # Red
+  "Beta (gkwreg)" = "#1976D2", # Blue
+  "Kumaraswamy" = "#388E3C", # Green
+  "Exp. Kumaraswamy" = "#7B1FA2" # Purple
 )
 
 MODEL_NAMES <- c(
@@ -161,92 +161,104 @@ Maximum likelihood estimation proceeds via numerical optimization of the log-lik
 
 # Store fitted model objects
 run_single_sim_with_fits <- function(n, dgp_fun, params, formula, test_prop = 0.2) {
-  data_full <- tryCatch({
-    dgp_fun(n, params)
-  }, error = function(e) NULL)
-  
+  data_full <- tryCatch(
+    {
+      dgp_fun(n, params)
+    },
+    error = function(e) NULL
+  )
+
   if (is.null(data_full)) {
     return(list(
       betareg = list(success = FALSE),
       gkw_kw = list(success = FALSE)
     ))
   }
-  
+
   n_test <- floor(n * test_prop)
   test_idx <- sample(seq_len(n), n_test)
   train_data <- data_full[-test_idx, ]
   test_data <- data_full[test_idx, ]
-  
+
   results <- list()
-  
+
   # Beta regression
-  results$betareg <- tryCatch({
-    t0 <- proc.time()[3]
-    fit <- betareg(formula, data = train_data)
-    time_elapsed <- proc.time()[3] - t0
-    
-    pred <- predict(fit, newdata = test_data, type = "response")
-    
-    if (any(is.na(pred)) || any(is.infinite(pred))) {
-      return(list(success = FALSE))
+  results$betareg <- tryCatch(
+    {
+      t0 <- proc.time()[3]
+      fit <- betareg(formula, data = train_data)
+      time_elapsed <- proc.time()[3] - t0
+
+      pred <- predict(fit, newdata = test_data, type = "response")
+
+      if (any(is.na(pred)) || any(is.infinite(pred))) {
+        return(list(success = FALSE))
+      }
+
+      rmse <- sqrt(mean((test_data$y - pred)^2))
+
+      list(
+        fit = fit,
+        coef = coef(fit),
+        se = sqrt(diag(vcov(fit))),
+        loglik = as.numeric(logLik(fit)),
+        aic = AIC(fit),
+        bic = BIC(fit),
+        rmse = rmse,
+        time = time_elapsed,
+        converged = isTRUE(fit$converged),
+        success = TRUE
+      )
+    },
+    error = function(e) {
+      list(success = FALSE)
     }
-    
-    rmse <- sqrt(mean((test_data$y - pred)^2))
-    
-    list(
-      fit = fit,
-      coef = coef(fit),
-      se = sqrt(diag(vcov(fit))),
-      loglik = as.numeric(logLik(fit)),
-      aic = AIC(fit),
-      bic = BIC(fit),
-      rmse = rmse,
-      time = time_elapsed,
-      converged = isTRUE(fit$converged),
-      success = TRUE
-    )
-  }, error = function(e) {
-    list(success = FALSE)
-  })
-  
+  )
+
[…]

[…]
