+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   gkwreg-vs-betareg: gkwreg-vs-betareg.html
-last_built: 2025-10-20T21:08Z
+last_built: 2025-10-21T02:57Z
 urls:
   reference: https://evandeilton.github.io/gkwreg/reference
   article: https://evandeilton.github.io/gkwreg/articles
