### 2025-10-31 #8991 @evandeilton
Hi @jbytecode,

Due to the extent of the changes requested by @JonasMoss, I'm finding it challenging to complete all modifications within the initially expected timeframe. The refactoring process is taking considerably longer than I previously anticipated.

I've developed a new package, [gkwdist](https://github.com/evandeilton/gkwdist), to publish the GKW distributions separately. This package has been submitted to CRAN and is currently awaiting approval. Once approved, I will add it as a dependency to [gkwreg](https://github.com/evandeilton/gkwreg).

I'm currently addressing several issues and am committed to accommodating all requested tasks. Progress is being tracked [here](https://github.com/evandeilton/gkwreg/issues) and [here](https://github.com/evandeilton/gkwdist/issues).

A comprehensive revision is in progress, and I appreciate your patience during this refactoring phase.
