### 2025-11-02 d83216d
Fix README and paper PDF
