+++ README.md
@@ -33,7 +33,7 @@ speed, numerical accuracy, and optimization stability.
 
 ## Key Features
 
-### 📊 Flexible Distribution Hierarchy
+### Flexible Distribution Hierarchy
 
 Model bounded data using the **5-parameter Generalized Kumaraswamy
 (GKw)** distribution and its **seven nested subfamilies**:
@@ -52,7 +52,7 @@ Each family offers distinct flexibility-parsimony tradeoffs. Start
 simple (`kw` or `beta`) and compare nested models using likelihood ratio
 tests or information criteria.
 
-### 🎯 Advanced Regression Modeling
+### Advanced Regression Modeling
 
 - **Extended formula syntax** for parameter-specific linear predictors:
 
@@ -80,7 +80,7 @@ tests or information criteria.
     only)
   - Debugging tools (verbose output, trace levels)
 
-### ⚡ Computational Efficiency
+### Computational Efficiency
 
 - **TMB-powered estimation**: Compiled C++ templates with automatic
   differentiation
@@ -92,7 +92,7 @@ tests or information criteria.
   - Vectorized operations via Eigen/Armadillo
   - Memory-efficient for large datasets (n \> 100,000)
 
-### 🔬 Comprehensive Inference Tools
+### Comprehensive Inference Tools
 
 **Standard R Methods** (familiar workflow): - `summary()`, `print()`,
 `coef()`, `vcov()`, `confint()` - `logLik()`, `AIC()`, `BIC()`,
@@ -111,7 +111,7 @@ under alternative distributional assumptions
 Information criteria: `AIC()`, `BIC()` with multi-model comparison -
 Automated nesting detection and proper test statistics
 
-### 📈 Sophisticated Diagnostics
+### Sophisticated Diagnostics
 
 **6 Diagnostic Plot Types** (`plot.gkwreg`): 1. **Residuals vs
 Observation Indices**: Detect autocorrelation, temporal patterns 2.
@@ -129,7 +129,7 @@ for titles, themes, arrangement - **Performance**: Automatic sampling
 for large datasets, adjustable envelope simulations - **Programmatic
 access**: `save_diagnostics = TRUE` returns computed measures
 
-### 🔗 Ecosystem Integration
+### Ecosystem Integration
 
 - **Distribution functions** via companion package
   [`gkwdist`](https://github.com/evandeilton/gkwdist):
@@ -416,7 +416,7 @@ $$F(x; \alpha, \beta, \gamma, \delta, \lambda) = I_{[1-(1-x^{\alpha})^{\beta}]^{
 where $I_z(a,b)$ is the regularized incomplete beta function. The
 probability density function is:
 
-$$f(x; \alpha, \beta, \gamma, \delta, \lambda) = \frac{\lambda \alpha \beta x^{\alpha-1}}{B(\gamma, \delta)} (1-x^{\alpha})^{\beta-1} [1-(1-x^{\alpha})^{\beta}]^{\gamma\lambda-1} \{1-[1-(1-x^{\alpha})^{\beta}]^{\lambda}\}^{\delta-1}$$
+$$f(x; \alpha, \beta, \gamma, \delta, \lambda) = \frac{\lambda \alpha \beta x^{\alpha-1}}{B(\gamma, \delta)} (1-x^{\alpha})^{\beta-1} \left[1-(1-x^{\alpha})^{\beta}\right]^{\gamma\lambda-1} \left\{1-\left[1-(1-x^{\alpha})^{\beta}\right]^{\lambda} \right \}^{\delta-1}$$
 
 **Parameter Roles**: - **α, β**: Control basic shape (inherited from
 Kumaraswamy) - **γ, δ**: Govern tail behavior and concentration - **λ**:
[…]
