+++ vignettes/gkwreg-vs-betareg.Rmd
@@ -1285,76 +1285,6 @@ This study contributes three methodological insights:
 
 These quantitative benchmarks provide practitioners with concrete decision rules rather than vague guidance.
 
-## Limitations and Future Directions
-
-Several limitations warrant acknowledgment:
-
-1. **Sample sizes**: Simulations employed moderate sample sizes (n = 300-400); asymptotics for very large (n > 10,000) or small (n < 50) samples require investigation
-
-2. **Covariate structures**: Simple linear predictors examined; nonlinear effects, smoothing splines, and high-dimensional settings deserve attention
-
-3. **Missing data**: Impact of missingness mechanisms (MCAR, MAR, MNAR) on relative performance unexplored
-
-4. **Bayesian inference**: Comparison focused on maximum likelihood; MCMC implementations of Kumaraswamy models and prior sensitivity analysis needed
-
-5. **Discrete-continuous mixtures**: Zero-inflated or one-inflated extensions not addressed
-
-Future research should examine:
-
-- **Asymptotic relative efficiency**: Formal Cramér-Rao lower bound comparisons under correct and misspecified models
-- **Diagnostic tools**: Development of residual plots, influence measures, and goodness-of-fit tests specific to Kumaraswamy regression  
-- **Time series extensions**: Adaptation to longitudinal, panel, and time series structures
-- **Multivariate generalizations**: Joint modeling of multiple bounded responses
-- **Machine learning integration**: Incorporation into ensemble methods, boosting, and neural networks
-
-## Recommendations for Practice
-
-For applied researchers analyzing bounded continuous data:
-
-1. **Default strategy**: Consider Kumaraswamy as the primary choice for exploratory analysis given computational advantages and robustness. Only switch to Beta if specific theoretical reasons warrant (e.g., maintaining consistency with established literature in your field)
-
-2. **Diagnostic protocol**: 
-   - Generate histogram with density overlay to assess tail behavior
-   - Create QQ plot against Beta distribution to identify departures
-   - Calculate proportion of observations within 0.1 of boundaries
-   - If any diagnostic suggests misspecification, employ Exponentiated Kumaraswamy
-
-3. **Inference guidelines**:
-   - For correctly specified models: Beta and Kumaraswamy provide equivalent inference
-   - For misspecified models: Kumaraswamy standard errors more accurately reflect uncertainty
-   - Always report 95% CIs, not just p-values, to make precision explicit
-
-4. **Computational planning**: For intensive procedures (cross-validation, bootstrap, permutation tests), the 2-9× Kumaraswamy speedup transforms feasibility. Budget computational resources accordingly.
-
-5. **Reporting standards**: Document:
-   - Which distributional family was used and why
-   - Convergence diagnostics (all models converged? optimization details?)
-   - Computational time (for reproducibility and methods comparison)
-   - Model selection criteria (AIC, BIC) across candidate families
-
-6. **Software implementation**: The `gkwreg` package provides accessible implementation with syntax closely paralleling `betareg`:
-
-```r
-# Beta regression
-fit_beta <- betareg(y ~ x1 + x2 | z, data = mydata)
-
-# Kumaraswamy regression (drop-in replacement)
-fit_kw <- gkwreg(y ~ x1 + x2 | z, data = mydata, family = "kw")
-
-# Compare
-AIC(fit_beta, fit_kw)
-```
-
-The syntactic similarity facilitates adoption and experimentation.
-
-## Final Perspective
-
-The dominance of Beta regression in bounded-data analysis reflects historical precedent and software availability more than inherent statistical superiority. This study demonstrates that **Kumaraswamy-based models are not merely "alternatives" but often superior choices** across the performance dimensions that matter in practice: reliability, speed, and inferential validity.
-
[…]

[…]
