### 2025-11-14 c6671cd
ADD gkwgist from R-CRAN and fix documentation.
