+++ inst/tmb/kwreg.cpp
@@ -414,561 +414,3 @@ Type objective_function<Type>::operator() () {
   
   return nll;
 }
-
-
-
-
-
-
-
-// #include <TMB.hpp>
-// #include <vector>
-// #include <unordered_map>
-// #include <cmath>
-// #include <cassert>
-// 
-// // ========================
-// // Numeric constants
-// // ========================
-// template <class Type>
-// struct Constants {
-//   static const Type eps_log;
-//   static const Type eps_pos;
-//   static const Type eps_prob;
-//   static const Type inf_repl;
-//   static const Type max_exp;
-// };
-// 
-// template <class Type>
-// const Type Constants<Type>::eps_log  = Type(1e-15);
-// template <class Type>
-// const Type Constants<Type>::eps_pos  = Type(1e-10);
-// template <class Type>
-// const Type Constants<Type>::eps_prob = Type(1e-12);
-// template <class Type>
-// const Type Constants<Type>::inf_repl = Type(1e10);
-// template <class Type>
-// const Type Constants<Type>::max_exp  = Type(30);
-// 
-// // ========================
-// // Safe numerical operations
-// // ========================
-// template <class Type>
-// Type safeLog(const Type &x) {
-//   // Avoid log(<=0)
-//   if (x <= Type(0.0)) return -Constants<Type>::inf_repl;
-//   return log(x + Constants<Type>::eps_log);
-// }
-// 
-// template <class Type>
-// Type safeExp(const Type &x) {
-//   // Clamp exponent to avoid overflow
-//   if (x >  Constants<Type>::max_exp)  return exp(Constants<Type>::max_exp);
-//   if (x < -Constants<Type>::max_exp)  return exp(-Constants<Type>::max_exp);
-//   return exp(x);
-// }
-// 
-// template <class Type>
-// Type safePow(const Type &base, const Type &exponent) {
-//   // Protect base^exponent in extreme cases
-//   if (base <= Constants<Type>::eps_pos) {
-//     return (exponent > Type(0.0)) ? Type(0.0) : Constants<Type>::inf_repl;
-//   }
-//   if (fabs(exponent) > Type(1.0)) {
-//     return safeExp(exponent * safeLog(base));
-//   }
-//   return pow(base, exponent);
-// }
-// 
-// // ========================
-// // Enforce domain
-// // ========================
-// template <class Type>
-// Type enforceMin(const Type &x, const Type &min_val) {
-//   // max(x, min_val)
-//   return CppAD::CondExpLt(x, min_val, min_val, x);
-// }
-// 
-// template <class Type>
-// Type enforceProbability(const Type &x) {
-//   // Clamps x into (eps_prob, 1 - eps_prob)
-//   Type eps   = Constants<Type>::eps_prob;
-//   Type upper = Type(1.0) - eps;
-//   Type clippedUp = CppAD::CondExpGt(x, upper, upper, x);
-//   return CppAD::CondExpLt(clippedUp, eps, eps, clippedUp);
-// }
-// 
-// // ========================
-// // Link functions
-// // (1=log, 2=logit, 3=probit, 4=cauchy, 5=cloglog,
-// //  6=identity, 7=sqrt, 8=inverse, 9=inverse-square)
-// // ========================
-// template <class Type>
-// Type fast_log_link(const Type &eta) {
-//   return safeExp(eta);
-// }
-// 
-// template <class Type>
-// Type fast_logit_link(const Type &eta) {
-//   if (eta >  Constants<Type>::max_exp) return Type(1.0) - Constants<Type>::eps_prob;
[…]
