+++ reference/LossAversion.md
@@ -1,0 +1,358 @@
+# (No) Myopic Loss Aversion in Adolescents
+
+Data from a behavioral economics experiment assessing the extent of
+myopic loss aversion among adolescents aged 11 to 19 years. The
+experiment tests whether short-term investment horizons lead to more
+conservative investment behavior.
+
+## Usage
+
+``` r
+LossAversion
+```
+
+## Format
+
+A data frame with 570 observations on 7 variables:
+
+- invest:
+
+  numeric. Average proportion of tokens invested across all 9 rounds of
+  the experiment (response variable).
+
+- gender:
+
+  factor. Gender of the player (or team of players).
+
+- male:
+
+  factor. Was (at least one of) the player(s) male (in the team)?
+
+- age:
+
+  numeric. Age of the player (or average age in case of team).
+
+- grade:
+
+  factor. School grade of the player(s).
+
+- arrangement:
+
+  factor. Investment horizon treatment with levels `short` (1 round),
+  `medium` (3 rounds), and `long` (9 rounds).
+
+- treatment:
+
+  factor. Type of treatment: long vs. short.
+
+## Source
+
+Data collected by Matthias Sutter and Daniela Glätzle-Rützler,
+Universität Innsbruck.
+
+## Details
+
+The data were collected by Matthias Sutter and Daniela Glätzle-Rützler
+(Universität Innsbruck) in an experiment with high-school students in
+Tyrol, Austria (Schwaz and Innsbruck). The experiment tests the theory
+of myopic loss aversion, which proposes that investors with shorter
+evaluation periods are more loss-averse and thus invest less in risky
+assets.
+
+Classical theory predicts that players with short investment horizons
+(myopic view) should invest less due to loss aversion. However, Sutter
+et al. (2015) found no evidence of myopic loss aversion in adolescents,
+contrary to findings in adult populations.
+
+The investment game structure: In each round, players could invest
+tokens in a risky asset with 50% chance of doubling or losing the
+investment. The treatment varied the feedback frequency (short = every
+round, medium = every 3 rounds, long = only at the end).
+
+## References
+
+Sutter, M., Kocher, M.G., Glätzle-Rützler, D., and Trautmann, S.T.
+(2015). No Myopic Loss Aversion in Adolescents? – An Experimental Note.
+*Journal of Economic Behavior & Organization*, **111**, 169–176.
+[doi:10.1016/j.jebo.2014.12.021](https://doi.org/10.1016/j.jebo.2014.12.021)
+
+Kosmidis, I., and Zeileis, A. (2024). Extended-Support Beta Regression
+for (0, 1) Responses. *arXiv:2409.07233*.
+[doi:10.48550/arXiv.2409.07233](https://doi.org/10.48550/arXiv.2409.07233)
+
+## Examples
+
+``` r
+# \donttest{
+require(gkwreg)
+require(gkwdist)
+
+data(LossAversion)
+# Control bounds
+
+LossAversion$invest <- with(
+  LossAversion,
+  ifelse(invest <= 0, 0.000001,
+    ifelse(invest >= 1, 0.999999, invest)
+  )
+)
+# Example 1: Test for myopic loss aversion
+# Do short-term players invest less? (They shouldn't, per Sutter et al.)
+fit_kw <- gkwreg(
+  invest ~ arrangement + age + male + grade |
+    arrangement + male,
+  data = LossAversion,
+  family = "kw"
+)
+summary(fit_kw)
+#> 
[…]
