+++ reference/retinal.md
@@ -1,0 +1,456 @@
+# Intraocular Gas Decay in Retinal Surgery
+
+Longitudinal data on the recorded decay of intraocular gas
+(perfluoropropane) in complex retinal surgeries. The dataset tracks the
+proportion of gas remaining over time following vitrectomy procedures.
+
+## Usage
+
+``` r
+retinal
+```
+
+## Format
+
+A data frame with 40 observations on 7 variables:
+
+- ID:
+
+  integer. Patient identification number for longitudinal tracking.
+
+- Gas:
+
+  numeric. Proportion of intraocular gas remaining (0-1 scale). Response
+  variable measuring the fraction of perfluoropropane gas still present
+  in the vitreous cavity.
+
+- Time:
+
+  numeric. Time point of measurement (days or weeks post-surgery).
+
+- LogT:
+
+  numeric. Logarithm of time, log(Time). Used to linearize the
+  exponential decay pattern.
+
+- LogT2:
+
+  numeric. Squared logarithm of time, (log(Time))^2. Captures nonlinear
+  decay patterns.
+
+- Level:
+
+  factor. Initial gas concentration level at the time of injection.
+  Different starting concentrations affect decay kinetics.
+
+## Source
+
+Based on clinical data from vitreoretinal surgery patients. Originally
+analyzed in Song and Tan (2000).
+
+## Details
+
+This longitudinal dataset comes from a study of gas decay following
+vitreoretinal surgery. Perfluoropropane (C3F8) is commonly used as a
+temporary tamponade agent in retinal detachment repair and other complex
+vitreoretinal procedures.
+
+**Clinical background:** During vitrectomy for retinal detachment, gas
+bubbles are injected into the vitreous cavity to help reattach the
+retina by providing internal tamponade. The gas gradually absorbs and
+dissipates over time. Understanding the decay rate is important for:
+
+- Predicting when patients can resume normal activities (esp. air
+  travel)
+
+- Assessing treatment efficacy
+
+- Planning follow-up examinations
+
+**Decay kinetics:** Gas decay typically follows a nonlinear pattern that
+can be approximated by exponential or power-law functions. The log
+transformation (LogT, LogT2) helps linearize these relationships for
+regression modeling.
+
+**Data structure:** This is a longitudinal/panel dataset with repeated
+measurements on the same patients over time. Correlation structures
+(exchangeable, AR(1), etc.) should be considered when modeling.
+
+The proportional nature of the gas variable (bounded between 0 and 1)
+makes this dataset ideal for:
+
+- Simplex marginal models (original application by Song & Tan 2000)
+
+- Beta regression with longitudinal correlation structures
+
+- Kumaraswamy regression with heteroscedastic errors
+
+## References
+
+Meyers, S.M., Ambler, J.S., Tan, M., Werner, J.C., and Huang, S.S.
[…]
