+++ reference/dot-process_link_scale.md
@@ -1,0 +1,31 @@
+# Process Link Scales for GKw Regression
+
+Process Link Scales for GKw Regression
+
+## Usage
+
+``` r
+.process_link_scale(link_scale, link_list, param_names, fixed_params)
+```
+
+## Arguments
+
+- link_scale:
+
+  A numeric value or list specifying scales for link functions.
+
+- link_list:
+
+  List of link functions for each parameter.
+
+- param_names:
+
+  Names of the parameters for the specified family.
+
+- fixed_params:
+
+  List of fixed parameters.
+
+## Value
+
+A list of link scales.

[…]
