+++ man/dot-get_cache_dir.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/gkwreg-compile-tmb.R
+\name{.get_cache_dir}
+\alias{.get_cache_dir}
+\title{Get Cache Directory for Compiled Models}
+\usage{
+.get_cache_dir(pkg_name)
+}
+\arguments{
+\item{pkg_name}{Package name}
+}
+\value{
+Path to cache directory
+}
+\description{
+Creates version-specific cache directory for TMB DLLs.
+Structure: <cache_root>/<pkg_name>/tmb_cache/R\if{html}{\out{<version>}}/<pkg_version>/
+}
+\keyword{internal}

[…]
