### 2025-11-16 c5ed731
Fix RcppEigen dep
