+++ reference/LossAversion.md
@@ -119,15 +119,15 @@ summary(fit_kw)
 #> -0.5002 -0.2002  0.0000  0.0037  0.1999  0.5000 
 #> 
 #> Coefficients:
-#>                       Estimate Std. Error z value Pr(>|z|)
-#> alpha:(Intercept)      0.00000    0.47351       0        1
-#> alpha:arrangementteam  0.00000    0.18358       0        1
-#> alpha:age              0.00000    0.03771       0        1
-#> alpha:maleyes          0.00000    0.12103       0        1
-#> alpha:grade10-12       0.00000    0.17325       0        1
-#> beta:(Intercept)       0.00000    0.09653       0        1
-#> beta:arrangementteam   0.00000    0.16937       0        1
-#> beta:maleyes           0.00000    0.13944       0        1
+#>                         Estimate Std. Error z value Pr(>|z|)
+#> alpha:(Intercept)     -2.841e-10  4.735e-01       0        1
+#> alpha:arrangementteam  2.126e-10  1.836e-01       0        1
+#> alpha:age             -4.949e-09  3.771e-02       0        1
+#> alpha:maleyes          1.889e-11  1.210e-01       0        1
+#> alpha:grade10-12      -4.135e-10  1.732e-01       0        1
+#> beta:(Intercept)       4.573e-10  9.653e-02       0        1
+#> beta:arrangementteam  -3.177e-11  1.694e-01       0        1
+#> beta:maleyes          -1.490e-11  1.394e-01       0        1
 #> 
 #> Confidence intervals (95%):
 #>                            3%    98%
@@ -155,15 +155,15 @@ summary(fit_kw)
 #> Number of observations: 570 
 #> Number of parameters: 8 
 #> Residual degrees of freedom: 562 
-#> Log-likelihood: -3e+21 
-#> AIC: 6e+21 
-#> BIC: 6e+21 
+#> Log-likelihood: -3e+11 
+#> AIC: 6e+11 
+#> BIC: 6e+11 
 #> RMSE: 0.2678 
 #> Efron's R2: 7.819e-05 
 #> Mean Absolute Error: 0.2217 
 #> 
-#> Convergence status: Successful 
-#> Iterations: 1 
+#> Convergence status: Failed 
+#> Iterations: 2 
 #> 
 
 # Interpretation:
@@ -195,18 +195,18 @@ summary(fit_kw_interact)
 #> -0.5002 -0.2002  0.0000  0.0037  0.2000  0.5000 
 #> 
 #> Coefficients:
-#>                                  Estimate Std. Error z value Pr(>|z|)
-#> alpha:(Intercept)                 0.00000    0.67535       0        1
-#> alpha:grade10-12                  0.00000    1.12039       0        1
-#> alpha:arrangementteam             0.00000    0.21132       0        1
-#> alpha:age                         0.00000    0.05342       0        1
-#> alpha:maleyes                     0.00000    0.12237       0        1
-#> alpha:grade10-12:arrangementteam  0.00000    0.22609       0        1
-#> alpha:grade10-12:age              0.00000    0.07630       0        1
-#> beta:(Intercept)                  0.00000    0.11264       0        1
-#> beta:arrangementteam              0.00000    0.17198       0        1
-#> beta:maleyes                      0.00000    0.13977       0        1
-#> beta:grade10-12                   0.00000    0.14404       0        1
+#>                                    Estimate Std. Error z value Pr(>|z|)
+#> alpha:(Intercept)                -1.698e-10  6.753e-01       0        1
+#> alpha:grade10-12                 -2.471e-10  1.120e+00       0        1
[…]
