+++ reference/dot-is_dll_functional.md
@@ -1,0 +1,23 @@
+# Check if DLL is Loaded and Functional
+
+Check if DLL is Loaded and Functional
+
+## Usage
+
+``` r
+.is_dll_functional(dll_name, verbose = FALSE)
+```
+
+## Arguments
+
+- dll_name:
+
+  Base name of DLL
+
+- verbose:
+
+  Logical
+
+## Value
+
+Logical
