+++ reference/list_tmb_models.md
@@ -1,0 +1,19 @@
+# List Available TMB Models
+
+Returns vector of TMB model names available in the package.
+
+## Usage
+
+``` r
+list_tmb_models(pkg_name = "gkwreg")
+```
+
+## Arguments
+
+- pkg_name:
+
+  Package name (default: "gkwreg")
+
+## Value
+
+Character vector of model names (without .cpp extension)

[…]
