### 2025-11-16 0023a87
Fix documentation and examples in gkwreg
