+++ R/gkwreg-core.R
@@ -1,547 +1,231 @@
 #' Fit Generalized Kumaraswamy Regression Models
 #'
 #' @description
-#' Fits regression models using the Generalized Kumaraswamy (GKw) family of
-#' distributions for modeling response variables strictly bounded in the interval
-#' (0, 1). The function provides a unified interface for fitting seven nested
-#' submodels of the GKw family, allowing flexible modeling of proportions, rates,
-#' and other bounded continuous outcomes through regression on distributional
-#' parameters.
-#'
-#' Maximum Likelihood Estimation is performed via automatic differentiation using
-#' the TMB (Template Model Builder) package, ensuring computational efficiency and
-#' numerical accuracy. The interface follows standard R regression modeling
-#' conventions similar to \code{\link[stats]{lm}}, \code{\link[stats]{glm}}, and
-#' \code{\link[betareg]{betareg}}, making it immediately familiar to R users.
-#'
-#' @param formula An object of class \code{\link[Formula]{Formula}} (or one that
-#'   can be coerced to that class). The formula uses extended syntax to specify
-#'   potentially different linear predictors for each distribution parameter:
-#'
-#'   \code{y ~ model_alpha | model_beta | model_gamma | model_delta | model_lambda}
-#'
-#'   where:
-#'   \itemize{
-#'     \item \code{y} is the response variable (must be in the open interval (0, 1))
-#'     \item \code{model_alpha} specifies predictors for the \eqn{\alpha} parameter
-#'     \item \code{model_beta} specifies predictors for the \eqn{\beta} parameter
-#'     \item \code{model_gamma} specifies predictors for the \eqn{\gamma} parameter
-#'     \item \code{model_delta} specifies predictors for the \eqn{\delta} parameter
-#'     \item \code{model_lambda} specifies predictors for the \eqn{\lambda} parameter
-#'   }
-#'
-#'   If a part is omitted or specified as \code{~ 1}, an intercept-only model is
-#'   used for that parameter. Parts corresponding to fixed parameters (determined
-#'   by \code{family}) are automatically ignored. See Details and Examples for
-#'   proper usage.
-#'
-#' @param data A data frame containing the variables specified in \code{formula}.
-#'   Standard R subsetting and missing value handling apply.
-#'
-#' @param family A character string specifying the distribution family from the
-#'   Generalized Kumaraswamy hierarchy. Must be one of:
-#'   \describe{
-#'     \item{\code{"gkw"}}{Generalized Kumaraswamy (default). Five parameters:
-#'       \eqn{\alpha, \beta, \gamma, \delta, \lambda}. Most flexible, suitable
-#'       when data show complex behavior not captured by simpler families.}
-#'     \item{\code{"bkw"}}{Beta-Kumaraswamy. Four parameters:
-#'       \eqn{\alpha, \beta, \gamma, \delta} (fixes \eqn{\lambda = 1}).
-#'       Combines Beta and Kumaraswamy flexibility.}
-#'     \item{\code{"kkw"}}{Kumaraswamy-Kumaraswamy. Four parameters:
-#'       \eqn{\alpha, \beta, \delta, \lambda} (fixes \eqn{\gamma = 1}).
-#'       Alternative four-parameter generalization.}
-#'     \item{\code{"ekw"}}{Exponentiated Kumaraswamy. Three parameters:
-#'       \eqn{\alpha, \beta, \lambda} (fixes \eqn{\gamma = 1, \delta = 0}).
-#'       Adds flexibility to standard Kumaraswamy.}
[…]
