+++ man/gkwreg.Rd
@@ -23,551 +23,234 @@ gkwreg(
 )
 }
 \arguments{
-\item{formula}{An object of class \code{\link[Formula]{Formula}} (or one that
-can be coerced to that class). The formula uses extended syntax to specify
-potentially different linear predictors for each distribution parameter:
-
+\item{formula}{An object of class \code{\link[Formula]{Formula}}. The formula
+uses extended syntax to specify different linear predictors for each
+parameter:
 \code{y ~ model_alpha | model_beta | model_gamma | model_delta | model_lambda}
 
-where:
-\itemize{
-\item \code{y} is the response variable (must be in the open interval (0, 1))
-\item \code{model_alpha} specifies predictors for the \eqn{\alpha} parameter
-\item \code{model_beta} specifies predictors for the \eqn{\beta} parameter
-\item \code{model_gamma} specifies predictors for the \eqn{\gamma} parameter
-\item \code{model_delta} specifies predictors for the \eqn{\delta} parameter
-\item \code{model_lambda} specifies predictors for the \eqn{\lambda} parameter
-}
-
-If a part is omitted or specified as \code{~ 1}, an intercept-only model is
-used for that parameter. Parts corresponding to fixed parameters (determined
-by \code{family}) are automatically ignored. See Details and Examples for
-proper usage.}
-
-\item{data}{A data frame containing the variables specified in \code{formula}.
-Standard R subsetting and missing value handling apply.}
-
-\item{family}{A character string specifying the distribution family from the
-Generalized Kumaraswamy hierarchy. Must be one of:
 \describe{
-\item{\code{"gkw"}}{Generalized Kumaraswamy (default). Five parameters:
-\eqn{\alpha, \beta, \gamma, \delta, \lambda}. Most flexible, suitable
-when data show complex behavior not captured by simpler families.}
-\item{\code{"bkw"}}{Beta-Kumaraswamy. Four parameters:
-\eqn{\alpha, \beta, \gamma, \delta} (fixes \eqn{\lambda = 1}).
-Combines Beta and Kumaraswamy flexibility.}
-\item{\code{"kkw"}}{Kumaraswamy-Kumaraswamy. Four parameters:
-\eqn{\alpha, \beta, \delta, \lambda} (fixes \eqn{\gamma = 1}).
-Alternative four-parameter generalization.}
-\item{\code{"ekw"}}{Exponentiated Kumaraswamy. Three parameters:
-\eqn{\alpha, \beta, \lambda} (fixes \eqn{\gamma = 1, \delta = 0}).
-Adds flexibility to standard Kumaraswamy.}
-\item{\code{"mc"}}{McDonald (Beta Power). Three parameters:
-\eqn{\gamma, \delta, \lambda} (fixes \eqn{\alpha = 1, \beta = 1}).
-Generalization of Beta distribution.}
-\item{\code{"kw"}}{Kumaraswamy. Two parameters: \eqn{\alpha, \beta}
-(fixes \eqn{\gamma = 1, \delta = 0, \lambda = 1}). Computationally
-efficient alternative to Beta with closed-form CDF.}
-\item{\code{"beta"}}{Beta distribution. Two parameters: \eqn{\gamma, \delta}
-(fixes \eqn{\alpha = 1, \beta = 1, \lambda = 1}). Standard choice for
-proportions and rates, corresponds to shape1 = \eqn{\gamma},
-shape2 = \eqn{\delta}.}
+\item{\code{y}}{The response variable, which must be in the (0, 1) interval.}
+\item{\code{model_alpha}}{Predictors for the \eqn{\alpha} parameter.}
+\item{\code{model_beta}}{Predictors for the \eqn{\beta} parameter.}
+\item{\code{model_gamma}}{Predictors for the \eqn{\gamma} parameter.}
[…]
