+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(anova,gkwreg)
 S3method(coef,gkwreg)
 S3method(confint,gkwreg)
 S3method(family,gkwreg)
+S3method(fitted,gkwreg)
 S3method(formula,gkwreg)
 S3method(getCall,gkwreg)
 S3method(logLik,gkwreg)
@@ -26,7 +27,6 @@ S3method(update,gkwreg)
 S3method(vcov,gkwreg)
 export("%>%")
 export(clean_tmb_cache)
-export(fitted.gkwreg)
 export(gkw_control)
 export(gkwreg)
 export(list_tmb_models)
