+++ reference/formula.gkwreg.md
@@ -45,13 +45,13 @@ fit1 <- gkwreg(yield ~ batch + temp, data = GasolineYield, family = "kw")
 #> Warning: NaNs produced
 formula(fit1)
 #> yield ~ batch + temp
-#> <environment: 0x5607a78b54c0>
+#> <environment: 0x55ac5347fa08>
 
 # Two-part formula
 fit2 <- gkwreg(yield ~ temp | batch, data = GasolineYield, family = "kw")
 formula(fit2)
 #> yield ~ temp | batch
-#> <environment: 0x5607a78b54c0>
+#> <environment: 0x55ac5347fa08>
 
 # Five-part formula
 fit3 <- gkwreg(yield ~ temp | batch | temp | 1 | 1,
@@ -60,6 +60,6 @@ fit3 <- gkwreg(yield ~ temp | batch | temp | 1 | 1,
 #> using C++ compiler: ‘g++ (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
 formula(fit3)
 #> yield ~ temp | batch | temp | 1 | 1
-#> <environment: 0x5607a78b54c0>
+#> <environment: 0x55ac5347fa08>
 # }
 ```
