+++ reference/gkwreg.md
@@ -1,20 +1,10 @@
 # Fit Generalized Kumaraswamy Regression Models
 
-Fits regression models using the Generalized Kumaraswamy (GKw) family of
-distributions for modeling response variables strictly bounded in the
-interval (0, 1). The function provides a unified interface for fitting
-seven nested submodels of the GKw family, allowing flexible modeling of
-proportions, rates, and other bounded continuous outcomes through
-regression on distributional parameters.
-
-Maximum Likelihood Estimation is performed via automatic differentiation
-using the TMB (Template Model Builder) package, ensuring computational
-efficiency and numerical accuracy. The interface follows standard R
-regression modeling conventions similar to
-[`lm`](https://rdrr.io/r/stats/lm.html),
-[`glm`](https://rdrr.io/r/stats/glm.html), and
-[`betareg`](https://rdrr.io/pkg/betareg/man/betareg.html), making it
-immediately familiar to R users.
+Fits regression models for response variables strictly bounded in the
+(0, 1) interval using the Generalized Kumaraswamy (GKw) family of
+distributions. It provides a unified interface for fitting the GKw and
+six nested submodels, allowing flexible regression on multiple
+distribution parameters.
 
 ## Usage
 
@@ -43,739 +33,399 @@ gkwreg(
 - formula:
 
   An object of class
-  [`Formula`](https://rdrr.io/pkg/Formula/man/Formula.html) (or one that
-  can be coerced to that class). The formula uses extended syntax to
-  specify potentially different linear predictors for each distribution
+  [`Formula`](https://rdrr.io/pkg/Formula/man/Formula.html). The formula
+  uses extended syntax to specify different linear predictors for each
   parameter:
-
   `y ~ model_alpha | model_beta | model_gamma | model_delta | model_lambda`
 
-  where:
+  `y`
 
-  - `y` is the response variable (must be in the open interval (0, 1))
+  :   The response variable, which must be in the (0, 1) interval.
 
-  - `model_alpha` specifies predictors for the \\\alpha\\ parameter
+  `model_alpha`
 
-  - `model_beta` specifies predictors for the \\\beta\\ parameter
+  :   Predictors for the \\\alpha\\ parameter.
 
-  - `model_gamma` specifies predictors for the \\\gamma\\ parameter
+  `model_beta`
 
-  - `model_delta` specifies predictors for the \\\delta\\ parameter
+  :   Predictors for the \\\beta\\ parameter.
 
-  - `model_lambda` specifies predictors for the \\\lambda\\ parameter
+  `model_gamma`
 
-  If a part is omitted or specified as `~ 1`, an intercept-only model is
-  used for that parameter. Parts corresponding to fixed parameters
-  (determined by `family`) are automatically ignored. See Details and
-  Examples for proper usage.
+  :   Predictors for the \\\gamma\\ parameter.
+
+  `model_delta`
+
+  :   Predictors for the \\\delta\\ parameter.
+
+  `model_lambda`
+
+  :   Predictors for the \\\lambda\\ parameter.
+
+  See Details for examples.
 
 - data:
 
-  A data frame containing the variables specified in `formula`. Standard
-  R subsetting and missing value handling apply.
+  A data frame containing the variables specified in `formula`.
 
 - family:
 
-  A character string specifying the distribution family from the
-  Generalized Kumaraswamy hierarchy. Must be one of:
[…]
