+++ R/gkwreg-core.R
@@ -1,231 +1,547 @@
 #' Fit Generalized Kumaraswamy Regression Models
 #'
 #' @description
-#' Fits regression models for response variables strictly bounded in the (0, 1)
-#' interval using the Generalized Kumaraswamy (GKw) family of distributions.
-#' It provides a unified interface for fitting the GKw and six nested submodels,
-#' allowing flexible regression on multiple distribution parameters.
+#' Fits regression models using the Generalized Kumaraswamy (GKw) family of
+#' distributions for modeling response variables strictly bounded in the interval
+#' (0, 1). The function provides a unified interface for fitting seven nested
+#' submodels of the GKw family, allowing flexible modeling of proportions, rates,
+#' and other bounded continuous outcomes through regression on distributional
+#' parameters.
 #'
-#' @details
-#' Estimation is performed via Maximum Likelihood using automatic differentiation
-#' through the TMB (Template Model Builder) package for efficiency and accuracy.
-#' The modeling interface uses \code{\link[Formula]{Formula}} syntax, similar to
-#' \code{\link[stats]{glm}} and \code{\link[betareg]{betareg}}, allowing different
-#' linear predictors for each distributional parameter.
+#' Maximum Likelihood Estimation is performed via automatic differentiation using
+#' the TMB (Template Model Builder) package, ensuring computational efficiency and
+#' numerical accuracy. The interface follows standard R regression modeling
+#' conventions similar to \code{\link[stats]{lm}}, \code{\link[stats]{glm}}, and
+#' \code{\link[betareg]{betareg}}, making it immediately familiar to R users.
 #'
-#' \subsection{Distribution Families}{
-#' The \code{family} argument selects a distribution from the GKw hierarchy.
-#' Simpler families (\code{"beta"}, \code{"kw"}) are often sufficient and
-#' computationally faster. More complex families (\code{"bkw"}, \code{"gkw"})
-#' add flexibility but risk overfitting. Use \code{\link[stats]{AIC}},
-#' \code{\link[stats]{BIC}}, or \code{\link{anova.gkwreg}} to compare nested models.
-#' }
+#' @param formula An object of class \code{\link[Formula]{Formula}} (or one that
+#'   can be coerced to that class). The formula uses extended syntax to specify
+#'   potentially different linear predictors for each distribution parameter:
 #'
-#' \subsection{Formula Specification}{
-#' The extended formula syntax \code{y ~ model_alpha | model_beta | ...} allows
-#' each parameter's linear predictor to be specified. Parameters are in the
-#' order \eqn{\alpha, \beta, \gamma, \delta, \lambda}.
-#'
-#' If a part is omitted (e.g., \code{y ~ model_alpha}), the remaining parameters
-#' are modeled as intercept-only. Parts corresponding to fixed parameters
-#' (defined by \code{family}) are ignored.
-#'
-#' \preformatted{
-#' # Kw family ("kw") parameters: alpha, beta
-#'
-#' # alpha ~ x1 + x2, beta ~ x3
-#' y ~ x1 + x2 | x3
-#'
-#' # alpha ~ x1, beta ~ 1 (intercept only)
-#' y ~ x1 | 1
-#'
-#' # alpha ~ x1, beta ~ x1 (same formula for both)
-#' y ~ x1
-#' }
-#' }
-#'
-#' \subsection{Optimization and Convergence}{
-#' Fitting is performed by \code{\link[stats]{optim}} or \code{\link[stats]{nlminb}}.
-#' If convergence fails, consider:
[…]
