+++ man/gkwreg.Rd
@@ -23,237 +23,555 @@ gkwreg(
 )
 }
 \arguments{
-\item{formula}{An object of class \code{\link[Formula]{Formula}}. The formula
-uses extended syntax to specify different linear predictors for each
-parameter:
+\item{formula}{An object of class \code{\link[Formula]{Formula}} (or one that
+can be coerced to that class). The formula uses extended syntax to specify
+potentially different linear predictors for each distribution parameter:
+
 \code{y ~ model_alpha | model_beta | model_gamma | model_delta | model_lambda}
 
-\describe{
-\item{\code{y}}{The response variable, which must be in the (0, 1) interval.}
-\item{\code{model_alpha}}{Predictors for the \eqn{\alpha} parameter.}
-\item{\code{model_beta}}{Predictors for the \eqn{\beta} parameter.}
-\item{\code{model_gamma}}{Predictors for the \eqn{\gamma} parameter.}
-\item{\code{model_delta}}{Predictors for the \eqn{\delta} parameter.}
-\item{\code{model_lambda}}{Predictors for the \eqn{\lambda} parameter.}
+where:
+\itemize{
+\item \code{y} is the response variable (must be in the open interval (0, 1))
+\item \code{model_alpha} specifies predictors for the \eqn{\alpha} parameter
+\item \code{model_beta} specifies predictors for the \eqn{\beta} parameter
+\item \code{model_gamma} specifies predictors for the \eqn{\gamma} parameter
+\item \code{model_delta} specifies predictors for the \eqn{\delta} parameter
+\item \code{model_lambda} specifies predictors for the \eqn{\lambda} parameter
 }
-See Details for examples.}
 
-\item{data}{A data frame containing the variables specified in \code{formula}.}
+If a part is omitted or specified as \code{~ 1}, an intercept-only model is
+used for that parameter. Parts corresponding to fixed parameters (determined
+by \code{family}) are automatically ignored. See Details and Examples for
+proper usage.}
 
-\item{family}{A character string specifying the distribution family. Must be one of:
+\item{data}{A data frame containing the variables specified in \code{formula}.
+Standard R subsetting and missing value handling apply.}
+
+\item{family}{A character string specifying the distribution family from the
+Generalized Kumaraswamy hierarchy. Must be one of:
 \describe{
-\item{\code{"gkw"}}{Generalized Kumaraswamy (default). Parameters:
-\eqn{\alpha, \beta, \gamma, \delta, \lambda}.}
-\item{\code{"bkw"}}{Beta-Kumaraswamy. Parameters:
-\eqn{\alpha, \beta, \gamma, \delta} (fixes \eqn{\lambda = 1}).}
-\item{\code{"kkw"}}{Kumaraswamy-Kumaraswamy. Parameters:
-\eqn{\alpha, \beta, \delta, \lambda} (fixes \eqn{\gamma = 1}).}
-\item{\code{"ekw"}}{Exponentiated Kumaraswamy. Parameters:
-\eqn{\alpha, \beta, \lambda} (fixes \eqn{\gamma = 1, \delta = 0}).}
-\item{\code{"mc"}}{McDonald (Beta Power). Parameters:
-\eqn{\gamma, \delta, \lambda} (fixes \eqn{\alpha = 1, \beta = 1}).}
-\item{\code{"kw"}}{Kumaraswamy. Parameters: \eqn{\alpha, \beta}
-(fixes \eqn{\gamma = 1, \delta = 0, \lambda = 1}).}
-\item{\code{"beta"}}{Beta distribution. Parameters: \eqn{\gamma, \delta}
-(fixes \eqn{\alpha = 1, \beta = 1, \lambda = 1}). Corresponds to
-\code{shape1 = \eqn{\gamma}}, \code{shape2 = \eqn{\delta}}.}
[…]
