### 2025-11-16 d945dea
Fix documentation from core gkwreg function.
