+++ R/gkwreg-core.R
@@ -615,7 +615,6 @@
 #' @examples
 #' \donttest{
 #' # SECTION 1: Basic Usage - Getting Started
-#'
 #' # Load packages and data
 #' library(gkwreg)
 #' library(gkwdist)
@@ -882,8 +881,8 @@
 #' )
 #'
 #' # Example 9.2: Complex simulation with all five parameters
-#' set.seed(456)
-#' n <- 1000
+#' set.seed(2203)
+#' n <- 2000
 #' x <- runif(n, -1, 1)
 #'
 #' # True parameters
@@ -905,7 +904,7 @@
 #'   y ~ x | x | x | x | x,
 #'   data = sim_data2,
 #'   family = "gkw",
-#'   control = gkw_control(maxit = 1000)
+#'   control = gkw_control(method = "L-BFGS-B", maxit = 2000)
 #' )
 #' summary(fit_gkw)
 #'
@@ -939,9 +938,10 @@
 #'   data = GasolineYield,
 #'   family = "kw",
 #'   control = gkw_control(
-#'     silent = FALSE,
-#'     trace = 2, # Maximum verbosity
-#'     maxit = 100
+#'     method = "BFGS",
+#'     silent = TRUE,
+#'     trace = 0, # 2, Maximum verbosity
+#'     maxit = 1000
 #'   )
 #' )
 #'
@@ -1027,7 +1027,9 @@
 #'   train <- GasolineYield[folds != fold, ]
 #'   test <- GasolineYield[folds == fold, ]
 #'
-#'   fit_train <- gkwreg(yield ~ batch + temp, train, family = "kw")
+#'   fit_train <- gkwreg(yield ~ batch + temp, train,
+#'     family = "kw"
+#'   )
 #'   pred_test <- predict(fit_train, newdata = test, type = "response")
 #'
 #'   sqrt(mean((test$yield - pred_test)^2))
@@ -1035,7 +1037,6 @@
 #'
 #' cat("Cross-validated RMSE:", mean(cv_rmse), "\n")
 #' }
-#'
 #' @keywords regression models
 #' @export
 gkwreg <- function(formula,
