+++ reference/gkwreg.md
@@ -869,7 +869,6 @@ Maintainer: Lopes, J. E.
 ``` r
 # \donttest{
 # SECTION 1: Basic Usage - Getting Started
-
 # Load packages and data
 library(gkwreg)
 library(gkwdist)
@@ -1830,8 +1829,8 @@ cbind(
 #> beta:x2           -0.2 -0.2419618 0.04607844
 
 # Example 9.2: Complex simulation with all five parameters
-set.seed(456)
-n <- 1000
+set.seed(2203)
+n <- 2000
 x <- runif(n, -1, 1)
 
 # True parameters
@@ -1853,9 +1852,9 @@ fit_gkw <- gkwreg(
   y ~ x | x | x | x | x,
   data = sim_data2,
   family = "gkw",
-  control = gkw_control(maxit = 1000)
+  control = gkw_control(method = "L-BFGS-B", maxit = 2000)
 )
-#> Error: Optimization with nlminb failed: NA/NaN gradient evaluation
+#> Warning: NaNs produced
 summary(fit_gkw)
 #> 
 #> Generalized Kumaraswamy Regression Model Summary
@@ -1863,50 +1862,40 @@ summary(fit_gkw)
 #> Family: gkw 
 #> 
 #> Call:
-#> gkwreg(formula = yield ~ batch + temp, data = GasolineYield, 
-#>     family = "gkw")
+#> gkwreg(formula = y ~ x | x | x | x | x, data = sim_data2, family = "gkw", 
+#>     control = gkw_control(method = "L-BFGS-B", maxit = 2000))
 #> 
 #> Residuals:
 #>     Min  Q1.25%  Median    Mean  Q3.75%     Max 
-#> -0.0272 -0.0123 -0.0010 -0.0013  0.0079  0.0268 
+#> -0.3455 -0.0929  0.0205  0.0337  0.1502  0.6360 
 #> 
 #> Coefficients:
-#>                      Estimate Std. Error  z value Pr(>|z|)    
-#> alpha:(Intercept)  -3.381e+00  7.662e-01   -4.413 1.02e-05 ***
-#> alpha:batch1        9.204e-01  3.181e-02   28.930  < 2e-16 ***
-#> alpha:batch2        6.787e-01  3.407e-02   19.922  < 2e-16 ***
-#> alpha:batch3        7.832e-01  3.400e-02   23.039  < 2e-16 ***
-#> alpha:batch4        5.598e-01  3.185e-02   17.573  < 2e-16 ***
-#> alpha:batch5        5.632e-01  3.396e-02   16.582  < 2e-16 ***
-#> alpha:batch6        5.419e-01  3.403e-02   15.923  < 2e-16 ***
-#> alpha:batch7        3.113e-01  3.188e-02    9.767  < 2e-16 ***
-#> alpha:batch8        2.549e-01  3.394e-02    7.510 5.91e-14 ***
-#> alpha:batch9        1.990e-01  3.794e-02    5.246 1.55e-07 ***
-#> alpha:temp          5.408e-03  1.456e-06 3715.029  < 2e-16 ***
-#> beta:(Intercept)    2.622e+00  3.722e-01    7.043 1.88e-12 ***
-#> gamma:(Intercept)   2.261e+01  1.831e+00   12.346  < 2e-16 ***
-#> delta:(Intercept)  -1.457e-01  1.217e+00   -0.120    0.905    
-#> lambda:(Intercept) -9.556e+00  4.202e+00   -2.274    0.023 *  
+#>                    Estimate Std. Error z value Pr(>|z|)    
+#> alpha:(Intercept)   0.97085    0.61249   1.585   0.1129    
[…]
