### 2025-11-18 76bf936
Fix userChunkSize and userChunkSize with warns while compiling. Update README
