+++ inst/tmb/kwreg.cpp
@@ -286,8 +286,10 @@ Type objective_function<Type>::operator() () {
   DATA_SCALAR(scale2);
   
   DATA_INTEGER(useMeanCache);  // For compatibility
+  
   DATA_INTEGER(calcFitted);
   DATA_INTEGER(userChunkSize);  // For compatibility
+  (void) userChunkSize;
   
   // == PARAMETERS ==
   PARAMETER_VECTOR(beta1);
