+++ index.md
@@ -11,13 +11,14 @@ MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.or
 
 The **gkwreg** package provides a comprehensive and computationally
 efficient framework for regression modeling of data restricted to the
-standard unit interval (0, 1), including proportions, rates, fractions,
-percentages, and bounded indices. While Beta regression is the
-traditional approach for such data, **gkwreg** focuses on the
-**Generalized Kumaraswamy (GKw)** distribution family, offering
-exceptional flexibility by encompassing seven important bounded
-distributions—including Beta and Kumaraswamy—as special or limiting
-cases.
+standard unit interval $(0,1)$, including proportions, rates, fractions,
+percentages, and bounded indices.
+
+While Beta regression is the traditional approach for such data,
+**gkwreg** focuses on the **Generalized Kumaraswamy (GKw)** distribution
+family. This offers exceptional flexibility by encompassing seven
+important bounded distributions—including Beta and Kumaraswamy—as
+special or limiting cases.
 
 The package enables full **distributional regression**, where *all*
 relevant parameters can be modeled as functions of covariates through
@@ -25,6 +26,21 @@ flexible link functions. Maximum Likelihood estimation is performed
 efficiently via the **Template Model Builder (TMB)** framework,
 leveraging **Automatic Differentiation (AD)** for superior computational
 speed, numerical accuracy, and optimization stability.
+
+## Table of Contents
+
+- [Key Features](https://www.google.com/search?q=%23key-features)
+- [Installation](https://www.google.com/search?q=%23installation)
+- [Quick Start](https://www.google.com/search?q=%23quick-start)
+- [Real Data
+  Example](https://www.google.com/search?q=%23real-data-example)
+- [Advanced
+  Features](https://www.google.com/search?q=%23advanced-features)
+- [Mathematical
+  Background](https://www.google.com/search?q=%23mathematical-background)
+- [Comparison with Other
+  Packages](https://www.google.com/search?q=%23comparison-with-other-packages)
+- [Citation](https://www.google.com/search?q=%23citation)
 
 ------------------------------------------------------------------------
 
@@ -35,15 +51,15 @@ speed, numerical accuracy, and optimization stability.
 Model bounded data using the **5-parameter Generalized Kumaraswamy
 (GKw)** distribution and its **seven nested subfamilies**:
 
-| Distribution              | Code   | Parameters Modeled | Fixed Parameters    | \# Params |
-|:--------------------------|:-------|:-------------------|:--------------------|:----------|
-| Generalized Kumaraswamy   | `gkw`  | α, β, γ, δ, λ      | None                | 5         |
-| Beta-Kumaraswamy          | `bkw`  | α, β, γ, δ         | λ = 1               | 4         |
-| Kumaraswamy-Kumaraswamy   | `kkw`  | α, β, δ, λ         | γ = 1               | 4         |
-| Exponentiated Kumaraswamy | `ekw`  | α, β, λ            | γ = 1, δ = 0        | 3         |
-| McDonald (Beta Power)     | `mc`   | γ, δ, λ            | α = 1, β = 1        | 3         |
-| Kumaraswamy               | `kw`   | α, β               | γ = 1, δ = 0, λ = 1 | 2         |
-| Beta                      | `beta` | γ, δ               | α = 1, β = 1, λ = 1 | 2         |
+| Distribution              | Code   | Parameters Modeled                   | Fixed Parameters                    | \# Params |
[…]
