### 2025-11-24 17e335b
Fix main equation README.Rmd
