### 2025-11-24 85ad208
Fix paper.md and other files. Add CONTRIBUTIONG and CODE_OF_CONDUCT.md
