+++ README.md
@@ -487,6 +487,26 @@ optimized C++ code with automatic differentiation:
 
 ------------------------------------------------------------------------
 
+## Contributing
+
+Contributions are welcome! Please see our [Contributing
+Guidelines](CONTRIBUTING.md) and [Code of Conduct](CODE_OF_CONDUCT.md).
+
+### How to Contribute
+
+- Report bugs or request features via [GitHub
+  Issues](https://github.com/evandeilton/gkwreg/issues)
+- Submit pull requests for bug fixes or new features
+- Improve documentation or add examples
+- Share your use cases and feedback
+
+## Community Guidelines
+
+This project follows a [Code of Conduct](CODE_OF_CONDUCT.md). Please
+read it before contributing.
+
+------------------------------------------------------------------------
+
 ## Citation
 
 If you use **gkwreg** in your research, please cite:
@@ -510,7 +530,7 @@ Or use the BibTeX entry:
 ## License
 
 This package is licensed under the **MIT License**. See the
-[LICENSE](https://www.google.com/search?q=LICENSE) file for details.
+[LICENSE](LICENSE.md) file for details.
 
 ------------------------------------------------------------------------
 
