+++ reference/gkwreg.md
@@ -930,7 +930,6 @@ summary(fit_basic)
 # Example 1.2: Model with predictors (uses all defaults)
 # Default: family = "gkw", method = "nlminb", hessian = TRUE
 fit_default <- gkwreg(yield ~ batch + temp, data = GasolineYield)
-#> Warning: NaNs produced
 summary(fit_default)
 #> 
 #> Generalized Kumaraswamy Regression Model Summary
@@ -942,45 +941,45 @@ summary(fit_default)
 #> 
 #> Residuals:
 #>     Min  Q1.25%  Median    Mean  Q3.75%     Max 
-#> -0.0267 -0.0110 -0.0010 -0.0010  0.0079  0.0268 
+#> -0.0272 -0.0123 -0.0010 -0.0013  0.0079  0.0268 
 #> 
 #> Coefficients:
 #>                      Estimate Std. Error  z value Pr(>|z|)    
-#> alpha:(Intercept)  -3.269e+00        NaN      NaN      NaN    
-#> alpha:batch1        9.197e-01  3.191e-02   28.825  < 2e-16 ***
-#> alpha:batch2        6.780e-01  3.414e-02   19.857  < 2e-16 ***
-#> alpha:batch3        7.823e-01  3.414e-02   22.913  < 2e-16 ***
-#> alpha:batch4        5.590e-01  3.189e-02   17.525  < 2e-16 ***
-#> alpha:batch5        5.629e-01  3.410e-02   16.508  < 2e-16 ***
-#> alpha:batch6        5.411e-01  3.409e-02   15.873  < 2e-16 ***
-#> alpha:batch7        3.106e-01  3.191e-02    9.732  < 2e-16 ***
-#> alpha:batch8        2.546e-01  3.410e-02    7.465 8.34e-14 ***
-#> alpha:batch9        1.988e-01  3.809e-02    5.219 1.80e-07 ***
-#> alpha:temp          5.407e-03  1.367e-06 3955.503  < 2e-16 ***
-#> beta:(Intercept)    2.650e+00        NaN      NaN      NaN    
-#> gamma:(Intercept)   2.028e+01  1.089e+01    1.862   0.0627 .  
-#> delta:(Intercept)  -1.216e-01        NaN      NaN      NaN    
-#> lambda:(Intercept) -8.038e+00        NaN      NaN      NaN    
+#> alpha:(Intercept)  -3.381e+00  7.662e-01   -4.413 1.02e-05 ***
+#> alpha:batch1        9.204e-01  3.181e-02   28.930  < 2e-16 ***
+#> alpha:batch2        6.787e-01  3.407e-02   19.922  < 2e-16 ***
+#> alpha:batch3        7.832e-01  3.400e-02   23.039  < 2e-16 ***
+#> alpha:batch4        5.598e-01  3.185e-02   17.573  < 2e-16 ***
+#> alpha:batch5        5.632e-01  3.396e-02   16.582  < 2e-16 ***
+#> alpha:batch6        5.419e-01  3.403e-02   15.923  < 2e-16 ***
+#> alpha:batch7        3.113e-01  3.188e-02    9.767  < 2e-16 ***
+#> alpha:batch8        2.549e-01  3.394e-02    7.510 5.91e-14 ***
+#> alpha:batch9        1.990e-01  3.794e-02    5.246 1.55e-07 ***
+#> alpha:temp          5.408e-03  1.456e-06 3715.029  < 2e-16 ***
+#> beta:(Intercept)    2.622e+00  3.722e-01    7.043 1.88e-12 ***
+#> gamma:(Intercept)   2.261e+01  1.831e+00   12.346  < 2e-16 ***
[…]
