### 2025-11-24 2826fa7
Fix site, readme, add contribution guidelines and code of conduct.
