+++ README.Rmd
@@ -467,7 +467,7 @@ Standard Errors (Hessian inversion)
 
 ## Contributing
 
-Contributions are welcome! Please see our [Contributing Guidelines](CONTRIBUTING.md) and [Code of Conduct](CODE_OF_CONDUCT.md).
+Contributions are welcome! Please see our [Contributing Guidelines](https://github.com/evandeilton/gkwreg/blob/main/CONTRIBUTING.md) and [Code of Conduct](https://github.com/evandeilton/gkwreg/blob/main/CODE_OF_CONDUCT.md).
 
 ### How to Contribute
 
@@ -478,7 +478,7 @@ Contributions are welcome! Please see our [Contributing Guidelines](CONTRIBUTING
 
 ## Community Guidelines
 
-This project follows a [Code of Conduct](CODE_OF_CONDUCT.md). Please read it before contributing.
+This project follows a [Code of Conduct](https://github.com/evandeilton/gkwreg/blob/main/CODE_OF_CONDUCT.md). Please read it before contributing.
 
 -----
 
@@ -504,7 +504,7 @@ Or use the BibTeX entry:
 
 ## License
 
-This package is licensed under the **MIT License**. See the [LICENSE](LICENSE.md) file for details.
+This package is licensed under the **MIT License**. See the [LICENSE](https://github.com/evandeilton/gkwreg/blob/main/LICENSE) file for details.
 
 -----
 
