+++ reference/dot-get_cache_dir.md
@@ -1,7 +1,6 @@
 # Get Cache Directory for Compiled Models
 
-Creates version-specific cache directory for TMB DLLs. Structure:
-\<cache_root\>/\<pkg_name\>/tmb_cache/R/\<pkg_version\>/
+Creates version-specific cache directory for TMB DLLs.
 
 ## Usage
 
