+++ llms.txt
@@ -505,9 +505,9 @@ Standard Errors (Hessian inversion)
 ## Contributing
 
 Contributions are welcome! Please see our [Contributing
-Guidelines](https://evandeilton.github.io/gkwreg/CONTRIBUTING.md) and
-[Code of
-Conduct](https://evandeilton.github.io/gkwreg/CODE_OF_CONDUCT.md).
+Guidelines](https://github.com/evandeilton/gkwreg/blob/main/CONTRIBUTING.md)
+and [Code of
+Conduct](https://github.com/evandeilton/gkwreg/blob/main/CODE_OF_CONDUCT.md).
 
 ### How to Contribute
 
@@ -520,7 +520,7 @@ Conduct](https://evandeilton.github.io/gkwreg/CODE_OF_CONDUCT.md).
 ## Community Guidelines
 
 This project follows a [Code of
-Conduct](https://evandeilton.github.io/gkwreg/CODE_OF_CONDUCT.md).
+Conduct](https://github.com/evandeilton/gkwreg/blob/main/CODE_OF_CONDUCT.md).
 Please read it before contributing.
 
 ------------------------------------------------------------------------
@@ -548,8 +548,8 @@ Or use the BibTeX entry:
 ## License
 
 This package is licensed under the **MIT License**. See the
-[LICENSE](https://evandeilton.github.io/gkwreg/LICENSE.md) file for
-details.
+[LICENSE](https://github.com/evandeilton/gkwreg/blob/main/LICENSE) file
+for details.
 
 ------------------------------------------------------------------------
 
