### 2025-12-01 81199ce
Fix Issue: Package was archived on CRAN CRAN repository db overrides: X-CRAN-Comment: Archived on 2025-11-30 for policy violation. Misuse of ~/.cache/gkwreg and similar on other OSes.
