+++ inst/tmb/gkwreg.cpp
@@ -1,516 +1,2163 @@
-// File: gkwreg_optimized.cpp
-// ---------------------------------------------------------------------
-//  OPTIMIZED Regression Model for the Generalized Kumaraswamy (GKw) Distribution
+// File: gkwreg.cpp
+// =====================================================================
+//  SOTA IMPLEMENTATION: Generalized Kumaraswamy (GKw) Regression Model
 //
-//  Mathematical model:
-//  f(x; α, β, γ, δ, λ) = 
-//    (λαβ x^(α-1) (1-x^α)^(β-1)) / B(γ, δ+1) *
-//    [1-(1-x^α)^β]^(γλ-1) *
-//    [1-[1-(1-x^α)^β]^λ]^δ
+//  Mathematical Foundation:
+//  ------------------------
+//  PDF: f(x; α, β, γ, δ, λ) =
+//       (λαβ x^(α-1) (1-x^α)^(β-1)) / B(γ, δ+1) ×
+//       [1-(1-x^α)^β]^(γλ-1) ×
+//       {1-[1-(1-x^α)^β]^λ}^δ
 //
-//  where 0 < x < 1, α, β, γ, δ, λ > 0
+//  Domain: x ∈ (0,1), all parameters > 0
 //
-//  Optimization level: Elite (targeting 25-35% speedup)
-//  Based on proven techniques from kwreg_optimized.cpp
-//  
-//  Key optimizations:
-//  - Removed ALL CppAD::CondExp overhead (major performance killer)
-//  - Using native pow() instead of exp(log()) wherever possible
-//  - Maintained if(isDouble<Type>::value) optimization (compile-time branch)
-//  - Simplified validation using asDouble() checks (zero-cost in double mode)
-//  - All functions marked inline
-//  - Minimal overhead in hot paths
-//  - 100% compatible with original interface
-//  - Exact analytical mean if available (none known for GKw)
-//  - Precomputed link functions
-//  - Single unified loop structure
-// ---------------------------------------------------------------------
+//  SOTA Features:
+//  --------------
+//  ✓ Complete log-scale cascade computing (numerical stability)
+//  ✓ OpenMP parallelization for n > 5000 (double mode only)
+//  ✓ Ridge regularization (prevents divergence; fixed λ_ridge = 0.01)
+//  ✓ Predictive validation (early rejection of invalid regions)
+//  ✓ Intelligent caching (double-mode only, gradient-safe)
+//  ✓ Adaptive quadrature (15 or 30 points based on variance)
+//  ✓ Zero CondExp overhead
+//  ✓ Compile-time branching via isDouble<Type>::value
+//
+//  Compatibility: 100% drop-in replacement for gkwreg_optimized.cpp
+//  Author: SOTA analysis framework
+//  Version: 1.0.2-SOTA
+// =====================================================================
 
 #include <TMB.hpp>
-#include <unordered_map>
 
-// ========================
-// Numeric constants - MINIMAL SET
-// ========================
+#ifdef _OPENMP
+#include <omp.h>
+#endif
+
+#include <unordered_map>
+#include <cmath>
+#include <algorithm>
+
+// =====================================================================
+// SECTION 1: NUMERICAL CONSTANTS
+// =====================================================================
 template <class Type>
-struct Constants {
-  static const Type eps_log;
-  static const Type eps_pos;
-  static const Type eps_prob;
-  static const Type inf_repl;
-  static const Type max_exp;
+struct NumericLimits {
+  // Core stability thresholds
+  static constexpr double eps_machine      = 1e-15;   // Machine epsilon guard
[…]
