+++ reference/update.gkwreg.md
@@ -190,10 +190,10 @@ anova(m3_0, m3_1, m3_2, m3_3)
 #> Model 4: yield ~ temp | batch | temp
 #> 
 #>      Resid. Df Resid. Dev Df Deviance Pr(>Chi)  
-#> m3_0  27.00000  -48.98863                       
-#> m3_1  26.00000  -52.38707  1  3.39844 0.065258 .
-#> m3_2  17.00000  -52.38706  9  0.00000           
-#> m3_3  16.00000  -43.00466  1 -9.38240           
+#> m3_0  27.00000  -48.83211                       
+#> m3_1  26.00000  -52.22113  1  3.38901 0.065632 .
+#> m3_2  17.00000  -52.22112  9  0.00000           
+#> m3_3  16.00000  -42.82798  1 -9.39315           
 #> ---
 #> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
 
