+++ reference/gkwreg.md
@@ -930,6 +930,7 @@ summary(fit_basic)
 # Example 1.2: Model with predictors (uses all defaults)
 # Default: family = "gkw", method = "nlminb", hessian = TRUE
 fit_default <- gkwreg(yield ~ batch + temp, data = GasolineYield)
+#> Warning: NaNs produced
 summary(fit_default)
 #> 
 #> Generalized Kumaraswamy Regression Model Summary
@@ -941,45 +942,45 @@ summary(fit_default)
 #> 
 #> Residuals:
 #>     Min  Q1.25%  Median    Mean  Q3.75%     Max 
-#> -0.0272 -0.0123 -0.0010 -0.0013  0.0079  0.0268 
+#> -0.0235 -0.0071 -0.0017 -0.0002  0.0072  0.0285 
 #> 
 #> Coefficients:
-#>                      Estimate Std. Error  z value Pr(>|z|)    
-#> alpha:(Intercept)  -3.381e+00  7.662e-01   -4.413 1.02e-05 ***
-#> alpha:batch1        9.204e-01  3.181e-02   28.930  < 2e-16 ***
-#> alpha:batch2        6.787e-01  3.407e-02   19.922  < 2e-16 ***
-#> alpha:batch3        7.832e-01  3.400e-02   23.039  < 2e-16 ***
-#> alpha:batch4        5.598e-01  3.185e-02   17.573  < 2e-16 ***
-#> alpha:batch5        5.632e-01  3.396e-02   16.582  < 2e-16 ***
-#> alpha:batch6        5.419e-01  3.403e-02   15.923  < 2e-16 ***
-#> alpha:batch7        3.113e-01  3.188e-02    9.767  < 2e-16 ***
-#> alpha:batch8        2.549e-01  3.394e-02    7.510 5.91e-14 ***
-#> alpha:batch9        1.990e-01  3.794e-02    5.246 1.55e-07 ***
-#> alpha:temp          5.408e-03  1.456e-06 3715.029  < 2e-16 ***
-#> beta:(Intercept)    2.622e+00  3.722e-01    7.043 1.88e-12 ***
-#> gamma:(Intercept)   2.261e+01  1.831e+00   12.346  < 2e-16 ***
-#> delta:(Intercept)  -1.457e-01  1.217e+00   -0.120    0.905    
-#> lambda:(Intercept) -9.556e+00  4.202e+00   -2.274    0.023 *  
+#>                     Estimate Std. Error z value Pr(>|z|)    
+#> alpha:(Intercept)  -1.419893   0.666514  -2.130   0.0331 *  
+#> alpha:batch1        0.918498   0.032368  28.377  < 2e-16 ***
+#> alpha:batch2        0.676328   0.034627  19.532  < 2e-16 ***
+#> alpha:batch3        0.781489   0.034587  22.595  < 2e-16 ***
+#> alpha:batch4        0.557260   0.032509  17.142  < 2e-16 ***
+#> alpha:batch5        0.562416   0.034498  16.303  < 2e-16 ***
+#> alpha:batch6        0.539377   0.034694  15.547  < 2e-16 ***
+#> alpha:batch7        0.308696   0.032499   9.499  < 2e-16 ***
+#> alpha:batch8        0.254507   0.034578   7.360 1.83e-13 ***
+#> alpha:batch9        0.198134   0.038561   5.138 2.77e-07 ***
+#> alpha:temp          0.005403        NaN     NaN      NaN    
+#> beta:(Intercept)    5.511544   2.754889   2.001   0.0454 *  
+#> gamma:(Intercept)   2.560940   2.294913   1.116   0.2645    
[…]
