+++ reference/dot-check_and_compile_TMB_code.md
@@ -1,7 +1,7 @@
 # Check and Compile TMB Model Code
 
-Ensures a TMB model is compiled and loaded. Uses persistent cache to
-avoid recompilation across sessions.
+Ensures a TMB model is compiled and loaded. Uses session-specific cache
+within tempdir() to avoid recompilation during the same R session.
 
 ## Usage
 
