+++ DESCRIPTION
@@ -1,12 +1,20 @@
 Package: gkwreg
 Title: Generalized Kumaraswamy Regression Models for Bounded Data
-Version: 2.1.7
+Version: 2.1.10
 Authors@R: c(person(given = "José Evandeilton",
                     family = "Lopes",
                     role = c("aut", "cre"),
                     email = "evandeilton@gmail.com",
                     comment = c(ORCID = "0009-0007-5887-4084")))
-Description: Implements regression models for bounded continuous data in the open interval (0,1) using the five-parameter Generalized Kumaraswamy distribution. Supports modeling all distribution parameters (alpha, beta, gamma, delta, lambda) as functions of predictors through various link functions. Provides efficient maximum likelihood estimation via Template Model Builder ('TMB'), offering comprehensive diagnostics, model comparison tools, and simulation methods. Particularly useful for analyzing proportions, rates, indices, and other bounded response data with complex distributional features not adequately captured by simpler models.
+Description:
+    Implements regression models for bounded continuous data in the open interval (0,1) 
+    using the five-parameter Generalized Kumaraswamy distribution. Supports modeling 
+    all distribution parameters (alpha, beta, gamma, delta, lambda) as functions of 
+    predictors through various link functions. Provides efficient maximum likelihood 
+    estimation via Template Model Builder ('TMB'), offering comprehensive diagnostics,
+    model comparison tools, and simulation methods. Particularly useful for analyzing 
+    proportions, rates, indices, and other bounded response data with complex 
+    distributional features not adequately captured by simpler models.
 License: MIT + file LICENSE
 URL: https://github.com/evandeilton/gkwreg, https://evandeilton.github.io/gkwreg/
 BugReports: https://github.com/evandeilton/gkwreg/issues
