### 2025-12-05 cdb8e17
Fix poroblematic import for C++ compiling
