+++ R/gkwreg-compile-tmb.R
@@ -1,7 +1,7 @@
 # TMB Model Compilation and Loading System for gkwreg Package
 # 
-# Simple and robust TMB model compilation with caching.
-# Critical: Never unloads TMB model DLLs to avoid finalization errors.
+# Simple and robust TMB model compilation with caching in tempdir().
+# All functions are internal - called automatically by model fitting functions.
 
 # Internal package environment for DLL state management
 .gkwreg_env <- new.env(parent = emptyenv())
@@ -11,7 +11,8 @@
 #'
 #' @description
 #' Ensures a TMB model is compiled and loaded. Uses session-specific cache
-#' within tempdir() to avoid recompilation during the same R session.
+#' within tempdir() that is automatically cleaned when R exits.
+#' This is an internal function called automatically by model fitting functions.
 #'
 #' @param dll_name Character, base name of the C++ file (e.g., "kwreg", "ekwreg")
 #' @param pkg_name Character, package name (default: "gkwreg")
@@ -22,6 +23,7 @@
 #'
 #' @importFrom TMB compile dynlib
 #' @keywords internal
+#' @noRd
 .check_and_compile_TMB_code <- function(dll_name,
                                         pkg_name = "gkwreg",
                                         force_recompile = FALSE,
@@ -47,7 +49,7 @@
     stop("TMB source file not found: inst/tmb/", dll_name, ".cpp")
   }
   
-  # Setup cache directory
+  # Setup cache directory (in tempdir - auto-cleaned on exit)
   cache_dir <- .get_cache_dir(pkg_name)
   dll_ext <- .Platform$dynlib.ext
   cached_dll <- file.path(cache_dir, paste0(dll_name, dll_ext))
@@ -83,7 +85,7 @@
       warning("Failed to cache DLL (will need recompilation next session)")
     }
     
-    # Clean up temp directory
+    # Clean up temp compilation directory
     unlink(dirname(compiled_dll), recursive = TRUE, force = TRUE)
     
     log_msg("Compiled and loaded: ", dll_name)
@@ -114,6 +116,7 @@
 #' @return Logical, TRUE if successful
 #'
 #' @keywords internal
+#' @noRd
 .load_dll <- function(dll_path, dll_name, verbose = FALSE) {
   
   log_msg <- function(...) if (verbose) message("[TMB] ", ...)
@@ -166,6 +169,7 @@
 #' @return Logical
 #'
 #' @keywords internal
+#' @noRd
 .is_dll_functional <- function(dll_name, verbose = FALSE) {
   
   # Check if in R's DLL table
@@ -198,6 +202,7 @@
 #' @return Path to compiled DLL
 #'
 #' @keywords internal
+#' @noRd
 .compile_model <- function(cpp_file, dll_name, verbose = FALSE) {
   
   log_msg <- function(...) if (verbose) message("[TMB] ", ...)
@@ -205,7 +210,7 @@
   # Create temporary compilation directory
   temp_dir <- file.path(
     tempdir(),
-    paste0(dll_name, "_", format(Sys.time(), "%Y%m%d_%H%M%S"))
+    paste0(dll_name, "_compile_", format(Sys.time(), "%Y%m%d_%H%M%S"))
   )
   dir.create(temp_dir, recursive = TRUE, showWarnings = FALSE)
   
@@ -242,7 +247,8 @@
     dll_path <- expected_dll
   } else {
     # Search for any DLL in temp directory
-    dll_files <- list.files(temp_dir, pattern = paste0("\\", dll_ext, "$"), full.names = TRUE)
[…]
