+++ man/dot-compile_model.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gkwreg-compile-tmb.R
-\name{.compile_model}
-\alias{.compile_model}
-\title{Compile TMB Model}
-\usage{
-.compile_model(cpp_file, dll_name, verbose = FALSE)
-}
-\arguments{
-\item{cpp_file}{Path to source file}
-
-\item{dll_name}{Base name for output}
-
-\item{verbose}{Logical}
-}
-\value{
-Path to compiled DLL
-}
-\description{
-Compiles a TMB C++ template in a temporary directory.
-}
-\keyword{internal}
