+++ man/precompile_tmb_models.Rd
@@ -1,21 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gkwreg-compile-tmb.R
-\name{precompile_tmb_models}
-\alias{precompile_tmb_models}
-\title{Precompile All TMB Models}
-\usage{
-precompile_tmb_models(pkg_name = "gkwreg", verbose = TRUE)
-}
-\arguments{
-\item{pkg_name}{Package name (default: "gkwreg")}
-
-\item{verbose}{Logical}
-}
-\value{
-Invisibly returns list of compilation results
-}
-\description{
-Compiles and caches all TMB models found in inst/tmb/.
-Useful to run after package installation to avoid compilation delays
-during first use.
-}
