+++ man/dot-get_cache_dir.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gkwreg-compile-tmb.R
-\name{.get_cache_dir}
-\alias{.get_cache_dir}
-\title{Get Cache Directory for Compiled Models}
-\usage{
-.get_cache_dir(pkg_name)
-}
-\arguments{
-\item{pkg_name}{Package name}
-}
-\value{
-Path to cache directory within tempdir()
-}
-\description{
-Creates version-specific cache directory for TMB DLLs within tempdir().
-Uses temporary directory to comply with CRAN policies (no user directory access).
-Cache is session-specific and will not persist across R sessions.
-}
-\keyword{internal}
