+++ man/dot-check_and_compile_TMB_code.Rd
@@ -1,30 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gkwreg-compile-tmb.R
-\name{.check_and_compile_TMB_code}
-\alias{.check_and_compile_TMB_code}
-\title{Check and Compile TMB Model Code}
-\usage{
-.check_and_compile_TMB_code(
-  dll_name,
-  pkg_name = "gkwreg",
-  force_recompile = FALSE,
-  verbose = FALSE
-)
-}
-\arguments{
-\item{dll_name}{Character, base name of the C++ file (e.g., "kwreg", "ekwreg")}
-
-\item{pkg_name}{Character, package name (default: "gkwreg")}
-
-\item{force_recompile}{Logical, force recompilation (default: FALSE)}
-
-\item{verbose}{Logical, print diagnostic messages (default: FALSE)}
-}
-\value{
-Invisibly returns a list with DLL information
-}
-\description{
-Ensures a TMB model is compiled and loaded. Uses session-specific cache
-within tempdir() to avoid recompilation during the same R session.
-}
-\keyword{internal}
