+++ man/clean_tmb_cache.Rd
@@ -1,20 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gkwreg-compile-tmb.R
-\name{clean_tmb_cache}
-\alias{clean_tmb_cache}
-\title{Clean TMB Cache}
-\usage{
-clean_tmb_cache(pkg_name = "gkwreg", verbose = TRUE)
-}
-\arguments{
-\item{pkg_name}{Package name (default: "gkwreg")}
-
-\item{verbose}{Logical}
-}
-\value{
-Invisibly returns number of files deleted
-}
-\description{
-Removes all cached compiled TMB models. Useful for troubleshooting or
-after package updates.
-}
