+++ NAMESPACE
@@ -26,12 +26,9 @@ S3method(terms,gkwreg)
 S3method(update,gkwreg)
 S3method(vcov,gkwreg)
 export("%>%")
-export(clean_tmb_cache)
 export(gkw_control)
 export(gkwreg)
-export(list_tmb_models)
 export(lrtest)
-export(precompile_tmb_models)
 export(response)
 import(RcppArmadillo)
 import(gkwdist)

[…]
