### 2025-12-06 bc28ddc
CRAN Resubmission v2.1.12: Fix Imports and DESCRIPTION
- Remove RcppArmadillo and RcppEigen from Imports (kept in LinkingTo).
- Update NAMESPACE to remove unused imports.
- Add single quotes to technical term in DESCRIPTION.
- Verify compliance with CRAN cache policy (tempdir usage).
