+++ NAMESPACE
@@ -6,7 +6,6 @@ S3method(anova,gkwreg)
 S3method(coef,gkwreg)
 S3method(confint,gkwreg)
 S3method(family,gkwreg)
-S3method(fitted,gkwreg)
 S3method(formula,gkwreg)
 S3method(getCall,gkwreg)
 S3method(logLik,gkwreg)
@@ -26,11 +25,11 @@ S3method(terms,gkwreg)
 S3method(update,gkwreg)
 S3method(vcov,gkwreg)
 export("%>%")
+export(fitted.gkwreg)
 export(gkw_control)
 export(gkwreg)
 export(lrtest)
 export(response)
-import(RcppArmadillo)
 import(gkwdist)
 import(graphics)
 importFrom(Formula,as.Formula)
