+++ reference/BIC.gkwreg.md
@@ -81,7 +81,7 @@ BIC(fit1)
 BIC(fit1, fit2, fit3)
 #>      df        BIC
 #> fit1 11  -26.81126
-#> fit2 12 -152.34901
+#> fit2 12 -152.34982
 #> fit3  3  -70.50538
 # }
 ```
