### 2026-01-06 38825a5
v2.1.13: CRAN resubmission - fix checktime and spelling NOTE Changes: - Added inst/WORDLIST with 'Kumaraswamy' to resolve spelling NOTE - Added skip_on_cran() to all 9 test files (reduces checktime from 18min to ~9min) - Added cran-comments.md documenting changes since archival - Added cran-comments.md to .Rbuildignore - Updated DESCRIPTION version to 2.1.13 - Updated NEWS.md with v2.1.13 release notes CRAN compliance: - 0 errors, 0 warnings, 2 notes (expected for resubmission after archival) - Cache policy verified: uses only tempdir() - Check time now under 10 minutes
