+++ tests/testthat/test-gkwreg-beta-vs-betareg.R
@@ -6,6 +6,9 @@
 #              parameterization differences
 # ==============================================================================
 
+# Skip on CRAN to reduce check time
+skip_on_cran()
+
 library(testthat)
 library(gkwreg)
 library(gkwdist)
