+++ tests/testthat/test-gkwreg-anova-tests.R
@@ -4,6 +4,9 @@
 #              covering single models, nested model comparisons, likelihood ratio tests,
 #              and edge cases
 
+# Skip on CRAN to reduce check time
+skip_on_cran()
+
 library(testthat)
 library(gkwreg)
 library(gkwdist)
