+++ R/gkwreg-compile-tmb.R
@@ -1,5 +1,5 @@
 # TMB Model Compilation and Loading System for gkwreg Package
-# 
+#
 # Simple and robust TMB model compilation with caching in tempdir().
 # All functions are internal - called automatically by model fitting functions.
 
@@ -28,37 +28,36 @@
                                         pkg_name = "gkwreg",
                                         force_recompile = FALSE,
                                         verbose = FALSE) {
-  
   log_msg <- function(...) if (verbose) message("[TMB] ", ...)
-  
+
   # Check session cache first
   cache_key <- paste0("TMB_", dll_name, "_loaded")
-  
+
   if (!force_recompile && isTRUE(get0(cache_key, envir = .gkwreg_env, ifnotfound = FALSE))) {
     if (.is_dll_functional(dll_name, verbose)) {
       log_msg("Using already loaded model: ", dll_name)
       return(invisible(list(name = dll_name, cached = TRUE)))
     }
   }
-  
+
   log_msg("Preparing TMB model: ", dll_name)
-  
+
   # Get source file path
   cpp_file <- system.file("tmb", paste0(dll_name, ".cpp"), package = pkg_name)
   if (cpp_file == "" || !file.exists(cpp_file)) {
     stop("TMB source file not found: inst/tmb/", dll_name, ".cpp")
   }
-  
+
   # Setup cache directory (in tempdir - auto-cleaned on exit)
   cache_dir <- .get_cache_dir(pkg_name)
   dll_ext <- .Platform$dynlib.ext
   cached_dll <- file.path(cache_dir, paste0(dll_name, dll_ext))
-  
+
   # Determine if compilation needed
-  need_compile <- force_recompile || 
+  need_compile <- force_recompile ||
     !file.exists(cached_dll) ||
     file.info(cpp_file)$mtime > file.info(cached_dll)$mtime
-  
+
   if (!need_compile) {
     # Try loading cached DLL
     if (.load_dll(cached_dll, dll_name, verbose)) {
@@ -70,30 +69,30 @@
       need_compile <- TRUE
     }
   }
-  
+
   # Compile if necessary
   if (need_compile) {
     compiled_dll <- .compile_model(cpp_file, dll_name, verbose)
-    
+
     # Load the compiled DLL
     if (!.load_dll(compiled_dll, dll_name, verbose)) {
       stop("Failed to load freshly compiled DLL: ", dll_name)
     }
-    
+
     # Copy to cache
     if (!file.copy(compiled_dll, cached_dll, overwrite = TRUE)) {
       warning("Failed to cache DLL (will need recompilation next session)")
     }
-    
+
     # Clean up temp compilation directory
     unlink(dirname(compiled_dll), recursive = TRUE, force = TRUE)
-    
+
     log_msg("Compiled and loaded: ", dll_name)
   }
-  
+
   # Mark as loaded in session cache
   assign(cache_key, TRUE, envir = .gkwreg_env)
-  
+
   invisible(list(
     name = dll_name,
     path = cached_dll,
@@ -118,45 +117,52 @@
 #' @keywords internal
 #' @noRd
 .load_dll <- function(dll_path, dll_name, verbose = FALSE) {
-  
   log_msg <- function(...) if (verbose) message("[TMB] ", ...)
-  
+
   if (!file.exists(dll_path)) {
     log_msg("DLL not found: ", dll_path)
[…]
