+++ llms.txt
@@ -551,18 +551,11 @@ This package is licensed under the **MIT License**. See the
 [LICENSE](https://github.com/evandeilton/gkwreg/blob/main/LICENSE) file
 for details.
 
-------------------------------------------------------------------------
-
-## Author and Maintainer
-
-**José Evandeilton Lopes (Lopes, J. E.)** \| <evandeilton@gmail.com> \|
-[GitHub](https://github.com/evandeilton) \|
-[ORCID](https://orcid.org/0009-0007-5887-4084)
-
-LEG - Laboratório de Estatística e Geoinformação \| UFPR - Universidade
-Federal do Paraná, Brazil
-
-------------------------------------------------------------------------
+|                                                                                                                                                                    |
+|--------------------------------------------------------------------------------------------------------------------------------------------------------------------|
+| \## Author and Maintainer                                                                                                                                          |
+| **José Evandeilton Lopes (Lopes, J. E.)** \| <evandeilton@gmail.com> \| [GitHub](https://github.com/evandeilton) \| [ORCID](https://orcid.org/0009-0007-5887-4084) |
+| LEG - Laboratório de Estatística e Geoinformação \| UFPR - Universidade Federal do Paraná, Brazil                                                                  |
 
 # Package index
 
@@ -599,7 +592,7 @@ Federal do Paraná, Brazil
   Parameters
 - [`family(`*`<gkwreg>`*`)`](https://evandeilton.github.io/gkwreg/reference/family.gkwreg.md)
   : Extract Family from GKw Regression Model
-- [`fitted.gkwreg()`](https://evandeilton.github.io/gkwreg/reference/fitted.gkwreg.md)
+- [`fitted(`*`<gkwreg>`*`)`](https://evandeilton.github.io/gkwreg/reference/fitted.gkwreg.md)
   : Extract Fitted Values from a Generalized Kumaraswamy Regression
   Model
 - [`formula(`*`<gkwreg>`*`)`](https://evandeilton.github.io/gkwreg/reference/formula.gkwreg.md)
