+++ authors.md
@@ -10,14 +10,14 @@
 Source:
 [`DESCRIPTION`](https://github.com/evandeilton/gkwreg/blob/main/DESCRIPTION)
 
-Lopes J (2025). *gkwreg: Generalized Kumaraswamy Regression Models for
-Bounded Data*. R package version 2.1.12,
+Lopes J (2026). *gkwreg: Generalized Kumaraswamy Regression Models for
+Bounded Data*. R package version 2.1.13,
 <https://github.com/evandeilton/gkwreg>.
 
     @Manual{,
       title = {gkwreg: Generalized Kumaraswamy Regression Models for Bounded Data},
       author = {José Evandeilton Lopes},
-      year = {2025},
-      note = {R package version 2.1.12},
+      year = {2026},
+      note = {R package version 2.1.13},
       url = {https://github.com/evandeilton/gkwreg},
     }
