### 2026-01-07 1d8deba
Fix paper.md due to JOSS reviewing process
