### 2026-01-08 ee2c6be
**FIXED**: clang-san runtime error (integer overflow)    - Fixed  overflow in cache key generation for TMB models    - Added  function to prevent undefined behavior when      distribution parameters reach extreme values during optimization    - Affects: ekwreg.cpp, kwreg.cpp, mcreg.cpp, kkwreg.cpp, bkwreg.cpp, gkwreg.cpp
