+++ gkwreg.Rcheck/00_pkg_src/gkwreg/R/gkwreg-utils.R
@@ -1,0 +1,565 @@
+#' Get Parameter Information for a GKw Family Distribution
+#'
+#' @param family The GKw family distribution name.
+#' @return A list with parameter information.
+#' @keywords internal
+.get_family_param_info <- function(family) {
+  # Define parameter information and TMB parameter positions for each family
+  family_params <- list(
+    gkw = list(
+      names = c("alpha", "beta", "gamma", "delta", "lambda"),
+      n = 5,
+      fixed = list(),
+      positions = list(alpha = 1, beta = 2, gamma = 3, delta = 4, lambda = 5)
+    ),
+    bkw = list(
+      names = c("alpha", "beta", "gamma", "delta"),
+      n = 4,
+      fixed = list(lambda = 1),
+      positions = list(alpha = 1, beta = 2, gamma = 3, delta = 4)
+    ),
+    kkw = list(
+      names = c("alpha", "beta", "delta", "lambda"),
+      n = 4,
+      fixed = list(gamma = 1),
+      positions = list(alpha = 1, beta = 2, delta = 3, lambda = 4)
+    ),
+    ekw = list(
+      names = c("alpha", "beta", "lambda"),
+      n = 3,
+      fixed = list(gamma = 1, delta = 0),
+      positions = list(alpha = 1, beta = 2, lambda = 3)
+    ),
+    mc = list(
+      names = c("gamma", "delta", "lambda"),
+      n = 3,
+      fixed = list(alpha = 1, beta = 1),
+      positions = list(gamma = 1, delta = 2, lambda = 3)
+    ),
+    kw = list(
+      names = c("alpha", "beta"),
+      n = 2,
+      fixed = list(gamma = 1, delta = 0, lambda = 1),
+      positions = list(alpha = 1, beta = 2)
+    ),
+    beta = list(
+      names = c("gamma", "delta"),
+      n = 2,
+      fixed = list(alpha = 1, beta = 1, lambda = 1),
+      positions = list(gamma = 1, delta = 2)
+    )
+  )
+
+  # Return parameter information for the specified family
+  return(family_params[[family]])
+}
+
+
+#' Process Formula Parts from a Formula Object
+#'
+#' @param formula_obj Formula object created with the Formula package.
+#' @param param_names Names of the parameters for the specified family.
+#' @param fixed_params List of fixed parameters.
+#' @param data Data frame containing the variables.
+#' @return A list of formula objects for each parameter.
+#' @keywords internal
+.process_formula_parts <- function(formula_obj, param_names, fixed_params, data) {
+  # Get non-fixed parameters
+  non_fixed_params <- setdiff(param_names, names(fixed_params))
+
+  # Extract the response variable from the formula
+  resp_var <- as.character(formula_obj[[2]])
+
+  # Create list to store formulas for each parameter
+  formula_list <- list()
+
+  # Get the max number of RHS parts in the formula
+  n_parts <- length(attr(Formula::Formula(formula_obj), "rhs"))
+
+  # Process each non-fixed parameter
+  for (i in seq_along(non_fixed_params)) {
+    param <- non_fixed_params[i]
+
+    if (i <= n_parts) {
+      # Extract the ith part of the formula
+      rhs_part <- stats::formula(formula_obj, rhs = i, lhs = 1)[[3]]
+
+      # Check if this part is just a dot
+      if (identical(as.character(rhs_part), ".") || identical(as.character(rhs_part), "1")) {
+        # Use intercept-only model
+        formula_list[[param]] <- Formula::as.Formula(paste(resp_var, "~", "1"))
+      } else {
+        # Use the specified formula part
[…]
