+++ gkwreg.Rcheck/00_pkg_src/gkwreg/inst/paper/paper.bib
@@ -1,57 +1,0 @@
-@article{kumaraswamy1980,
-  author = {Kumaraswamy, P.},
-  title = {A generalized probability density function for double-bounded random processes},
-  journal = {Journal of Hydrology},
-  volume = {46},
-  number = {1-2},
-  pages = {79--88},
-  year = {1980},
-  doi = {10.1016/0022-1694(80)90036-0}
-}
-
-@article{carrasco2010,
-  author = {Carrasco, J. M. F. and Ferrari, S. L. P. and Cordeiro, G. M.},
-  title = {A new generalized {K}umaraswamy distribution},
-  journal = {arXiv preprint arXiv:1004.0911},
-  year = {2010}
-}
-
-@Manual{gkwreg,
-  title = {gkwreg: An R Package for Generalized Kumaraswamy Regression Models for Bounded Data},
-  author = {Lopes, JE},
-  year = {2025},
-  note = {R package version 1.0.11},
-}
-
-@article{Kristensen2016,
-  author = {Kristensen, K. and Nielsen, A. and Berg, C. W. and Skaug, H. and Bell, B. M.},
-  title = {{TMB}: Automatic Differentiation and {L}aplace Approximation},
-  journal = {Journal of Statistical Software},
-  volume = {70},
-  number = {5},
-  pages = {1--21},
-  year = {2016},
-  doi = {10.18637/jss.v070.i05}
-}
-
-@article{Zeileis2010,
-  author = {Zeileis, A. and Croissant, Y.},
-  title = {Extended Model Formulas in {R}: Multiple Parts and Multiple Responses},
-  journal = {Journal of Statistical Software},
-  volume = {34},
-  number = {1},
-  pages = {1--13},
-  year = {2010},
-  doi = {10.18637/jss.v034.i01}
-}
-
-@article{Dunn1996,
-  author = {Dunn, P. K. and Smyth, G. K.},
-  title = {Randomized Quantile Residuals},
-  journal = {Journal of Computational and Graphical Statistics},
-  volume = {5},
-  number = {3},
-  pages = {236--244},
-  year = {1996},
-  doi = {10.1080/10618600.1996.10474708}
-}
