+++ gkwreg.Rcheck/00_pkg_src/gkwreg/inst/tmb/kwreg.cpp
@@ -1,429 +1,0 @@
-// File: kwreg.cpp
-// ---------------------------------------------------------------------
-//  Regression Model for the Kumaraswamy (Kw) Distribution
-//
-//  This is a two-parameter special case of the Generalized Kumaraswamy distribution
-//  with γ = 1, δ = 0, and λ = 1 fixed. The probability density function is:
-//
-//  f(x; α, β) = αβx^(α-1)(1-x^α)^(β-1)
-//
-//  where 0 < x < 1, and α, β > 0.
-//
-//  The cumulative distribution function has a simple closed form:
-//  F(x; α, β) = 1-(1-x^α)^β
-//
-//  This distribution is similar to the beta distribution but with a simpler
-//  mathematical form and closed-form expressions for both CDF and inverse CDF.
-// ---------------------------------------------------------------------
-
-
-#include <TMB.hpp>
-#include <unordered_map>
-
-// ========================
-// Numeric constants - MINIMAL SET
-// ========================
-template <class Type>
-struct Constants {
-  static const Type eps_log;
-  static const Type eps_pos;
-  static const Type eps_prob;
-  static const Type inf_repl;
-  static const Type max_exp;
-};
-
-template <class Type>
-const Type Constants<Type>::eps_log  = Type(1e-15);
-template <class Type>
-const Type Constants<Type>::eps_pos  = Type(1e-10);
-template <class Type>
-const Type Constants<Type>::eps_prob = Type(1e-12);
-template <class Type>
-const Type Constants<Type>::inf_repl = Type(1e10);
-template <class Type>
-const Type Constants<Type>::max_exp  = Type(30);
-
-// ========================
-// Safe integer cast - prevents overflow when converting large doubles to int
-// ========================
-inline int safe_int_cast(double value) {
-  const double max_int = 2147483647.0;  // INT_MAX
-  const double min_int = -2147483648.0; // INT_MIN
-  if (value > max_int) return static_cast<int>(max_int);
-  if (value < min_int) return static_cast<int>(min_int);
-  return static_cast<int>(value);
-}
-
-// ========================
-// Safe numerical operations - INLINE, NO CondExp
-// ========================
-template <class Type>
-inline Type safeLog(const Type &x) {
-  return (x <= Constants<Type>::eps_pos) ? 
-  -Constants<Type>::inf_repl : log(x);
-}
-
-template <class Type>
-inline Type safeExp(const Type &x) {
-  double x_d = asDouble(x);
-  if (x_d >  30.0) return exp(Type(30.0));
-  if (x_d < -30.0) return exp(Type(-30.0));
-  return exp(x);
-}
-
-// ========================
-// Link functions - SIMPLIFIED, INLINE, NO CondExp
-// ========================
-template <class Type>
-inline Type apply_positive_link(const Type &eta, const int link_type, 
-                                const Type &scale_factor) {
-  Type result;
-  Type min_val = Constants<Type>::eps_pos;
-  
-  switch (link_type) {
-  case 1: // log
-    result = exp(eta);
-    break;
-  case 2: // logit
-    result = scale_factor / (Type(1.0) + exp(-eta));
-    break;
-  case 3: // probit
[…]

[…]
