### 2026-01-08 a2c76cf
fix: resolve macOS ARM64 binary download issue by forcing source installation
