+++ .github/workflows/R-CMD-check.yaml
@@ -29,27 +29,28 @@ jobs:
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
           
-      # Configuração específica para Windows - adiciona Rtools
+      # Configuração para Windows - adiciona Rtools
       - uses: r-lib/actions/setup-r-dependencies@v2
-        if: matrix.config.os == 'windows-latest'
+        if: runner.os == 'Windows'
         with:
           extra-packages: any::rcmdcheck
           needs: check
           windows-path-include-mingw: true
-          
-      # Instalação de dependências do sistema para Windows
-      - name: Install system dependencies (Windows)
-        if: matrix.config.os == 'windows-latest'
-        run: |
-          # Tentar pacotes específicos do MinGW que estão disponíveis via Chocolatey
-          choco install --no-progress mingw-w64 -y || echo "MinGW installation attempted"
-          # Verificar se temos acesso ao compilador
-          g++ --version || echo "G++ not found in PATH"
-          
-      - name: Install pak
-        run: Rscript -e 'install.packages("pak", repos = "https://cran.rstudio.com")'
-        
+
+      # Configuração para macOS - força source quando binário falhar
       - uses: r-lib/actions/setup-r-dependencies@v2
+        if: runner.os == 'macOS'
+        with:
+          extra-packages: any::rcmdcheck
+          needs: check
+          pak-version: devel
+        env:
+          PKG_SYSREQS: true
+          R_PKG_TYPE: source
+
+      # Configuração para Linux
+      - uses: r-lib/actions/setup-r-dependencies@v2
+        if: runner.os == 'Linux'
         with:
           extra-packages: any::rcmdcheck
           needs: check
