### 2026-01-08 607543a
fix(ci): disable sysreqs lookup on macOS ARM64 to resolve pak error
