### 2026-01-18 #8991 @jbytecode
@evandeilton - Hi, I've read the manuscript. Looks almost good to me. Could you please perform these tasks:

- Create a tagged release in the repo. This tagged release is supposed to contain the reviewed repository.
- Use this tagged release to create a Zenodo archive.
- Please report me the Zenodo DOI and the tagged release.
- The metadata in Zenodo archive (author list, their ORCIDs, license information) should match those in the paper and software.

Thank you in advance.
