+++ reference/fitted.gkwreg.md
@@ -138,12 +138,12 @@ comparison <- data.frame(
 )
 head(comparison)
 #>          EKW         KW       Beta Diff_EKW_KW Diff_EKW_Beta
-#> 1 0.12232274 0.12232274 0.12232274           0             0
+#> 1 0.12232728 0.12232728 0.12232728           0             0
 #> 2 0.00010000 0.00010000 0.00010000           0             0
-#> 3 0.35297369 0.35297369 0.35297369           0             0
+#> 3 0.35297443 0.35297443 0.35297443           0             0
 #> 4 0.46181798 0.46181798 0.46181798           0             0
-#> 5 0.06734392 0.06734392 0.06734392           0             0
-#> 6 0.12847689 0.12847689 0.12847689           0             0
+#> 5 0.06734871 0.06734871 0.06734871           0             0
+#> 6 0.12862292 0.12862292 0.12862292           0             0
 
 # Visualize differences
 par(mfrow = c(1, 2))
