### 2026-01-18 #8991 @evandeilton
Hi @jbytecode, Thank you for the review.
Here are the requested items:

- Tagged Release: v2.1.16 - https://github.com/evandeilton/gkwreg/releases/tag/v2.1.16
- Zenodo DOI: [10.5281/zenodo.18292325](https://doi.org/10.5281/zenodo.18292325)
- Zenodo Archive: https://doi.org/10.5281/zenodo.18292325

The metadata in the Zenodo archive (authors, ORCIDs, license) has been verified to match the paper and software repository.
