### 2026-01-18 #8991 @jbytecode
@evandeilton - The Zenodo metadata including the title, author list, orchids, license should match those in the paper and repo.

- the title doesn't match
- the author list doesn't match
- their orcid should be set
-  In the package DESCRIPTION the license is MIT. Please set the license of the repo. The license section should also be set.
