### 2026-01-18 #8991 @jbytecode
the title is "evandeilton/gkwreg: JOSS submission - Reviewed version [Zenodo]" which is wrong, please replace it with the paper's title.

the author list is wrong, please add the other authors and their orcids.

the license is Creative Commons Attribution 4.0 International, which is wrong, please set it MIT (as it is in the DESCRIPTION file of package) in both **repo** and **Zenodo** archive.
