### 2026-01-18 #8991 @evandeilton
Hi @jbytecode,

Since I'm new to Zenodo, I overlooked those details. I've now fixed everything. Please let me know if everything looks correct now.

Thanks!
