### 2026-01-18 #8991 @jbytecode
@evandeilton - Thank you for the changes. Please set the license properly in the repo as MIT.
