### 2026-01-18 #8991 @jbytecode
@evandeilton - it's still  `Unknown, MIT licenses found` in the repo
