### 2026-01-18 #8991 @jbytecode
@evandeilton - please fix the license issue in the repo, then, please have a full proof-read of the manuscript. Fix anything if exists. Please make minor changes if required. Ping me when you've done with them. Thank you in advance.
