### 2026-01-19 515f172
FIX Paper and MIT License for JOSS
