+++ inst/paper/paper.bib
@@ -18,9 +18,9 @@
 
 @Manual{gkwreg,
   title = {gkwreg: An R Package for Generalized Kumaraswamy Regression Models for Bounded Data},
-  author = {Lopes, JE},
-  year = {2025},
-  note = {R package version 1.0.11},
+  author = {Lopes, José Evandeilton and Bonat, Wagner Hugo},
+  year = {2026},
+  note = {R package version 2.1.16},
 }
 
 @article{Kristensen2016,
