+++ inst/paper/paper.md
@@ -1,5 +1,4 @@
 ---
-# Example from https://joss.readthedocs.io/en/latest/submitting.html
 title: 'gkwreg: An R Package for Generalized Kumaraswamy Regression Models for Bounded Data'
 tags:
   - R
@@ -17,13 +16,11 @@ authors:
     orcid: 0000-0002-0349-7054
     affiliation: "1"
 affiliations:
- - name: Paraná Federal University, Brazil
+ - name: Federal University of Paraná, Brazil
    index: 1
- # - name: Universidade Federal do Paraná (UFPR)
- #   index: 2
 citation_author: Lopes, JE & Bonat, WH
-date: '19 julho 2025'
-year: '2025'
+date: '19 January 2026'
+year: '2026'
 bibliography: paper.bib
 output: rticles::joss_article
 csl: apa.csl
@@ -70,15 +67,14 @@ and lacks a closed-form cumulative distribution function (CDF).
 The Kumaraswamy (Kw) distribution [@kumaraswamy1980] offers an analytically 
 simple CDF, yet its two-parameter form may be overly restrictive. 
 To overcome these limitations, the Generalized Kumaraswamy (GKw) distribution,
-a flexible, five-parameter family incorporating the Beta and Kw distributions
-introduced by [@carrasco2010] was developed. However, practical applications
+a flexible, five-parameter family incorporating the Beta and Kw distributions,
+introduced by [@carrasco2010], was developed. However, practical applications
 of GKw in regression contexts have faced computational challenges. 
 Its complex likelihood function makes Maximum Likelihood Estimation (MLE) 
 computationally demanding and unstable, necessitating efficient and user-friendly
 computational tools.
 
-This work introduces the `R` package `gkwreg`, which addresses this need. Built on the Template
-Model Builder (`TMB`) package [@Kristensen2016], it leverages automatic 
+This work introduces the `R` package `gkwreg`, which addresses this need. Built on the Template Model Builder (`TMB`) package [@Kristensen2016], it leverages automatic 
 differentiation (AD) in `C++` to efficiently compute gradients and Hessians,
 enhancing speed, accuracy, and stability of MLE, especially when 
 distribution parameters vary with covariates. `gkwreg` offers an intuitive
@@ -91,23 +87,16 @@ robust goodness-of-fit assessments.
 
 # Mathematics
 
-The Probability Density Function (PDF) of the five-parameter Generalized Kumaraswamy (GKw) 
-distribution is given by: 
-$$f(y; \boldsymbol{\theta}) = \frac{\lambda\,\alpha\,\beta\,y^{\alpha-1}}{B(\gamma, \delta+1)}\,\bigl(1-y^\alpha\bigr)^{\beta-1}\,\bigl[1-\bigl(1-y^\alpha\bigr)^\beta\bigr]^{\gamma\lambda-1}\,\left\{1-\bigl[1-\bigl(1-y^\alpha\bigr)^\beta\bigr]^\lambda\right\}^\delta$$
-where $\boldsymbol{\theta} = (\alpha, \beta, \gamma, \delta, \lambda)^\top$ is the vector of positive shape parameters and $B(\cdot, \cdot)$ is the beta function.
+The Probability Density Function (PDF) of the five-parameter Generalized Kumaraswamy (GKw) distribution is given by: 
[…]
