### 2026-01-19 #8991 @evandeilton
Hi @jbytecode ,

The license issue has been fixed using GitHub's license template ("Add file" → "Create new file" → LICENSE → "Choose a license template").
I have also added Wagner Bonat as a package co-author to reflect his contributions to both the paper and the package.

If any license-related issues persist, please let me know and I would be happy to address them.

The manuscript has been proofread and corrected accordingly.

The submission is now ready for your review.

Best regards
