+++ authors.md
@@ -5,19 +5,22 @@
 - **José Evandeilton Lopes**. Author, maintainer.
   [](https://orcid.org/0009-0007-5887-4084)
 
+- **Wagner Hugo Bonat**. Author.
+  [](https://orcid.org/0000-0002-0349-7054)
+
 ## Citation
 
 Source:
 [`DESCRIPTION`](https://github.com/evandeilton/gkwreg/blob/main/DESCRIPTION)
 
-Lopes J (2026). *gkwreg: Generalized Kumaraswamy Regression Models for
-Bounded Data*. R package version 2.1.14,
+Lopes J, Bonat W (2026). *gkwreg: Generalized Kumaraswamy Regression
+Models for Bounded Data*. R package version 2.1.16,
 <https://github.com/evandeilton/gkwreg>.
 
     @Manual{,
       title = {gkwreg: Generalized Kumaraswamy Regression Models for Bounded Data},
-      author = {José Evandeilton Lopes},
+      author = {José Evandeilton Lopes and Wagner Hugo Bonat},
       year = {2026},
-      note = {R package version 2.1.14},
+      note = {R package version 2.1.16},
       url = {https://github.com/evandeilton/gkwreg},
     }
