### 2026-01-19 #8991 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
