### 2026-01-19 3b64a2c
Fix README and MIT License
