+++ llms.txt
@@ -210,7 +210,7 @@ newdata <- data.frame(
 # Predict different quantities
 pred_mean <- predict(fit_kw, newdata, type = "response") # E(Y|X)
 pred_var <- predict(fit_kw, newdata, type = "variance") # Var(Y|X)
-pred_alpha <- predict(fit_kw, newdata, type = "alpha") # α parameter
+pred_alpha <- predict(fit_kw, newdata, type = "alpha") # <U+03B1> parameter
 pred_params <- predict(fit_kw, newdata, type = "parameter") # All parameters
 
 # Evaluate density at y = 0.5 for each observation
@@ -592,7 +592,7 @@ for details.
   Parameters
 - [`family(`*`<gkwreg>`*`)`](https://evandeilton.github.io/gkwreg/reference/family.gkwreg.md)
   : Extract Family from GKw Regression Model
-- [`fitted.gkwreg()`](https://evandeilton.github.io/gkwreg/reference/fitted.gkwreg.md)
+- [`fitted(`*`<gkwreg>`*`)`](https://evandeilton.github.io/gkwreg/reference/fitted.gkwreg.md)
   : Extract Fitted Values from a Generalized Kumaraswamy Regression
   Model
 - [`formula(`*`<gkwreg>`*`)`](https://evandeilton.github.io/gkwreg/reference/formula.gkwreg.md)
