+++ reference/gkwreg.md
@@ -726,7 +726,7 @@ for detailed interpretation guidance.
 - Set `model = FALSE`, `x = FALSE`, `y = FALSE` to reduce object size
   (but limits some post-fitting capabilities)
 
-- Hessian matrix scales as O(p²) where p = number of parameters
+- Hessian matrix scales as O(p\<U+00B2\>) where p = number of parameters
 
 ## Methods
 
@@ -946,41 +946,41 @@ summary(fit_default)
 #> 
 #> Coefficients:
 #>                     Estimate Std. Error z value Pr(>|z|)    
-#> alpha:(Intercept)  -1.419893   0.666514  -2.130   0.0331 *  
-#> alpha:batch1        0.918498   0.032368  28.377  < 2e-16 ***
-#> alpha:batch2        0.676328   0.034627  19.532  < 2e-16 ***
-#> alpha:batch3        0.781489   0.034587  22.595  < 2e-16 ***
-#> alpha:batch4        0.557260   0.032509  17.142  < 2e-16 ***
-#> alpha:batch5        0.562416   0.034498  16.303  < 2e-16 ***
-#> alpha:batch6        0.539377   0.034694  15.547  < 2e-16 ***
-#> alpha:batch7        0.308696   0.032499   9.499  < 2e-16 ***
-#> alpha:batch8        0.254507   0.034578   7.360 1.83e-13 ***
-#> alpha:batch9        0.198134   0.038561   5.138 2.77e-07 ***
+#> alpha:(Intercept)  -1.419609   0.656437  -2.163   0.0306 *  
+#> alpha:batch1        0.918493   0.032363  28.381  < 2e-16 ***
+#> alpha:batch2        0.676321   0.034621  19.535  < 2e-16 ***
+#> alpha:batch3        0.781482   0.034584  22.596  < 2e-16 ***
+#> alpha:batch4        0.557254   0.032499  17.147  < 2e-16 ***
+#> alpha:batch5        0.562413   0.034496  16.304  < 2e-16 ***
+#> alpha:batch6        0.539372   0.034684  15.551  < 2e-16 ***
+#> alpha:batch7        0.308690   0.032490   9.501  < 2e-16 ***
+#> alpha:batch8        0.254506   0.034577   7.361 1.83e-13 ***
+#> alpha:batch9        0.198131   0.038558   5.138 2.77e-07 ***
 #> alpha:temp          0.005403        NaN     NaN      NaN    
-#> beta:(Intercept)    5.511544   2.754889   2.001   0.0454 *  
-#> gamma:(Intercept)   2.560940   2.294913   1.116   0.2645    
-#> delta:(Intercept)  -2.153490   2.535825  -0.849   0.3958    
-#> lambda:(Intercept)  2.190526   2.324758   0.942   0.3461    
+#> beta:(Intercept)    5.512391   2.714173   2.031   0.0423 *  
+#> gamma:(Intercept)   2.559268   2.292375   1.116   0.2642    
+#> delta:(Intercept)  -2.153376   2.502826  -0.860   0.3896    
+#> lambda:(Intercept)  2.191087   2.320556   0.944   0.3451    
 #> ---
 #> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
 #> 
 #> Confidence intervals (95%):
 #>                         3%     98%
-#> alpha:(Intercept)  -2.7262 -0.1135
[…]
