### 2026-01-19 #8991 @evandeilton
Dear Chris Vernon, Editors, and Reviewers,

I am writing to express my gratitude for the acceptance and publication of my software, gkwreg, in JOSS.

I would like to thank Chris Vernon for the final approval and @jbytecode for editing and guiding this submission.

A special thanks to the reviewers, @JonasMoss and @vipulg13. Your constructive feedback and expertise were fundamental in refining the package.

Best regards,

José Evandeilton Lopes
